% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{angle.str2end}
\alias{angle.str2end}
\title{The angles to draw arcs between two line segments

Gives the two pairs of angles in radians or degrees to draw arcs between two vectors or line segments
for the draw.arc function in the plotrix package.
The angles are provided with respect to the x-axis in the coordinate system.
The line segments are [ba] and [bc] when the argument is given as a,b,c in the function.
radian is a logical argument (default=TRUE) which yields the angle in radians if TRUE, and in degrees if FALSE.
The first pair of angles is for drawing arcs in the smaller angle between [ba] and [bc]
and the second pair of angles is for drawing arcs in the counter-clockwise order from [ba] to [bc].}
\usage{
angle.str2end(a, b, c, radian = TRUE)
}
\arguments{
\item{a, b, c}{Three 2D points which represent the intersecting line segments [ba] and [bc].}

\item{radian}{A logical argument (default=TRUE). If TRUE, the smaller angle or counter-clockwise angle
between the line segments [ba] and [bc] is provided in radians, else it is provided in degrees.}
}
\value{
A list with two elements
\item{small.arc.angles}{Angles of [ba] and [bc] with the x-axis so that difference between them
is the smaller angle between [ba] and [bc] }
\item{ccw.arc.angles}{Angles of [ba] and [bc] with the x-axis so that difference between them
is the counter-clockwise angle between [ba] and [bc]}
}
\description{
The angles to draw arcs between two line segments

Gives the two pairs of angles in radians or degrees to draw arcs between two vectors or line segments
for the draw.arc function in the plotrix package.
The angles are provided with respect to the x-axis in the coordinate system.
The line segments are [ba] and [bc] when the argument is given as a,b,c in the function.
radian is a logical argument (default=TRUE) which yields the angle in radians if TRUE, and in degrees if FALSE.
The first pair of angles is for drawing arcs in the smaller angle between [ba] and [bc]
and the second pair of angles is for drawing arcs in the counter-clockwise order from [ba] to [bc].
}
\examples{
A<-c(.3,.2); B<-c(.6,.3); C<-c(1,1)
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
pts<-rbind(A,B,C)

angle3pnts(A,B,C)

angle3pnts(A,B,A)
round(angle3pnts(A,B,A),7)

angle3pnts(A,B,C,radian=F)

#plot of the line segments
Xlim<-range(pts[,1])
Ylim<-range(pts[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

ang1<-angle3pnts(A,B,C,radian=F)
ang2<-angle3pnts(B+c(1,0),B,C,radian=F)

sa<-angle.str2end(A,B,C,radian=F)$s #small arc angles
ang1<-sa[1]
ang2<-sa[2]

plot(pts,asp=1,pch=1,xlab="x",ylab="y",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
L<-rbind(B,B); R<-rbind(A,C)
segments(L[,1], L[,2], R[,1], R[,2], lty=2)
library(plotrix)
draw.arc(B[1],B[2],radius=xd*.1,deg1=ang1,deg2=ang2)
txt<-rbind(A,B,C)
text(txt+cbind(rep(xd*.05,nrow(txt)),rep(-xd*.02,nrow(txt))),c("A","B","C"))

text(rbind(B)+.15*xd*c(cos(pi*(ang2+ang1)/360),sin(pi*(ang2+ang1)/360)),paste(round(abs(ang1-ang2),2)," degrees"))

}
\seealso{
\code{\link{angle3pnts}}
}
