% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{ArcsAStri}
\alias{ArcsAStri}
\title{An object of class "Digraphs"

The arcs of Arc Slice Proximity Catch Digraph (AS-PCD) for 2D data - one triangle case}
\usage{
ArcsAStri(Xp, tri, M = "CC")
}
\arguments{
\item{Xp}{A set of 2D points which constitute the vertices of the AS-PCD}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{M}{"CC" which stand for circumcenter of the triangle tri or a 2D point in Cartesian coordinates or a 3D point in
barycentric coordinates which serves as a center in the interior of the triangle Tb;
default is M="CC" i.e. the circumcenter of tri}
}
\value{
A list with the elements
\item{type}{A description of the type of the digraph}
\item{parameters}{Parameters of the digraph, here, it is the center used to construct the vertex regions
i.e., circumcenter}
\item{tess.points}{Points on which the tessellation of the study region is performed, here, tessellation
is the support triangle.}
\item{tess.name}{Name of data set used in tessellation (i.e., vertices of the triangle).}
\item{vertices}{Vertices of the digraph, Xp.}
\item{vert.name}{Name of the data set which constitute the vertices of the digraph}
\item{S}{tails (or sources) of the arcs of AS-PCD for 2D data Xp as vertices of the digraph}
\item{E}{heads (or arrow ends) of the arcs of AS-PCD for 2D data Xp as vertices of the digraph}
\item{mtxt}{Text for "main" title in the plot of the digraph}
\item{quant}{Various quantities for the digraph: number of vertices, number of partition points,
number of intervals, number of arcs, and arc density.}
}
\description{
Returns arcs as tails (or sources) and heads (or arrow ends) for data set Xp as the vertices
of AS-PCD. AS proximity regions are constructed with respect to the triangle tri, i.e.,
arcs may exist for points only inside tri.
It also provides various descriptions and quantities about the arcs of the AS-PCD
such as number of arcs, arc density, etc. Vertex regions are based on the center M="CC" for
circumcenter of tri; or M=(m1,m2) in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric
coordinates in the interior of the triangle tri; default is M="CC" the circumcenter of tri.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
n<-10

#set.seed(1)
dat<-runif.tri(n,Tr)$g
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,2))

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.2)
#M<-circ.cent.tri(Tr); #M<-"CC"; #M<-"CM"
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

Gam2AStri(dat[1,],dat[2,],dat,Tr,M)
Gam2AStri(dat[1,],dat[1,],dat,Tr,M) #same two points cannot be a dominating set of size 2

Gam2AStri(c(.2,.4),dat[2,],dat,Tr,M)
Gam2AStri(c(.2,.4),c(.2,.5),dat,Tr,M)
Gam2AStri(c(.2,.4),c(.2,.5),rbind(c(.2,.4),c(.2,.5)),Tr,M)
#or try
rv1<-rv.triCC(c(.2,.4),Tr)$rv
rv2<-rv.triCC(c(.2,.5),Tr)$rv
Gam2AStri(c(.2,.4),c(.2,.5),rbind(c(.2,.4),c(.2,.5)),Tr,M,rv1,rv2)

ind.gam2<-vector()
for (i in 1:(n-1))
  for (j in (i+1):n)
  {if (Gam2AStri(dat[i,],dat[j,],dat,Tr,M)==1)
    ind.gam2<-rbind(ind.gam2,c(i,j))}
ind.gam2

#or try
rv1<-rv.triCC(dat[1,],Tr)$rv
rv2<-rv.triCC(dat[2,],Tr)$rv
Gam2AStri(dat[1,],dat[2,],dat,Tr,M,rv1,rv2)

#or try
rv1<-rv.triCC(dat[1,],Tr)$rv
Gam2AStri(dat[1,],dat[2,],dat,Tr,M,rv1)

#or try
Rv2<-rv.triCC(dat[2,],Tr)$rv
Gam2AStri(dat[1,],dat[2,],dat,Tr,M,rv2=Rv2)

Gam2AStri(c(1.3,1.2),c(1.35,1.25),dat,Tr,M)

dat.fr<-data.frame(a=dat)
Gam2AStri(c(.3,.2),c(.35,.25),dat.fr,Tr,M)

dat.fr<-data.frame(a=Tr)
Gam2AStri(c(.3,.2),c(.35,.25),dat,dat.fr,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ArcsASMT}}, \code{\link{ArcsPEtri}}, \code{\link{ArcsCStri}}, \code{\link{ArcsPEMT}},
 and \code{\link{ArcsCSMT}}
}
