% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{ArcsCSend1D}
\alias{ArcsCSend1D}
\title{An object of class "Digraphs"

The arcs of Central Similarity Proximity Catch Digraph (CS-PCD) for 1D data - end interval case}
\usage{
ArcsCSend1D(Xp, Yp, t)
}
\arguments{
\item{Xp}{A set or vector of 1D points which constitute the vertices of the CS-PCD}

\item{Yp}{A set or vector of 1D points which constitute the end points of the intervals}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}
}
\value{
A list with the elements
\item{type}{A description of the type of the digraph}
\item{parameters}{Parameters of the digraph, here, it is the expansion parameter.}
\item{tess.points}{Points on which the tessellation of the study region is performed, here, tessellation
is the intervalization based on Yp.}
\item{tess.name}{Name of data set used in tessellation, it is Yp for this function}
\item{vertices}{Vertices of the digraph, Xp points}
\item{vert.name}{Name of the data set which constitutes the vertices of the digraph}
\item{S}{tails (or sources) of the arcs of CS-PCD for 1D data in the end intervals}
\item{E}{heads (or arrow ends) of the arcs of CS-PCD for 1D data in the end intervals}
\item{mtxt}{Text for "main" title in the plot of the digraph}
\item{quant}{Various quantities for the digraph: number of vertices, number of partition points,
number of intervals (which is 2 for end intervals), number of arcs, and arc density.}
}
\description{
Returns arcs as tails (or sources) and heads (or arrow ends) for 1D data set Xp as the vertices
of CS-PCD.  Yp determines the end points of the end intervals.
For this function, CS proximity regions are constructed data points outside the intervals based on
Yp points with expansion parameter t>0. That is, for this function,
arcs may exist for points only inside end intervals.
It also provides various descriptions and quantities about the arcs of the CS-PCD
such as number of arcs, arc density, etc.
See also (\insertCite{ceyhan:revstat-2016;textual}{pcds}).
}
\examples{
asyvarCSend1D(1.2)

tseq<-seq(.01,5,by=.05)
ltseq<-length(tseq)

var.end<-vector()
for (i in 1:ltseq)
{
  var.end<-c(var.end,asyvarCSend1D(tseq[i]))
}

par(mar=c(5,5,4,2))
plot(tseq, var.end,type="l",xlab="t",ylab=expression(paste(sigma^2,"(t)")),lty=1,xlim=range(tseq))

}
\seealso{
\code{\link{ArcsCSmid1D}}, \code{\link{ArcsCS1D}} , \code{\link{ArcsPEmid1D}},
\code{\link{ArcsPEend1D}}  and \code{\link{ArcsPE1D}}
}
