% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{ArcsCSMI}
\alias{ArcsCSMI}
\title{An object of class "Digraphs"

The arcs of Central Similarity Proximity Catch Digraph (CS-PCD) for 1D data - multiple interval case}
\usage{
ArcsCSMI(Xp, Yp, t, c)
}
\arguments{
\item{Xp}{A set or vector of 1D points which constitute the vertices of the CS-PCD}

\item{Yp}{A set or vector of 1D points which constitute the end points of the intervals}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a)}
}
\value{
A list with the elements
\item{type}{A description of the type of the digraph}
\item{parameters}{Parameters of the digraph, here, they are expansion and centrality parameters.}
\item{tess.points}{Points on which the tessellation of the study region is performed, here, tessellation
is the intervalization of the real line based on Yp points.}
\item{tess.name}{Name of data set used in tessellation, it is Yp for this function}
\item{vertices}{Vertices of the digraph, Xp points}
\item{vert.name}{Name of the data set which constitute the vertices of the digraph}
\item{S}{tails (or sources) of the arcs of CS-PCD for 1D data}
\item{E}{heads (or arrow ends) of the arcs of CS-PCD for 1D data}
\item{mtxt}{Text for "main" title in the plot of the digraph}
\item{quant}{Various quantities for the digraph: number of vertices, number of partition points,
number of intervals, number of arcs, and arc density.}
}
\description{
Returns arcs as tails (or sources) and heads (or arrow ends) for 1D data set Xp as the vertices
of CS-PCD.  Yp determines the end points of the intervals.
For this function, CS proximity regions are constructed data points inside or outside the intervals based
on Yp points with expansion parameter t>0 and centrality parameter c in (0,1). That is, for this function,
arcs may exist for points in the middle or end intervals.
It also provides various descriptions and quantities about the arcs of the CS-PCD
such as number of arcs, arc density, etc.
Equivalent to function ArcsCS1D.
See also (\insertCite{ceyhan:revstat-2016;textual}{pcds}).
}
\examples{
c<-.4
t<-2
a<-0; b<-10; int<-c(a,b)

n<-10
dat<-runif(n,a,b)
NumArcsCSint(dat,t,c,int)

NumArcsCSint(3,t,c,int)

NumArcsCSint(dat,t,c=.3,int)

NumArcsCSint(dat,t=1.5,c,int)

n<-20
dat<-runif(n,a,b)
#dat<-runif(n,a-5,b+5)
#dat<-runif(n,a+5,b+5)
#dat<-runif(n,a+10,b+10)

NumArcsCSint(dat,t,c,int)

n<-10
dat<-runif(n,a,b)
NumArcsCSint(dat,t,c,int)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ArcsCSend1D}}, \code{\link{ArcsCSmid1D}}, \code{\link{ArcsCS1D}},
and \code{\link{ArcsPEMI}}
}
