% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{ArcsCStri}
\alias{ArcsCStri}
\title{An object of class "Digraphs"

The arcs of Central Similarity Proximity Catch Digraphs (CS-PCD) for 2D data - one triangle case}
\usage{
ArcsCStri(Xp, tri, t, M = c(1, 1, 1))
}
\arguments{
\item{Xp}{A set of 2D points which constitute the vertices of the CS-PCD}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri or the circumcenter of tri;
default is M=c(1,1,1) i.e. the center of mass of tri}
}
\value{
A list with the elements
\item{type}{A description of the type of the digraph}
\item{parameters}{Parameters of the digraph, here, it is the center used to construct the vertex regions.}
\item{tess.points}{Points on which the tessellation of the study region is performed, here, tessellation
is the support triangle.}
\item{tess.name}{Name of data set used in tessellation (i.e., vertices of the triangle)}
\item{vertices}{Vertices of the digraph, Xp points}
\item{vert.name}{Name of the data set which constitute the vertices of the digraph}
\item{S}{tails (or sources) of the arcs of CS-PCD for 2D data Xp as vertices of the digraph}
\item{E}{heads (or arrow ends) of the arcs of CS-PCD for 2D data Xp as vertices of the digraph}
\item{mtxt}{Text for "main" title in the plot of the digraph}
\item{quant}{Various quantities for the digraph: number of vertices, number of partition points,
number of intervals, number of arcs, and arc density.}
}
\description{
Returns arcs as tails (or sources) and heads (or arrow ends) for data set Xp as the vertices
of CS-PCD. CS proximity regions are constructed with respect to the triangle tri with expansion
parameter t>0, i.e., arcs may exist for points only inside tri.
It also provides various descriptions and quantities about the arcs of the CS-PCD
such as number of arcs, arc density, etc. Edge regions are based on center M=(m1,m2)
in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the interior of
the triangle tri; default is M=c(1,1,1) i.e. the center of mass of tri.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-CS,ceyhan:test2014;textual}{pcds}).
}
\examples{
nx<-100
ny<-10

#set.seed(1)
Xp<-cbind(runif(nx),runif(nx))
#Xp<-cbind(runif(nx,0.5,1.5),runif(nx,0,1))
#Xp<-cbind(runif(nx,1,2),runif(nx,0,1))
Yp<-cbind(runif(ny),runif(ny))

plotDeltri(Xp,Yp,xlab="",ylab = "")

NumArcsCSMT(Xp,Yp,t=.5)

TSArcDensCSMT(Xp,Yp,t=.5)
TSArcDensCSMT(Xp,Yp,t=.5,ch=TRUE)

NumArcsCSMT(Xp,Yp,t=1.0)
TSArcDensCSMT(Xp,Yp,t=1.0)
TSArcDensCSMT(Xp,Yp,t=1.0,ch=TRUE)

NumArcsCSMT(Xp,Yp,t=1.5)
TSArcDensCSMT(Xp,Yp,t=1.5)
TSArcDensCSMT(Xp,Yp,t=1.5,ch=TRUE)

TSArcDensCSMT(c(.4,.2),Yp,t=.5)

TSArcDensCSMT(Xp1,Yp1,t)

TSArcDensCSMT(Xp1,rbind(Yp1,Yp1),t)

dat.fr<-data.frame(a=Xp1)
TSArcDensCSMT(dat.fr,Yp1,t)

dat.fr<-data.frame(a=Yp1)
TSArcDensCSMT(Xp1,dat.fr,t)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ArcsCSMT}}, \code{\link{ArcsAStri}} and \code{\link{ArcsPEtri}}
}
