% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{ArcsPE1D}
\alias{ArcsPE1D}
\title{An object of class "Digraphs"

The arcs of Proportional Edge Proximity Catch Digraph (PE-PCD) for 1D data - multiple interval case}
\usage{
ArcsPE1D(Xp, Yp, r, c)
}
\arguments{
\item{Xp}{A set or vector of 1D points which constitute the vertices of the PE-PCD}

\item{Yp}{A set or vector of 1D points which constitute the end points of the intervals}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >=1}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a)}
}
\value{
A list with the elements
\item{type}{A description of the type of the digraph}
\item{parameters}{Parameters of the digraph, here, they are expansion and centrality parameters.}
\item{tess.points}{Points on which the tessellation of the study region is performed, here, tessellation
is the intervalization of the real line based on Yp points.}
\item{tess.name}{Name of data set used in tessellation, it is Yp for this function}
\item{vertices}{Vertices of the digraph, Xp points}
\item{vert.name}{Name of the data set which constitute the vertices of the digraph}
\item{S}{tails (or sources) of the arcs of PE-PCD for 1D data}
\item{E}{heads (or arrow ends) of the arcs of PE-PCD for 1D data}
\item{mtxt}{Text for "main" title in the plot of the digraph}
\item{quant}{Various quantities for the digraph: number of vertices, number of partition points,
number of intervals, number of arcs, and arc density.}
}
\description{
Returns arcs as tails (or sources) and heads (or arrow ends) for 1D data set Xp as the vertices
of PE-PCD.  Yp determines the end points of the intervals.
For this function, PE proximity regions are constructed data points inside or outside the intervals based
on Yp points with expansion parameter r>=1 and centrality parameter c in (0,1). That is, for this function,
arcs may exist for points in the middle or end intervals.
It also provides various descriptions and quantities about the arcs of the PE-PCD
such as number of arcs, arc density, etc.
Equivalent to function ArcsPEMI.
See also (\insertCite{ceyhan:metrika-2012;textual}{pcds}).
}
\examples{
r<-2
a<-0; b<-10;
#a<-runif(1); b<-runif(1); int<-range(a,b); a<-int[1]; b<-int[2]
nx<-40; ny<-10

#set.seed(1)
xr<-range(a,b)
xf<-(xr[2]-xr[1])*.5
Xp<-runif(nx,a-xf,b+xf)
#Xp<-runif(nx,a+5,b+5)
#Xp<-runif(nx,a+10,b+10)
Yp<-runif(ny,a,b)
#Yp<-runif(ny,a,b)+c(-10,10)

Arcs<-ArcsPEend1D(Xp,Yp,r)
Arcs
summary(Arcs)
plot(Arcs)

S<-Arcs$S
E<-Arcs$E

jit<-.1
yjit<-runif(nx,-jit,jit)

Xlim<-range(a,b,Xp,Yp)
xd<-Xlim[2]-Xlim[1]

plot(cbind(a,0),pch=".",main="arcs of PE-PCDs for points (jittered along y-axis)\\n in end intervals ", xlab=" ", ylab=" ",
     xlim=Xlim+xd*c(-.05,.05),ylim=3*c(-jit,jit))
abline(h=0,lty=1)
points(Xp, yjit,pch=".",cex=3)
abline(v=Yp,lty=2)
arrows(S, yjit, E, yjit, length = .05, col= 4)

ArcsPEend1D(Xp,Yp,r)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ArcsPEmid1D}}, \code{\link{ArcsPEend1D}}, \code{\link{ArcsPEMI}},
\code{\link{ArcsCSmid1D}}, \code{\link{ArcsCSend1D}} and \code{\link{ArcsCS1D}}
}
