% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{ArcsPEend1D}
\alias{ArcsPEend1D}
\title{An object of class "Digraphs"

The arcs of Proportional Edge Proximity Catch Digraph (PE-PCD) for 1D data - end interval case}
\usage{
ArcsPEend1D(Xp, Yp, r)
}
\arguments{
\item{Xp}{A set or vector of 1D points which constitute the vertices of the PE-PCD}

\item{Yp}{A set or vector of 1D points which constitute the end points of the intervals}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}
}
\value{
A list with the elements
\item{type}{A description of the type of the digraph}
\item{parameters}{Parameters of the digraph, here, it is the expansion parameter.}
\item{tess.points}{Points on which the tessellation of the study region is performed, here, tessellation
is the intervalization based on Yp.}
\item{tess.name}{Name of data set used in tessellation, it is Yp for this function}
\item{vertices}{Vertices of the digraph, Xp points}
\item{vert.name}{Name of the data set which constitutes the vertices of the digraph}
\item{S}{tails (or sources) of the arcs of PE-PCD for 1D data in the end intervals}
\item{E}{heads (or arrow ends) of the arcs of PE-PCD for 1D data in the end intervals}
\item{mtxt}{Text for "main" title in the plot of the digraph}
\item{quant}{Various quantities for the digraph: number of vertices, number of partition points,
number of intervals (which is 2 for end intervals), number of arcs, and arc density.}
}
\description{
Returns arcs as tails (or sources) and heads (or arrow ends) for 1D data set Xp as the vertices
of PE-PCD.  Yp determines the end points of the end intervals.
For this function, PE proximity regions are constructed data points outside the intervals based on
Yp points with expansion parameter r>=1. That is, for this function,
arcs may exist for points only inside end intervals.
It also provides various descriptions and quantities about the arcs of the PE-PCD
such as number of arcs, arc density, etc.
See also (\insertCite{ceyhan:metrika-2012;textual}{pcds}).
}
\examples{
asyvarPEend1D(1.2)

rseq<-seq(1.01,5,by=.05)
lrseq<-length(rseq)

var.end<-vector()
for (i in 1:lrseq)
{
  var.end<-c(var.end,asyvarPEend1D(rseq[i]))
}

par(mar=c(5,5,4,2))
plot(rseq, var.end,type="l",xlab="r",ylab=expression(paste(sigma^2,"(r)")),lty=1,xlim=range(rseq))

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ArcsPEmid1D}}, \code{\link{ArcsPE1D}} , \code{\link{ArcsCSmid1D}},
\code{\link{ArcsCSend1D}}  and \code{\link{ArcsCS1D}}
}
