% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{ArcsPEmid1D}
\alias{ArcsPEmid1D}
\title{An object of class "Digraphs"

The arcs of Proportional Edge Proximity Catch Digraph (PE-PCD) for 1D data - middle intervals case}
\usage{
ArcsPEmid1D(Xp, Yp, r, c)
}
\arguments{
\item{Xp}{A set or vector of 1D points which constitute the vertices of the PE-PCD}

\item{Yp}{A set or vector of 1D points which constitute the end points of the intervals}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a)}
}
\value{
A list with the elements
\item{type}{A description of the type of the digraph}
\item{parameters}{Parameters of the digraph, here, they are expansion and centrality parameters.}
\item{tess.points}{Points on which the tessellation of the study region is performed, here, tessellation
is the intervalization based on Yp points.}
\item{tess.name}{Name of data set used in tessellation, it is Yp for this function}
\item{vertices}{Vertices of the digraph, i.e. Xp points}
\item{vert.name}{Name of the data set which constitute the vertices of the digraph}
\item{S}{tails (or sources) of the arcs of PE-PCD for 1D data in the middle intervals}
\item{E}{heads (or arrow ends) of the arcs of PE-PCD for 1D data in the middle intervals}
\item{mtxt}{Text for "main" title in the plot of the digraph}
\item{quant}{Various quantities for the digraph: number of vertices, number of partition points,
number of intervals, number of arcs, and arc density.}
}
\description{
Returns arcs as tails (or sources) and heads (or arrow ends) for 1D data set Xp as the vertices
of PE-PCD. For this function, PE proximity regions are constructed with respect to the intervals
based on Yp points with expansion parameter r>=1 and centrality parameter c in (0,1). That is, for this function,
arcs may exist for points only inside the intervals.
It also provides various descriptions and quantities about the arcs of the PE-PCD
such as number of arcs, arc density, etc. Vertex regions are based on center Mc of each middle interval.
See also (\insertCite{ceyhan:metrika-2012;textual}{pcds}).
}
\examples{
c<-.4
r<-2
a<-0; b<-10; int<-c(a,b)

n<-20
dat<-runif(n,a,b)

NumArcsPEint(dat,r,c,int)
TSArcDensPE1D(dat,r,c,int)
TSArcDensPE1D(dat,r,c,int,alt="g")
TSArcDensPE1D(dat,r,c,int,alt="l")

NumArcsPEint(dat,r,c=.3,int)
TSArcDensPE1D(dat,r,c=.3,int)

NumArcsPEint(dat,r=1.5,c,int)
TSArcDensPE1D(dat,r=1.5,c,int)

dat<-runif(n,a-1,b+1)
NumArcsPEint(dat,r,c,int)
TSArcDensPE1D(dat,r,c,int)

c<-.4
r<-2
a<-0; b<-10; int<-c(a,b)
n<-20
dat<-runif(n,a,b)
TSArcDensPE1D(dat,r,c,int)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ArcsPEend1D}}, \code{\link{ArcsPE1D}}, \code{\link{ArcsCSmid1D}},
 \code{\link{ArcsCSend1D}} and \code{\link{ArcsCS1D}}
}
