% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{ArcsPEtri}
\alias{ArcsPEtri}
\title{An object of class "Digraphs"

The arcs of Proportional Edge Proximity Catch Digraph (PE-PCD) for 2D data - one triangle case}
\usage{
ArcsPEtri(Xp, tri, r, M = c(1, 1, 1))
}
\arguments{
\item{Xp}{A set of 2D points which constitute the vertices of the PE-PCD}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri or the circumcenter of tri;
default is M=c(1,1,1) i.e. the center of mass of tri}
}
\value{
A list with the elements
\item{type}{A description of the type of the digraph}
\item{parameters}{Parameters of the digraph, here, it is the center used to construct the vertex regions.}
\item{tess.points}{Points on which the tessellation of the study region is performed, here, tessellation
is the support triangle.}
\item{tess.name}{Name of data set used in tessellation (i.e., vertices of the triangle)}
\item{vertices}{Vertices of the digraph, Xp points}
\item{vert.name}{Name of the data set which constitute the vertices of the digraph}
\item{S}{tails (or sources) of the arcs of PE-PCD for 2D data Xp as vertices of the digraph}
\item{E}{heads (or arrow ends) of the arcs of PE-PCD for 2D data Xp as vertices of the digraph}
\item{mtxt}{Text for "main" title in the plot of the digraph}
\item{quant}{Various quantities for the digraph: number of vertices, number of partition points,
number of intervals, number of arcs, and arc density.}
}
\description{
Returns arcs as tails (or sources) and heads (or arrow ends) for data set Xp as the vertices
of PE-PCD. PE proximity regions are constructed with respect to the triangle tri with expansion
parameter r>=1, i.e., arcs may exist for points only inside tri.
It also provides various descriptions and quantities about the arcs of the PE-PCD
such as number of arcs, arc density, etc. Vertex regions are based on center M=(m1,m2)
in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the interior of
the triangle tri or based on the circumcenter of tri; default is M=c(1,1,1) i.e. the center of mass of tri.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-PE;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
n<-10

#set.seed(1)
dat<-runif.tri(n,Tr)$gen.points
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.0)
#M<-circ.cent.tri(Tr)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)
r<-1.5

S<-rbind(dat[1,],dat[2,])
IndNPEtri.domset(S,dat,r,Tr,M)

S<-rbind(dat[1,],dat[2,],dat[3,],dat[5,])
IndNPEtri.domset(S,dat,r,Tr,M)

S<-rbind(c(.1,.1),c(.3,.4),c(.5,.3))
IndNPEtri.domset(S,dat,r,Tr,M)

IndNPEtri.domset(c(.2,.5),dat,r,Tr,M)
IndNPEtri.domset(c(.2,.5),c(.2,.5),r,Tr,M)

IndNPEtri.domset(dat[5,],dat[2,],r,Tr,M)

S<-rbind(dat[1,],dat[2,],dat[3,],dat[5,],c(.2,.5))
IndNPEtri.domset(S,dat[3,],r,Tr,M)

IndNPEtri.domset(dat,dat,r,Tr,M)

P<-c(.4,.2)
S<-dat[c(1,3,4),]
IndNPEtri.domset(dat,P,r,Tr,M)

IndNPEtri.domset(rbind(S,S),dat,r,Tr,M)

dat.fr<-data.frame(a=dat)
IndNPEtri.domset(S,dat.fr,r,Tr,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ArcsPEMT}}, \code{\link{ArcsAStri}} and \code{\link{ArcsCStri}}
}
