% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{centMc}
\alias{centMc}
\title{(Parameterized) Center of an interval}
\usage{
centMc(int, c)
}
\arguments{
\item{int}{A vector with two entries representing an interval}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a)}
}
\value{
(parameterized) center inside int
}
\description{
Returns the (parameterized) center, Mc, of the interval, int=(a,b), parameterized by c in (0,1)
so that 100c % of the length of interval is to the left of Mc and 100(1-c) % of the length of the interval
is to the right of Mc. That is, for the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a).
See also (\insertCite{ceyhan:metrika-2012,ceyhan:revstat-2016;textual}{pcds}).
}
\examples{
nx<-10
ny<-10

#set.seed(1)
Xp<-cbind(runif(nx,0,1),runif(nx,0,1))
#Xp<-cbind(runif(nx,0.5,1.5),runif(nx,0,1))
#Xp<-cbind(runif(nx,1,2),runif(nx,0,1))
Yp<-cbind(runif(ny,0,1),runif(ny,0,1))

M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-"CC"
#M<-c(1.3,1.3)

plot.regsASMT(Xp,Yp,M,xlab="",ylab="")

Xlim<-range(Xp[,1],Yp[,1])
Ylim<-range(Xp[,2],Yp[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot.regsASMT(Xp,Yp,M,xlab="",ylab="",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))

Xp<-c(.5,.5)
Xlim<-Ylim<-range(Xp,Yp)
plot.regsASMT(Xp,Yp,M,xlab="",ylab="",xlim=Xlim,ylim=Ylim)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{centersMc}}
}
