% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{circ.cent.bastri}
\alias{circ.cent.bastri}
\title{Circumcenter of a basic triangle}
\usage{
circ.cent.bastri(c1, c2)
}
\arguments{
\item{c1, c2}{Positive real numbers representing the top vertex in basic triangle Tb=T((0,0),(1,0),(c1,c2)),
c1 must be in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1}
}
\value{
circumcenter of the basic triangle Tb=T((0,0),(1,0),(c1,c2)) given c1,c2 as the arguments of the
function
}
\description{
Returns the circumcenter of a basic triangle Tb=T((0,0),(1,0),(c1,c2)) given c1,c2
where c1 is in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1.
Any given triangle can be mapped to the basic triangle by a combination of rigid body motions
(i.e., translation, rotation and reflection) and scaling, preserving uniformity of the points in the
original triangle. Hence basic triangle is useful for simulation
studies under the uniformness hypothesis.
See (\insertCite{weisstein-tri-centers,ceyhan:comp-geo-2010;textual}{pcds}) for triangle centers and
(\insertCite{ceyhan:arc-density-PE,ceyhan:arc-density-CS,ceyhan:dom-num-NPE-Spat2011;textual}{pcds}) for the basic triangle.
}
\examples{
n<-10
M<-matrix(sample(c(0,1),n^2,replace=T),nrow=n)
diag(M)<-1

dom.greedy(M)
IndUBdom(M,2)
dom.exact(M)

}
\seealso{
\code{\link{circ.cent.tri}}
}
