% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{circ.cent.tetra}
\alias{circ.cent.tetra}
\title{Circumcenter of a general tetrahedron}
\usage{
circ.cent.tetra(th)
}
\arguments{
\item{th}{Four 3D points, stacked row-wise, each row representing a vertex of the tetrahedron}
}
\value{
circumcenter of the tetrahedron th
}
\description{
Returns the circumcenter a given tetrahedron th with vertices stacked row-wise.
}
\examples{
n<-100

#set.seed(1)
Xdt<-runif.tetra(n,tetra)
Xdt
summary(Xdt)
plot(Xdt)

Dt<-Xdt$g
# library(scatterplot3d)
# s3d<-scatterplot3d(Dt, highlight.3d=TRUE,xlab="x",ylab="y",zlab="z",
#                    col.axis="blue", col.grid="lightblue",
#                    main="3D Scatterplot of the Data", pch=20)
# s3d$points3d(tetra,pch=20,col="blue")

Xlim<-range(tetra[,1],Dt[,1])
Ylim<-range(tetra[,2],Dt[,2])
Zlim<-range(tetra[,3],Dt[,3])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]
zd<-Zlim[2]-Zlim[1]

library(plot3D)
scatter3D(Dt[,1],Dt[,2],Dt[,3], theta =225, phi = 30, bty = "g", xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05), zlim=Zlim+zd*c(-.05,.05),
          pch = 20, cex = 1, ticktype = "detailed")
#add the vertices of the tetrahedron
points3D(tetra[,1],tetra[,2],tetra[,3], add=TRUE)
L<-rbind(A,A,A,B,B,C); R<-rbind(B,C,D,C,D,D)
segments3D(L[,1], L[,2], L[,3], R[,1], R[,2],R[,3], add=TRUE,lwd=2)

text3D(tetra[,1],tetra[,2],tetra[,3], labels=c("A","B","C","D"), add=TRUE)

runif.tetra(n,tetra)

dat.fr<-data.frame(a=tetra)
runif.tetra(n,dat.fr)

}
\seealso{
\code{\link{circ.cent.tri}}
}
