% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{cl2eVRCM}
\alias{cl2eVRCM}
\title{An object of class "Extrema"

The closest points in a data set to edges in each CM-vertex region in a triangle}
\usage{
cl2eVRCM(dat, tri)
}
\arguments{
\item{dat}{A set of 2D points representing the set of data points}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}
}
\value{
A list with the elements
\item{txt1}{Vertex labels are A=1, B=2, and C=3 (corresponds to row number in Extrema Points).}
\item{txt2}{A short description of the distances as "Distances to Edges in the Respective
CM-Vertex Regions".}
\item{type}{Type of the extrema points}
\item{desc}{A short description of the extrema points}
\item{mtxt}{The "main" title for the plot of the exrema}
\item{Ext}{The extrema points, here, closest points to edges in the respective vertex region.}
\item{X}{The input data, dat, can be a matrix or data frame}
\item{num.points}{The number of data points, i.e., size of dat}
\item{supp}{Support of the data points, here, it is tri}
\item{cent}{The center point used for construction of vertex regions}
\item{ncent}{Name of the center, cent, it is "CM" for this function}
\item{regions}{Vertex regions inside the triangle, tri, provided as a list}
\item{region.names}{Names of the vertex regions as "vr=1","vr=2","vr=3"}
\item{region.centers}{Centers of mass of the vertex regions inside tri}
\item{dist2ref}{Distances of closest points in the vertex regions to corresponding edges}
}
\description{
Returns the closest data points among the data set, dat, to edge j in CM-vertex region j for j=1,2,3
in the triangle, tri=T(A,B,C), where CM stands for center of mass. Vertex labels are A=1, B=2, and C=3,
and corresponding edge labels are BC=1, AC=2, and AB=3.
Function yields NA if there are no data points in a CM-vertex region.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:masa-2007,ceyhan:comp-geo-2010,ceyhan:dom-num-NPE-Spat2011;textual}{pcds}).
}
\seealso{
\code{\link{cl2e.inTbVR.CM}}, \code{\link{cl2eVRCC}}, \code{\link{cl2eVRcent}},
and \code{\link{cl2edgesTe}}
}
