% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{cl2Mc.int}
\alias{cl2Mc.int}
\title{An object of class "Extrema"

The closest points to center in each vertex region in an interval}
\usage{
cl2Mc.int(Dt, int, c)
}
\arguments{
\item{Dt}{A set or vector of 1D points from which closest points to Mc are found
in the interval int}

\item{int}{A vector of two real numbers representing an interval}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a)}
}
\value{
A list with the elements
\item{txt1}{Vertex Labels are a=1 and b=2 for the interval (a,b).}
\item{txt2}{A short description of the distances as "Distances from ..."}
\item{type}{Type of the extrema points}
\item{desc}{A short description of the extrema points}
\item{mtxt}{The "main" title for the plot of the exrema}
\item{Ext}{The extrema points, here, closest points to Mc in each vertex region}
\item{X}{The input data vector, Dt,}
\item{num.points}{The number of data points, i.e., size of Dt}
\item{supp}{Support of the data points, here, it is int.}
\item{cent}{The (parameterized) center point used for construction of vertex regions.}
\item{ncent}{Name of the (parameterized) center, cent, it is "Mc" for this function.}
\item{regions}{Vertex regions inside the interval, int, provided as a list.}
\item{region.names}{Names of the vertex regions as "vr=1","vr=2"}
\item{region.centers}{Centers of mass of the vertex regions inside int.}
\item{dist2ref}{Distances from closest points in each vertex region to Mc.}
}
\description{
Returns the closest data points among the data set, Dt, in each Mc-vertex region
i.e., finds the closest points from right and left to Mc among points of the 1D data set Dt which reside in
in the interval int=(a,b). Mc is based on the centrality parameter c in (0,1),
so that 100c % of the length of interval is to the left of Mc and 100(1-c) % of the length of the interval
is to the right of Mc. That is, for the interval (a,b), Mc=a+c*(b-a). If there are no points from Dt to
the left of Mc in the interval, then it yields NA, and likewise for the right of Mc in the interval.
See also (\insertCite{ceyhan:metrika-2012;textual}{pcds}).
}
\examples{
c<-.4
r<-2
a<-0; b<-10; int<-c(a,b)

n<-10
dat<-runif(n,a,b)
NumArcsPEmid1D(dat,r,c,int)

NumArcsPEmid1D(3,r,c,int)

NumArcsPEmid1D(dat,r,c=.3,int)

NumArcsPEmid1D(dat,r=1.5,c,int)

n<-20
dat<-runif(n,a-5,b+5)
NumArcsPEmid1D(dat,r,c,int)

dat<-runif(n,a+10,b+10)
NumArcsPEmid1D(dat,r,c,int)

n<-10
dat<-runif(n,a,b)
NumArcsPEmid1D(dat,r,c,int)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{cl2CC.TbVR}} and \code{\link{cl2CC.VR}}
}
