% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{cp2e.bastri}
\alias{cp2e.bastri}
\title{Projections of a point inside the basic triangle to its edges}
\usage{
cp2e.bastri(c1, c2, M)
}
\arguments{
\item{c1, c2}{Positive real numbers which constitute the vertex of the basic triangle
adjacent to the shorter edges; c1 must be in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the basic triangle}
}
\value{
Three projection points (stacked row-wise) from a general center M=(m1,m2) in Cartesian coordinates
or M=(alpha,beta,gamma) in barycentric coordinates in the interior of a basic triangle to the edges on
the extension of the lines joining M to the vertices;
row i is the projection  point into edge i, for i=1,2,3.
}
\description{
Returns the projections from a general center M=(m1,m2) in Cartesian coordinates or M=(alpha,beta,gamma) in
barycentric coordinates in the interior of the basic triangle Tb=T((0,0),(1,0),(c1,c2))
to the edges on the extension of the lines joining M to the vertices (see the examples for an illustration).
In the basic triangle Tb, c1 is in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1.
Any given triangle can be mapped to the basic triangle by a combination of rigid body motions
(i.e., translation, rotation and reflection) and scaling, preserving uniformity of the points in the
original triangle. Hence basic triangle is useful for simulation studies under the uniformness hypothesis.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010;textual}{pcds}).
}
\examples{
a<-0; b<-10; int<-c(a,b)
c<-.4
#a<-runif(1); b<-runif(1); int<-range(a,b); a<-int[1]; b<-int[2]
#c<-runif(1);
r<-1/max(c,1-c)
nx<-40; ny<-10

#set.seed(1)
Xp<-runif(nx,a,b)
Yp<-runif(ny,a,b)
PEdom1D(Xp,Yp,r,c)

plot.intervals(Xp,Yp,xlab="",ylab="")
plot.regsPEMI(Xp,Yp,r,c)

TSDomPEBin.1D(Xp,Yp,int,c,alt="t")
TSDomPEBin.1D(Xp,int=int,c=c,alt="t")

TSDomPEBin.1D(Xp,Yp,int,c,alt="l")
TSDomPEBin.1D(Xp,Yp,int,c,alt="g")
TSDomPEBin.1D(Xp,Yp,int,c,end=TRUE)
TSDomPEBin.1D(Xp,Yp,int,c=.25)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{cp2e.tri}} and \code{\link{cp2edges.nd}}
}
