% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{cp2e.tri}
\alias{cp2e.tri}
\title{Projections of a point inside a triangle to its edges}
\usage{
cp2e.tri(tri, M)
}
\arguments{
\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri}
}
\value{
Three projection points (stacked row-wise) from a general center M=(m1,m2) in Cartesian coordinates
or M=(alpha,beta,gamma) in barycentric coordinates in the interior of a triangle to the edges on
the extension of the lines joining M to the vertices;
row i is the projection  point into edge i, for i=1,2,3.
}
\description{
Returns the projections from a general center M=(m1,m2) in Cartesian coordinates or M=(alpha,beta,gamma) in
barycentric coordinates in the interior of a triangle to the edges on the extension of the lines joining
M to the vertices (see the examples for an illustration).
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(0,0); B<-c(1,0); C<-c(.5,sqrt(3.)/2);
Te<-rbind(A,B,C)
n<-10

#set.seed(1)
dat<-runifTe(n)$gen.points
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

M<-as.numeric(runifTe(1)$g)
#M<-c(.6,.2)
#M<-circ.cent.tri(Te)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

NumArcsPETe(dat,r=1.25)

inc.mat<-IncMatPETe(dat,r=1.25,M)
inc.mat
sum(inc.mat)-n

dom.greedy(inc.mat)
IndUBdom(inc.mat,2)
dom.exact(inc.mat)

inc.mat<-IncMatPETe(rbind(dat,c(0,1)),r=1.25,M)
inc.mat
sum(inc.mat)-(n+1)

IncMatPETe(dat,r=1.5,M);

IncMatPETe(rbind(dat,dat),r=1.5,M)

dat.fr<-data.frame(a=dat)
IncMatPETe(dat.fr,r=1.5,M);

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{cp2e.bastri}} and \code{\link{cp2edges.nd}}
}
