% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{cp2edges.nd}
\alias{cp2edges.nd}
\title{Projections of Centers for non-degenerate asymptotic distribution of domination number of
Proportional Edge Proximity Catch Digraphs (PE-PCDs) to its edges}
\usage{
cp2edges.nd(tri, r, cent = 1)
}
\arguments{
\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region;
must be in (1,1.5] for this function}

\item{cent}{Index of the center (as 1,2,3 corresponding to M1,M2,M3) which gives nondegenerate asymptotic
distribution of the domination number of PE-PCD for uniform data in tri for expansion parameter r>=1 in (1,1.5];
default cent=1}
}
\value{
Three projection points (stacked row-wise) from one of the centers (as 1,2,3 corresponding to
M1,M2,M3) which gives nondegenerate asymptotic distribution of the domination number of PE-PCD for
uniform data in tri for expansion parameter r>=1 in (1,1.5].
}
\description{
Returns the projections from  center M to the edges on the extension of the lines joining M to the vertices
in the triangle, tri. Here M is one of the three centers which gives nondegenerate asymptotic distribution
of the domination number of PE-PCD for uniform data in tri for a given expansion parameter r>=1 in (1,1.5).
The center label cent values 1,2,3 correspond to the vertices M1,M2,and M3 (i.e., row numbers in the
output of cent.nondeg(tri,r)); default for cent is 1. M becomes center of mass CM for r=1.5.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:masa-2007,ceyhan:dom-num-NPE-Spat2011;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#c1<-.4; c2<-.6; A<-c(0,0); B<-c(1,0); C<-c(c1,c2);
#A<-c(0,0); B<-c(1,0); C<-c(.5,sqrt(3.)/2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
r<-1.35

Ms<-cent.nondeg(Tr,r)
Ms

Xlim<-range(Tr[,1])
Ylim<-range(Tr[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Tr,pch=".",xlab="",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Tr)
points(Ms,pch=".",col=1)
polygon(Ms,lty=2)

xc<-Tr[,1]#+c(-.02,.02,.02)
yc<-Tr[,2]#+c(.02,.02,.03)
txt.str<-c("A","B","C")
text(xc,yc,txt.str)

xc<-Ms[,1]#+c(-.04,.04,.03)
yc<-Ms[,2]#+c(.02,.02,.05)
txt.str<-c("M1","M2","M3")
text(xc,yc,txt.str)

cent.nondeg(Tr,r)

dat.fr<-data.frame(a=Tr)
cent.nondeg(dat.fr,r)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{cp2e.bastri}} and \code{\link{cp2e.tri}}
}
