% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{CSarcdens.tri}
\alias{CSarcdens.tri}
\title{Arc density of Central Similarity Proximity Catch Digraphs (CS-PCDs) - one triangle case}
\usage{
CSarcdens.tri(Xp, tri, t, M = c(1, 1, 1), tri.cor = TRUE)
}
\arguments{
\item{Xp}{A set of 2D points which constitute the vertices of the CS-PCD}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri;
default is M=c(1,1,1) i.e. the center of mass of tri}

\item{tri.cor}{A logical argument for computing the arc density for only the points inside the triangle, tri
(default=TRUE), i.e., if TRUE only the induced digraph with the vertices inside tri are considered in the
computation of arc density}
}
\value{
A list of elements
\item{arc.dens}{Arc density of CS-PCD whose vertices are the 2D numerical data set, Xp;
CS proximity regions are defined with respect to the triangle tri and M-edge regions}
\item{std.arc.dens}{Arc density standardized by the mean and asymptotic variance of the arc
density of CS-PCD for uniform data in the triangle tri.}
\item{caveat}{The warning as "The standardized arc density is only correct when M is the center of mass
in the current version".}
}
\description{
Returns the arc density of CS-PCD whose vertex set is the given 2D numerical data set, Xp,
(some of its members are) in the triangle tri. CS proximity regions is defined with respect to tri with
expansion parameter t>0 and edge regions are based on center M=(m1,m2) in Cartesian coordinates or
M=(alpha,beta,gamma) in barycentric coordinates in the interior of the triangle tri; default is
M=c(1,1,1) i.e. the center of mass of tri.
The function also provides arc density standardized by the mean and asymptotic variance of the arc density
of CS-PCD for uniform data in the triangle tri. For the number of arcs, loops are not allowed.
tri.cor is a logical argument for triangle correction (default is TRUE), if TRUE, only the points inside the
triangle are considered (i.e., digraph induced by these vertices are considered) in computing the arc density,
otherwise all points are considered (for the number of vertices in the denominator of arc density).
See (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-CS,ceyhan:test2014;textual}{pcds}) for more on CS-PCDs.
Caveat: The standardized arc density is only correct when M is the center of mass in the current version.
}
\seealso{
\code{\link{ASarcdens.tri}}, \code{\link{PEarcdens.tri}}, and \code{\link{NumArcsCStri}}
}
