% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{dimension}
\alias{dimension}
\title{The dimension of a vector or matrix or a data frame}
\usage{
dimension(x)
}
\arguments{
\item{x}{A vector or a matrix or a data frame whose dimension is to be determined.}
}
\value{
Dimension (i.e., # of columns) of \code{x}
}
\description{
Returns the dimension (i.e., # of columns) of x, which is a matrix or a vector or a data frame.
This is different than the dim function in the base distribution of R, in the sense that, dimension gives
only the number of columns of the argument x, while dim gives the number of rows and columns of x.
dimension also works for a scalar or a vector, while dim yields NULL for such arguments.
}
\examples{
A<-c(-1.22,-2.33); B<-c(2.55,3.75,4)
is.point(A)
is.point(A,2)
is.point(A,1)
is.point(A,3)

is.point(B)
is.point(B,3)

dat<-rbind(A,B[1:2])
is.point(dat)
is.point(as.vector(dat),4)

lis<-list(a=A,b=B)
is.point(lis)

dat.fr<-data.frame(a=A,b=B[1:2])
is.point(dat.fr)

is.point(c("a","b"))

}
\seealso{
\code{\link{is.point}} and  \code{\link{dim}} from the base distribution of R
}
