% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{dist.pt2set}
\alias{dist.pt2set}
\title{Distance from a point to a set of finite cardinality}
\usage{
dist.pt2set(x, Y)
}
\arguments{
\item{x}{A vector (i.e., a point in R^d)}

\item{Y}{A set of d-dimensional points}
}
\value{
A list of elements
\item{distance}{Distance from point x to set Y}
\item{ind.cl.point}{Index of the closest point in set Y to the point x}
\item{closest.point}{The closest point in set Y to the point x}
}
\description{
Returns the Euclidean distance between a point x and set of points Y and the closest point in set Y to x.
Distance between a point and a set is by definition the distance from the point to the closest point in the
set. x should be of finite dimension and Y should be of finite cardinality and x and elements of Y
must have the same dimension.
}
\examples{
B<-c(1,0); C<-c(.5,sqrt(3.)/2);
Dist(B,C);
dist(rbind(B,C))
dist(rbind(as.vector(B),as.vector(C)))
Dist(B,B);

x<-runif(10)
y<-runif(10)
Dist(x,y)

xm<-matrix(x,ncol=2)
ym<-matrix(y,ncol=2)
Dist(xm,ym)
dist(rbind(as.vector(xm),as.vector(ym)))

Dist(xm,xm)

xm<-matrix(x,ncol=2)
ym<-matrix(y,ncol=1)
Dist(xm,ym) #gives error, since vectors are not of the same dimension

Dist(B,xm)

Dist(B,x) #gives error, since vectors are not of the same dimension

dat.fr<-data.frame(b=B,c=C)
Dist(B,dat.fr)
Dist(dat.fr,dat.fr)
Dist(dat.fr,cbind(B,C))


}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{dp2l}} and \code{\link{dp2pl}}
}
