% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{dom.greedy}
\alias{dom.greedy}
\title{Approximate domination number and approximate dominating set by the greedy algorithm}
\usage{
dom.greedy(Inc.Mat)
}
\arguments{
\item{Inc.Mat}{A square matrix consisting of 0s and 1s which represents the incidence matrix of
a graph or digraph}
}
\value{
A list with two elements
\item{dom.num}{The cardinality of the (approximate) minimum dominating set found by the greedy algorithm.
i.e. (approximate) domination number of the graph or digraph whose incidence matrix Inc.Mat is given
as input.}
\item{ind.domset}{Indices of the rows in the incidence matrix Inc.Mat for the ((approximate) minimum
dominating set). The row numbers in the incidence matrix correspond to the indices of the vertices
(i.e. indices of the data points).}
}
\description{
Returns the (approximate) domination number and the indices (i.e. row numbers) for the corresponding
(approximate) minimum dominating set based on the incidence matrix Inc.Mat of a graph or a digraph
by using the greedy algorithm (\insertCite{chvatal:1979;textual}{pcds}).
Here the row number in the incidence matrix corresponds to the index of the vertex (i.e. index of the
data point). The function works whether loops are allowed or not (i.e., whether the first diagonal
is all 1 or all 0). This function may yield the actual domination number or overestimates it.
}
\examples{
nx<-1000 #number of X points (target) #try nx<-1, 10, 100 also
ny<-10 #number of Y points (nontarget)

#set.seed(1)
Yp<-cbind(runif(ny),runif(ny))
del<-.4

Xdt<-rascMT(nx,Yp,del)
Xdt
summary(Xdt)
plot(Xdt)

Yp<-cbind(runif(ny),runif(ny))
del<-.3 #try .5, .75, .85
dat<-rascMT(nx,Yp,del) #data under CSR in the convex hull of Ypoints
#or use
DTY<-tri.mesh(Yp[,1],Yp[,2],duplicate="remove") #Delaunay triangulation based on Y points
dat<-rascMT(nx,Yp,del,DTY) #data under CSR in the convex hull of Ypoints

#or use
DTY<-tri.mesh(Yp[,1],Yp[,2],duplicate="remove") #Delaunay triangulation based on Y points
TRY<-triangles(DTY)[,1:3];
dat<-rascMT(nx,Yp,del,DTr=TRY) #data under CSR in the convex hull of Ypoints

#or use
DTY<-tri.mesh(Yp[,1],Yp[,2],duplicate="remove") #Delaunay triangulation based on Y points
TRY<-triangles(DTY)[,1:3];
dat<-rascMT(nx,Yp,del,DTY,TRY)$g #data under CSR in the convex hull of Ypoints

Xlim<-range(Yp[,1])
Ylim<-range(Yp[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

#plot of the data in the convex hull of Y points together with the Delaunay triangulation
DTY<-tri.mesh(Yp[,1],Yp[,2],duplicate="remove") #Delaunay triangulation based on Y points

plot(dat,main=" ", xlab=" ", ylab=" ",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05),type="n")
plot.tri(DTY, add=TRUE, do.points=TRUE,col="blue")
points(dat,pch=".",cex=3)

Yp<-rbind(c(.3,.2),c(.4,.5),c(.14,.15))
rascMT(nx,Yp,del)

dat.fr<-data.frame(a=Yp)
rascMT(nx,dat.fr,del)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{dom.exact}}, \code{\link{PEdom1D}}, \code{\link{PEdomtri}}, \code{\link{PEdomMTnd}},
and \code{\link{IndCSdomUBtri}}
}
