% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{dp2l}
\alias{dp2l}
\title{The distance from a point to a line}
\usage{
dp2l(p, a, b)
}
\arguments{
\item{p}{A 2D point, distance from p to the line passing through points a and b are to be computed.}

\item{a, b}{2D points that determine the straight line (i.e., through which the straight line is passing)}
}
\value{
A list with two elements
\item{dis}{Distance from point p to the line passing through a and b}
\item{cl2p}{The closest point on the line passing through a and b to the point p}
}
\description{
Returns the distance from a point p to the line joining points a and b in 2D space.
}
\examples{
A<-c(-1.22,-2.33); B<-c(2.55,3.75); C<-c(0,6); D<-c(3,-2)
#A<-runif(2); B<-runif(2); C<-runif(2); D<-runif(2)
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100); D<-runif(2,1,100)

ip<-int.2lines(A,B,C,D)
ip
pts<-rbind(A,B,C,D,ip)
xr<-range(pts)
xf<-abs(xr[2]-xr[1])*.1 #how far to go at the lower and upper ends in the x-coordinate
x<-seq(xr[1]-xf,xr[2]+xf,l=100)
lnAB<-Line(A,B,x)
lnCD<-Line(C,D,x)

y1<-lnAB$y
y2<-lnCD$y
Xlim<-range(x,pts)
Ylim<-range(y1,y2,pts)
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]
pf<-c(xd,-yd)*.025

#plot of the line joining A and B
plot(rbind(A,B,C,D),pch=1,xlab="x",ylab="y",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
lines(x,y1,lty=1,col=1)
lines(x,y2,lty=1,col=2)
text(rbind(A+pf,B+pf),c("A","B"))
text(rbind(C+pf,D+pf),c("C","D"))
text(rbind(ip+pf),c("intersection\\n point"))

int.2lines(c(1,2),c(1,3),C,D)

}
\seealso{
\code{\link{dp2pl}}, \code{\link{dist.pt2set}}  and \code{\link{Dist}}
}
