% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{dp2pl}
\alias{dp2pl}
\title{The distance from a point to a plane}
\usage{
dp2pl(p, a, b, c)
}
\arguments{
\item{p}{A 3D point, distance from p to the plane passing through points a, b and c are to be computed.}

\item{a, b, c}{3D points that determine the plane (i.e., through which the plane is passing)}
}
\value{
A list with two elements
\item{dis}{Distance from point p to the plane spanned by 3D points a, b and c}
\item{cl2p}{The closest point on the plane spanned by 3D points a, b and c to the point p}
}
\description{
Returns the distance from a point p to the plane passing through points a, b and c in 3D space.
}
\examples{
A<-c(1,10,3); B<-c(1,1,3); C<-c(3,9,12)
#A<-runif(3); B<-runif(3); C<-runif(3)
#A<-runif(3,1,100); B<-runif(3,1,100); C<-runif(3,1,100)
pts<-rbind(A,B,C)
Plane(A,B,C,.1,.2)

xr<-range(pts[,1]); yr<-range(pts[,2])
xf<-(xr[2]-xr[1])*.1 #how far to go at the lower and upper ends in the x-coordinate
yf<-(yr[2]-yr[1])*.1 #how far to go at the lower and upper ends in the y-coordinate
x<-seq(xr[1]-xf,xr[2]+xf,l=100)
y<-seq(yr[1]-yf,yr[2]+yf,l=100)

plABC<-Plane(A,B,C,x,y)
plABC
summary(plABC)
plot(plABC)

Plane(A,B,A+B,.1,.2)

z.grid<-plABC$z

persp(x,y,z.grid, xlab="x",ylab="y",zlab="z",theta = -30, phi = 30, expand = 0.5, col = "lightblue",
      ltheta = 120, shade = 0.05, ticktype = "detailed")

zr<-max(z.grid)-min(z.grid)
Pts<-rbind(A,B,C)+rbind(c(0,0,zr*.1),c(0,0,zr*.1),c(0,0,zr*.1))
Mn.pts<-apply(Pts,2,mean)

library(plot3D)
persp3D(z = z.grid, x = x, y = y, theta =225, phi = 30, ticktype = "detailed") #plane spanned by points A, B, C
#add the defining points
points3D(Pts[,1],Pts[,2],Pts[,3], add=TRUE)
text3D(Pts[,1],Pts[,2],Pts[,3], c("A","B","C"),add=TRUE)
text3D(Mn.pts[1],Mn.pts[2],Mn.pts[3],plABC$equation,add=TRUE)
polygon3D(Pts[,1],Pts[,2],Pts[,3], add=TRUE)

Plane(A,B,C,.1,.2)

}
\seealso{
\code{\link{dp2l}}, \code{\link{dist.pt2set}}  and \code{\link{Dist}}

#' @export dp2pl
}
