% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{funs10}
\alias{funs10}
\alias{muCS1D}
\alias{asyvarCS1D}
\title{Two functions of muCS1D and asyvarCS1D

The mean and (asymptotic) variance of arc density of Central Similarity Proximity Catch Digraph (CS-PCD)
for 1D data - middle interval case}
\usage{
muCS1D(t, c)

asyvarCS1D(t, c)
}
\arguments{
\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{c}{A positive real number in (0,1) parameterizing the center inside int}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a)}
}
\value{
x*y (prod1) or x+y (sum1).
}
\description{
muCS1D(t,c) returns the mean of the (arc) density of CS-PCD
and asyvarCS1D(r,c) returns the (asymptotic) variance of the arc density of CS-PCD
for a given centrality parameter c in (0,1) and an expansion parameter t>0 and 1D uniform data in a
finite interval (a,b), i.e. data from U(a,b) distribution.
See also (\insertCite{ceyhan:revstat-2016;textual}{pcds}).
}
\examples{
c<-.4
t<-2
a<-0; b<-10; int<-c(a,b)

n<-10
dat<-runif(n,a,b)
NumArcsCSmid1D(dat,t,c,int)

NumArcsCSmid1D(dat,t,c=.3,int)

NumArcsCSmid1D(dat,t=1.5,c,int)

NumArcsCSmid1D(dat,t,c,int+5)
NumArcsCSmid1D(dat,t,c,int+10)

n<-20
dat<-runif(n,a-5,b+5)
NumArcsCSint(dat,t,c,int)

dat<-runif(n,a+10,b+10)
NumArcsCSmid1D(dat,t,c,int)

n<-10
dat<-runif(n,a,b)
NumArcsCSmid1D(dat,t,c,int)

muCS1D(1.2,.4)
muCS1D(1.2,.6)

tseq<-seq(0.01,5,by=.05)
cseq<-seq(0.01,.99,by=.05)

ltseq<-length(tseq)
lcseq<-length(cseq)

mu.grid<-matrix(0,nrow=ltseq,ncol=lcseq)
for (i in 1:ltseq)
  for (j in 1:lcseq)
  {
    mu.grid[i,j]<-muCS1D(tseq[i],cseq[j])
  }

persp(tseq,cseq,mu.grid, xlab="t", ylab="c", zlab="mu(t,c)",theta = -30, phi = 30, expand = 0.5, col = "lightblue",
      ltheta = 120, shade = 0.05, ticktype = "detailed")

}
\references{
\insertAllCited{}

\insertAllCited{}
}
\seealso{
\code{\link{muPE1D}} and \code{\link{asyvarPE1D}}
}
