% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{funs11}
\alias{funs11}
\alias{muCSend1D}
\alias{asyvarCSend1D}
\title{Two functions of muCSend1D and asyvarCSend1D

The mean and (asymptotic) variance of arc density of Central Similarity Proximity Catch Digraph (CS-PCD)
for 1D data - end interval case}
\usage{
muCSend1D(t)

asyvarCSend1D(t)
}
\arguments{
\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}
}
\value{
muCSend1D returns the mean and asyvarCSend1D returns the asymptotic variance of the
arc density of CS-PCD for uniform data in end intervals

x*y (prod1) or x+y (sum1).
}
\description{
muCSend1D(t) returns the mean of the arc density of CS-PCD
and asyvarCSend1D(t) returns the asymptotic variance of the arc density of CS-PCD
for a given expansion parameter t>0  for 1D uniform data in the left and right end intervals
for the interval (a,b).
See also (\insertCite{ceyhan:revstat-2016;textual}{pcds}).
}
\examples{
a<-0; b<-10; int<-c(a,b)

n<-5
datL<-runif(n,a-5,a)
datR<-runif(n,b,b+5)
dat<-c(datL,datR)

NumArcsCSend1D(dat,t=2,int)

NumArcsCSend1D(dat,t=1.2,int)

NumArcsCSend1D(dat,t=4,int)

NumArcsCSend1D(dat,t=2,int+5)
NumArcsCSend1D(dat,t=2,int=c(-5,15))

n<-20
dat2<-runif(n,a-5,b+5)
NumArcsCSend1D(dat2,t=2,int)

t<-.5
NumArcsCSend1D(dat,t,int)

muCSend1D(1.2)

tseq<-seq(0.01,5,by=.05)
ltseq<-length(tseq)

mu.end<-vector()
for (i in 1:ltseq)
{
  mu.end<-c(mu.end,muCSend1D(tseq[i]))
}

par(mar = c(5,4,4,2) + 0.1)
plot(tseq, mu.end,type="l",ylab=expression(paste(mu,"(t)")),xlab="t",lty=1,xlim=range(tseq),ylim=c(0,1))

}
\references{
\insertAllCited{}

\insertAllCited{}
}
\seealso{
\code{\link{muPEend1D}} and \code{\link{asyvarPEend1D}}
}
