% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{funs4}
\alias{funs4}
\alias{cart2bary}
\alias{bary2cart}
\title{Two functions: cart2bary and bary2cart

The conversion of a point in Cartesian coordinates to Barycentric coordinates and vice versa}
\usage{
cart2bary(P, tri)

bary2cart(P, tri)
}
\arguments{
\item{P}{A 2D point for cart2bary, and a vector of three numeric entries for bary2cart}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}
}
\value{
cart2bary returns the barycentric coordinates of a given point P=(x,y) and
bary2cart returns the Cartesian coordinates of the point P=(t1,t2,t3) (not necessarily normalized)

x*y (prod1) or x+y (sum1).
}
\description{
cart2bary converts Cartesian coordinates of a given point P=(x,y) to barycentric coordinates
(in the normalized form) with respect to the triangle tri=(v1,v2,v3)
with vertex labeling done row-wise in tri (i.e., row i corresponds to vertex vi for i=1,2,3).
bary2cart converts barycentric coordinates of the point P=(t1,t2,t3) (not necessarily normalized) to
Cartesian coordinates according to the coordinates of the triangle, tri.
For information on barycentric coordinates, see (\insertCite{weisstein-barycentric;textual}{pcds})
}
\examples{
c1<-.4; c2<-.6
A<-c(0,0); B<-c(1,0); C<-c(c1,c2);

as.bastri(rbind(A,B,C))

as.bastri(rbind(B,C,A))

as.bastri(rbind(B,A,C))
as.bastri(rbind(A,C,B))

A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
as.bastri(rbind(A,B,C))
as.bastri(rbind(A,C,B))
as.bastri(rbind(B,A,C))

A<-runif(2); B<-runif(2); C<-runif(2)
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
as.bastri(rbind(A,B,C))

dat.fr<-data.frame(a=rbind(A,B,C))
as.bastri(dat.fr)

c1<-.4; c2<-.6
A<-c(0,0); B<-c(1,0); C<-c(c1,c2);

#A<-runif(2); B<-runif(2); C<-runif(2)
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);

Tr<-rbind(A,B,C)

cart2bary(A,Tr)
cart2bary(c(.3,.2),Tr)
cart2bary(c(.4,.2),Tr)
cart2bary(c(.5,.2),Tr)
cart2bary(c(.6,.2),Tr)

P<-c(.8,.2)
round(cart2bary(P,Tr),2)

P<-c(.5,.61)
cart2bary(P,Tr)

CM<-(A+B+C)/3
cart2bary(CM,Tr)

####
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
Tr<-rbind(A,B,C);
cart2bary(c(1.4,1.2),Tr)

cart2bary(c(.8,.2),Tr)

cart2bary(c(1.5,1.61),Tr)

cart2bary(c(1.5,1.61),rbind(Tr,Tr))

dat.fr<-data.frame(a=Tr)
cart2bary(c(.8,.2),dat.fr)

}
