% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{funs7}
\alias{funs7}
\alias{mu1PE1D}
\alias{fvar1}
\title{Two functions of muPE1D and asyvarPE1D

The mean and (asymptotic) variance of arc density of Proportional Edge Proximity Catch Digraph (PE-PCD)
for 1D data - middle interval case}
\usage{
mu1PE1D(r, c)

fvar1(r, c)
}
\arguments{
\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a)}
}
\value{
muPE1D returns the mean and asyvarPE1D returns the asymptotic variance of the
arc density of PE-PCD for U(a,b) data

x*y (prod1) or x+y (sum1).
}
\description{
muPE1D(r,c) returns the mean of the (arc) density of PE-PCD
and asyvarPE1D(r,c) returns the (asymptotic) variance of the arc density of PE-PCD
for a given centrality parameter c in (0,1) and an expansion parameter r>=1 and for 1D uniform data in a
finite interval (a,b), i.e. data from U(a,b) distribution.
muPE1D uses auxiliary (internal) function mu1PE1D(r,c) which yields mean (i.e., expected value)
of the arc density of PE-PCD for a given c in (0,1/2) and r>=1.
asyvarPE1D uses auxiliary (internal) functions fvar1 which yields asymptotic variance
of the arc density of PE-PCD for c in (1/4,1/2) and r>=1; and fvar2 which yields asymptotic variance
of the arc density of PE-PCD for c in (0,1/4) and r>=1.
See also (\insertCite{ceyhan:metrika-2012;textual}{pcds}).
}
\examples{
c<-.4
a<-0; b<-10; int<-c(a,b)
#c<-runif(1)
#a<-runif(1); b<-runif(1); int<-range(a,b); a<-int[1]; b<-int[2]
Mc<-centMc(int,c)

nx<-10
xr<-range(a,b,Mc)
xf<-(xr[2]-xr[1])*.5
dat<-runif(n,a,b)
#dat<-runif(n,a-xf,b+xf)
#dat<-runif(nx,a,b)+10

Ext<-cl2Mc.int(dat,int,c)
Ext
summary(Ext)
plot(Ext)

cl2Mc.int(dat[1],int,c)
cMc<-cl2Mc.int(dat,int,c)

Xlim<-range(a,b,dat)
xd<-Xlim[2]-Xlim[1]

plot(cbind(a,0),xlab="",pch=".",xlim=Xlim+xd*c(-.05,.05))
abline(h=0)
abline(v=c(a,b,Mc),col=c(1,1,2),lty=2)
points(cbind(dat,0))
points(cbind(c(cMc$Ext),0),pch=4,col=2)
text(cbind(c(a,b,Mc),-0.1),c("a","b","Mc"))

n<-20
dat<-runif(n,a-5,b+5)
cl2Mc.int(dat,c(a,b),c)

dat<-runif(n,a+b,b+10)
cl2Mc.int(dat,int,c)

c<-.4
a<-0; b<-10; int<-c(a,b)
n<-10
dat<-runif(n,a,b)
cl2Mc.int(dat,int,c)

muPE1D(1.2,.4)
muPE1D(1.2,.6)

rseq<-seq(1.01,5,by=.05)
cseq<-seq(0.01,.99,by=.05)

lrseq<-length(rseq)
lcseq<-length(cseq)

mu.grid<-matrix(0,nrow=lrseq,ncol=lcseq)
for (i in 1:lrseq)
  for (j in 1:lcseq)
  {
    mu.grid[i,j]<-muPE1D(rseq[i],cseq[j])
  }

persp(rseq,cseq,mu.grid, xlab="r", ylab="c", zlab="mu(r,c)", theta = -30, phi = 30, expand = 0.5, col = "lightblue",
      ltheta = 120, shade = 0.05, ticktype = "detailed")

}
\references{
\insertAllCited{}

\insertAllCited{}
}
\seealso{
\code{\link{muCS1D}} and \code{\link{asyvarCS1D}}
}
