% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{funs8}
\alias{funs8}
\alias{muPEend1D}
\alias{asyvarPEend1D}
\title{Two functions of muPEend1D and asyvarPEend1D

The mean and (asymptotic) variance of arc density of Proportional Edge Proximity Catch Digraph (PE-PCD)
for 1D data - end interval case}
\usage{
muPEend1D(r)

asyvarPEend1D(r)
}
\arguments{
\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >=1}
}
\value{
muPEend1D returns the mean and asyvarPEend1D returns the asymptotic variance of the
arc density of PE-PCD for uniform data in end intervals

x*y (prod1) or x+y (sum1).
}
\description{
muPEend1D(r) returns the mean of the arc density of PE-PCD
and asyvarPEend1D(r) returns the asymptotic variance of the arc density of PE-PCD
for a given expansion parameter r>=1 for 1D uniform data in the left and right end intervals
for the interval (a,b).
See also (\insertCite{ceyhan:metrika-2012;textual}{pcds}).
}
\examples{
a<-0; b<-10; int<-c(a,b)

n<-5
datL<-runif(n,a-5,a)
datR<-runif(n,b,b+5)
dat<-c(datL,datR)

r<-2
NumArcsPEend1D(dat,r,int)

NumArcsPEend1D(dat,r=1.2,int)
NumArcsPEend1D(dat,r=4,int)
NumArcsPEend1D(dat,r=.4,int)

n<-20
dat2<-runif(n,a-5,b+5)
NumArcsPEend1D(dat2,r,int)

NumArcsPEend1D(dat,r,int)

muPEend1D(1.2)

rseq<-seq(1.01,5,by=.05)
lrseq<-length(rseq)

mu.end<-vector()
for (i in 1:lrseq)
{
  mu.end<-c(mu.end,muPEend1D(rseq[i]))
}

plot(rseq, mu.end,type="l",ylab=expression(paste(mu,"(r)")),xlab="r",lty=1,xlim=range(rseq),ylim=c(0,1))

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{muCSend1D}} and \code{\link{asyvarCSend1D}}
}
