% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{funs9}
\alias{funs9}
\alias{muPE}
\alias{asyvarPE}
\title{Two functions of muPE and asyvarPE

The mean and (asymptotic) variance of arc density of Proportional Edge Proximity Catch Digraph (PE-PCD)
for 2D uniform data in one triangle}
\usage{
muPE(r)

asyvarPE(r)
}
\arguments{
\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >=1}
}
\value{
muPE returns the mean and asyvarPE returns the (asymptotic) variance of the
arc density of PE-PCD for uniform data in any triangle

x*y (prod1) or x+y (sum1).
}
\description{
muPE(r) returns the mean of the (arc) density of PE-PCD
and asyvarPE(r) returns the asymptotic variance of the arc density of PE-PCD
with expansion parameter r>=1 for 2D uniform data in a triangle.
PE proximity regions are defined with respect to the triangle and
vertex regions are based on center of mass, CM of the triangle.
See also (\insertCite{ceyhan:arc-density-PE;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
n<-20

#set.seed(1)
dat<-runif.tri(n,Tr)$g
#dat<-cbind(runif(n,1,2),runif(n,0,2))
#dat<-cbind(runif(n),runif(n))

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.0)
#M<-circ.cent.tri(Tr)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

Ext<-cl2eVRcent(dat,Tr,M)
Ext
summary(Ext)
plot(Ext)

cl2eVRcent(dat[1,],Tr,M)
cl2e<-cl2eVRcent(dat,Tr,M)
cl2e

Ds<-cp2e.tri(Tr,M)

Xlim<-range(Tr[,1],dat[,1])
Ylim<-range(Tr[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

if (dimension(M)==3) {M<-bary2cart(M,Tr)} #need to run this when M is given in barycentric coordinates

plot(Tr,pch=".",xlab="",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Tr)
points(dat,pch=1,col=1)
L<-rbind(M,M,M); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)
points(cl2e$Ext,pch=3,col=2)

xc<-Tr[,1]#+c(-.02,.03,.02)
yc<-Tr[,2]#+c(.02,.02,.04)
txt.str<-c("A","B","C")
text(xc,yc,txt.str)

txt<-rbind(M,Ds)
xc<-txt[,1]#+c(-.02,.05,-.02,-.01)
yc<-txt[,2]#+c(-.03,.02,.08,-.07)
txt.str<-c("M","D1","D2","D3")
text(xc,yc,txt.str)

cl2eVRcent(dat,Tr,M)

dat.fr<-data.frame(a=dat)
cl2eVRcent(dat.fr,Tr,M)

dat.fr<-data.frame(a=Tr)
cl2eVRcent(dat,dat.fr,M)

muPE(1.2)

rseq<-seq(1.01,5,by=.05)
lrseq<-length(rseq)

mu<-vector()
for (i in 1:lrseq)
{
  mu<-c(mu,muPE(rseq[i]))
}

plot(rseq, mu,type="l",xlab="r",ylab=expression(mu(r)),lty=1,xlim=range(rseq),ylim=c(0,1))

}
\references{
\insertAllCited{}

\insertAllCited{}
}
\seealso{
\code{\link{muCS}} and \code{\link{asyvarCS}}
}
