% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{Gam1ASbastri}
\alias{Gam1ASbastri}
\title{The indicator for a point being a dominating point for Arc Slice Proximity Catch Digraphs
(AS-PCDs) - basic triangle case}
\usage{
Gam1ASbastri(p, Dt, c1, c2, M = "CC", rv = NULL, ch.data.pnt = FALSE)
}
\arguments{
\item{p}{A 2D point that is to be tested for being a dominating point or not of the AS-PCD}

\item{Dt}{A set of 2D points which constitutes the vertices of the AS-PCD}

\item{c1, c2}{Positive real numbers which constitute the vertex of the basic triangle
adjacent to the shorter edges; c1 must be in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1}

\item{M}{"CC" which stand for circumcenter of the triangle Tb or a 2D point in Cartesian coordinates or
a 3D point in barycentric coordinates which serves as a center in the interior of the triangle Tb;
default is M="CC" i.e. the circumcenter of Tb}

\item{rv}{Index of the vertex whose region contains point p, rv takes the vertex labels as 1,2,3 as
in the row order of the vertices in Tb}

\item{ch.data.pnt}{A logical argument for checking whether point p is a data point in Dt or not
(default is FALSE)}
}
\value{
I(p is a dominating point of the 2D data set Dt for AS-PCD),
that is, returns 1 if p is a dominating point, returns 0 otherwise
}
\description{
Returns I(p is a dominating point of the 2D data set Dt for AS-PCD), that is, returns 1 if p is a dominating
point for AS-PCD, returns 0 otherwise. AS proximity regions are defined with respect to the basic triangle,
Tb, c1 is in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1.
Any given triangle can be mapped to the basic triangle by a combination of rigid body motions
(i.e., translation, rotation and reflection) and scaling, preserving uniformity of the points in the
original triangle. Hence basic triangle is useful for simulation
studies under the uniformness hypothesis. Vertex regions are based on the center M="CC" for circumcenter
of Tb; or M=(m1,m2) in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the
interior of Tb; default is M="CC".
Point, p, is in the vertex region of vertex rv (default is NULL); vertices are labeled as 1,2,3
in the order they are stacked row-wise.
ch.data.pnt is for checking whether point p is a data point in Dt or not (default is FALSE),
so by default this function checks whether the point p would be a dominating point
if it actually were in the data set.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
n<-20

#set.seed(1)
dat<-runif.tri(n,Tr)$g
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,2))

Ext<-cl2CC.VR(dat,Tr)
Ext
summary(Ext)
plot(Ext)

cl2CC.VR(dat[1,],Tr)

dat2<-rbind(dat,c(.2,.4))
cl2CC.VR(dat2,Tr)
cl2CC.VR(dat2,Tr,ch.all.intri = TRUE)

c2CC<-cl2CC.VR(dat,Tr)

CC<-circ.cent.tri(Tr) #the circumcenter
D1<-(B+C)/2; D2<-(A+C)/2; D3<-(A+B)/2;
Ds<-rbind(D1,D2,D3)

Xlim<-range(Tr[,1],dat[,1])
Ylim<-range(Tr[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(A,pch=".",asp=1,xlab="",ylab="",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Tr)
points(dat)
L<-matrix(rep(CC,3),ncol=3,byrow=T); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)
points(c2CC$Ext,pch=4,col=2)

txt<-rbind(Tr,CC,Ds)
xc<-txt[,1]+c(-.07,.08,.06,.12,-.1,-.1,-.09)
yc<-txt[,2]+c(.02,-.02,.03,.0,.02,.06,-.04)
txt.str<-c("A","B","C","CC","D1","D2","D3")
text(xc,yc,txt.str)

cl2CC.VR(dat,Tr)

cl2CC.VR(c(1.4,1.2),Tr)

cl2CC.VR(rbind(c("a","b"),dat),Tr)

dat.fr<-data.frame(a=dat)
cl2CC.VR(dat.fr,Tr)

dat.fr<-data.frame(a=Tr)
cl2CC.VR(dat,dat.fr)


}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{Gam1AStri}} and \code{\link{Gam1PEbastri}}
}
