% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{Gam1PEbastri}
\alias{Gam1PEbastri}
\title{The indicator for a point being a dominating point or not for Proportional Edge Proximity Catch
Digraphs (PE-PCDs) - basic triangle case}
\usage{
Gam1PEbastri(p, Dt, r, c1, c2, M = c(1, 1, 1), rv = NULL,
  ch.data.pnt = FALSE)
}
\arguments{
\item{p}{A 2D point that is to be tested for being a dominating point or not of the PE-PCD}

\item{Dt}{A set of 2D points which constitutes the vertices of the PE-PCD}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{c1, c2}{Positive real numbers which constitute the vertex of the basic triangle
adjacent to the shorter edges; c1 must be in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the basic triangle Tb or the circumcenter of Tb;
default is M=c(1,1,1) i.e. the center of mass of Tb}

\item{rv}{Index of the vertex whose region contains point p, rv takes the vertex labels as 1,2,3 as
in the row order of the vertices in Tb}

\item{ch.data.pnt}{A logical argument for checking whether point p is a data point in Dt or not
(default is FALSE)}
}
\value{
I(p is a dominating point of the 2D data set Dt for PE-PCD),
that is, returns 1 if p is a dominating point, returns 0 otherwise
}
\description{
Returns I(p is a dominating point of the 2D data set Dt for PE-PCD) for data in the basic triangle
Tb=T(c(0,0),c(1,0),c(c1,c2)), that is, returns 1 if p is a dominating point for PE-PCD, returns 0 otherwise.
PE proximity regions are defined with respect to the basic triangle Tb.
In the basic triangle, Tb, c1 is in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1.
Any given triangle can be mapped to the basic triangle by a combination of rigid body motions
(i.e., translation, rotation and reflection) and scaling, preserving uniformity of the points in the
original triangle. Hence basic triangle is useful for simulation
studies under the uniformness hypothesis. Vertex regions are based on center M=(m1,m2) in Cartesian
coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the interior of a basic triangle
to the edges on the extension of the lines joining M to the vertices or based on the circumcenter of Tb;
default is M=c(1,1,1) i.e. the center of mass of Tb.
Point, p, is in the vertex region of vertex rv (default is NULL); vertices are labeled as 1,2,3
in the order they are stacked row-wise.
ch.data.pnt is for checking whether point p is a data point in Dt or not (default is FALSE),
so by default this function checks whether the point p would be a dominating point
if it actually were in the data set.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:dom-num-NPE-Spat2011;textual}{pcds}).
}
\examples{
n<-10
dat<-cbind(runif(n),runif(n))

P<-dat[7,]
is.in.data(P,dat)
is.in.data(P,dat[7,])

P<-dat[7,]+10^(-7)
is.in.data(P,dat)

P<-dat[7,]+10^(-9)
is.in.data(P,dat)

is.in.data(P,P)

is.in.data(c(2,2),c(2,2))

is.in.data(P,vector())

###for 1D data
n<-10
dat<-runif(n)

P<-dat[7]
is.in.data(P,dat)  #this does not work because both entries are treated as vectors of different dimensions
is.in.data(P,dat[7]) #this works because both entries are treated as 1D vectors

dat<-as.matrix(dat)
is.in.data(P,dat) #this works, because P is a 1D point, and dat is treated as a set of 10 1D points

P<-dat[7]+10^(-7)
is.in.data(P,dat)

P<-dat[7]+10^(-9)
is.in.data(P,dat)

is.in.data(P,P)

is.in.data(P,vector())

###for 3D data
n<-10
dat<-cbind(runif(n),runif(n),runif(n))

P<-dat[7,]
is.in.data(P,dat)
is.in.data(P,dat[7,])

P<-dat[7,]+10^(-7)
is.in.data(P,dat)

P<-dat[7,]+10^(-9)
is.in.data(P,dat)

is.in.data(P,P)

n<-10
dat<-cbind(runif(n),runif(n))
P<-dat[7,]
is.in.data(P,dat)

dat.fr<-data.frame(a=dat)
is.in.data(P,dat.fr)


}
\seealso{
\code{\link{Gam1ASbastri}} and \code{\link{Gam1AStri}}
}
