% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{Gam1PEtri}
\alias{Gam1PEtri}
\title{The indicator for a point being a dominating point for Proportional Edge Proximity Catch Digraphs (PE-PCDs)
- one triangle case}
\usage{
Gam1PEtri(p, Dt, tri, r, M = c(1, 1, 1), rv = NULL,
  ch.data.pnt = FALSE)
}
\arguments{
\item{p}{A 2D point that is to be tested for being a dominating point or not of the PE-PCD}

\item{Dt}{A set of 2D points which constitutes the vertices of the PE-PCD}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri or the circumcenter of tri;
default is M=c(1,1,1) i.e. the center of mass of tri}

\item{rv}{Index of the vertex whose region contains point p, rv takes the vertex labels as 1,2,3 as
in the row order of the vertices in tri}

\item{ch.data.pnt}{A logical argument for checking whether point p is a data point
in Dt or not (default is FALSE)}
}
\value{
I(p is a dominating point of the 2D data set Dt for PE-PCD),
that is, returns 1 if p is a dominating point, returns 0 otherwise
}
\description{
Returns I(p is a dominating point of the 2D data set Dt for PE-PCD) in the triangle tri, that is,
returns 1 if p is a dominating point for PE-PCD, returns 0 otherwise.
Point, p, is in the vertex region of vertex rv (default is NULL); vertices are labeled as 1,2,3
in the order they are stacked row-wise in tri.
PE proximity region is constructed with respect to the triangle tri with expansion parameter r>=1
and vertex regions are based on center M=(m1,m2) in Cartesian coordinates or
M=(alpha,beta,gamma) in barycentric coordinates in the interior of the triangle tri
or based on the circumcenter of tri; default is M=c(1,1,1) i.e. the center of mass of tri.
ch.data.pnt is for checking whether point p is a data point in Dt or not (default is FALSE),
so by default this function checks whether the point p would be a dominating point
if it actually were in the data set.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:masa-2007,ceyhan:dom-num-NPE-Spat2011,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
n<-10

#set.seed(1)
dat<-runif.tri(n,Tr)$g
#dat<-cbind(runif(n,1,2),runif(n,0,2))
#dat<-cbind(runif(n),runif(n))

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.0)
#M<-circ.cent.tri(Tr)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

#dat<-c(.5,.5)
IM<-IncMatPEtri(dat,Tr,r=1.25,M)

IM
dom.greedy(IM)
dom.exact(IM)
IndUBdom(IM,2)
IndUBdom(IM,3)

IncMatPEtri(dat,Tr,r=1.5,M)
IncMatPEtri(dat,Tr,r=2,M)

r<-2
IncMatPEtri(dat,Tr,r,M)

dat.fr<-data.frame(a=dat)
IncMatPEtri(dat.fr,Tr,r,M)

dat.fr<-data.frame(a=Tr)
IncMatPEtri(dat,dat.fr,r,M)


}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{Gam1PEbastri}} and \code{\link{Gam1AStri}}
}
