% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{in.triangle}
\alias{in.triangle}
\title{Check whether a point is inside a triangle}
\usage{
in.triangle(p, tri, boundary = FALSE)
}
\arguments{
\item{p}{A 2D point to be checked whether it is inside the triangle or not.}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{boundary}{A logical parameter (default=FALSE) to include boundary or not, so if it is TRUE,
the function checks if the point, p, lies in the closure of the triangle (i.e. interior and boundary combined)
else it checks if p lies in the interior of the triangle.}
}
\value{
A list with two elements
\item{inside.tri}{A logical output, it is TRUE, if the point, p, is inside the triangle, tri,
else it is FALSE.}
\item{barycentric}{The barycentric coordinates (alpha, beta, gamma) of the point p with respect to the triangle, tri.}
}
\description{
Checks if the point p lies in the triangle, tri, using the barycentric coordinates, generally denoted as
(alpha,beta,gamma). If all (normalized or non-normalized) barycentric coordinates are positive then the point p is inside the triangle,
if all are nonnegative with one or more are zero, then p falls in the boundary. If some of the barycentric
coordinates are negative, then p falls outside the triangle.
boundary is a logical argument (default=FALSE) to include boundary or not, so if it is TRUE,
the function checks if the point, p, lies in the closure of the triangle (i.e. interior and boundary combined)
else it checks if p lies in the interior of the triangle.
}
\examples{
cent<-c(1,1); rad<-1; p<-c(1.4,1.2)

#cent<-runif(2); rad<-runif(1); p<-runif(2);
in.circle(p,cent,rad)

p<-c(.4,-.2)
in.circle(p,cent,rad)

p<-c(1,0)
in.circle(p,cent,rad)
in.circle(p,cent,rad,boundary=TRUE)

#for a NA entry point
p<-c(NA,.2)
in.circle(p,cent,rad)

dat.fr<-data.frame(a=Tr)
in.triangle(p,dat.fr)

}
\seealso{
\code{\link{inTriAll}} and \code{\link{in.convex.hull}} from the tripack package
}
