% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IncMatASMT}
\alias{IncMatASMT}
\title{Incidence matrix for Arc Slice Proximity Catch Digraphs (AS-PCDs) - multiple triangle case}
\usage{
IncMatASMT(Xp, Yp, M = "CC")
}
\arguments{
\item{Xp}{A set of 2D points which constitute the vertices of the AS-PCD}

\item{Yp}{A set of 2D points which constitute the vertices of the Delaunay triangles}

\item{M}{"CC" which stand for circumcenter of each Delaunay triangle tri or 3D point in barycentric
coordinates which serves as a center in the interior of each Delaunay triangle;
default is M="CC" i.e. the circumcenter of each triangle}
}
\value{
Incidence matrix for the AS-PCD whose vertices are the 2D data set, dat,
and AS proximity regions are defined in the Delaunay triangles based on Yp points.
}
\description{
Returns the incidence matrix for the AS-PCD whose vertices are a given 2D numerical data set, Xp,
in the convex hull of Yp which is partitioned by the Delaunay triangles based on Yp points.
AS proximity regions are defined with respec to the Delaunay triangles based on Yp points and vertex
regions are based on the center M="CC"
for circumcenter of each Delaunay triangle or M=(alpha,beta,gamma) in barycentric coordinates in the
interior of each Delaunay triangle; default is M="CC" i.e. circumcenter of each triangle.
Loops are allowed, so the diagonal entries are all equal to 1.
See (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}) for more on AS-PCDs.
Also see (\insertCite{okabe:2000,ceyhan:comp-geo-2010,renka:1996;textual}{pcds}) for more on Delaunay triangulation
and the tripack algorithm.
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
n<-10

#set.seed(1)
dat<-runif.tri(n,Tr)$g
#dat<-cbind(runif(n,1,2),runif(n,0,2))
#dat<-cbind(runif(n),runif(n))
#dat<-c(.5,.5)

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.2)
#M<-circ.cent.tri(Tr); #M<-"CC"; #M<-"CM"
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

IM<-IncMatAStri(dat,Tr,M)
IM

dom.greedy(IM)
dom.exact(IM)

dat.fr<-data.frame(a=dat)
IncMatAStri(dat.fr,Tr,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IncMatAStri}}, \code{\link{IncMatPEMT}}, and \code{\link{IncMatCSMT}}
}
