% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IncMatCS1D}
\alias{IncMatCS1D}
\title{Incidence matrix for Central Similarity Proximity Catch Digraphs (CS-PCDs)
for 1D data - multiple interval case}
\usage{
IncMatCS1D(Xp, Yp, t, c)
}
\arguments{
\item{Xp}{a set of 1D points which constitutes the vertices of the digraph}

\item{Yp}{a set of 1D points which constitutes the end points of the intervals
that partition the real line.}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a)}
}
\value{
Incidence matrix for the CS-PCD with vertices being 1D data set, Xp,
and Yp determines the end points of the intervals (the multi-interval case)
}
\description{
Returns the incidence matrix for the CS-PCD for a given 1D numerical data set, Xp,
as the vertices of the digraph and Yp determines the end points of the intervals (in the multi-interval case).
Loops are allowed, so the diagonal entries are all equal to 1. CS proximity region is constructed
with an expansion parameter t>0 and a centrality parameter c in (0,1).
See also (\insertCite{ceyhan:revstat-2016;textual}{pcds}).
}
\examples{
t<-2
c<-.4
a<-0; b<-10;
#c<-runif(1)
#a<-runif(1); b<-runif(1); int<-range(a,b); a<-int[1]; b<-int[2]
nx<-40; ny<-10

#set.seed(1)
xr<-range(a,b)
xf<-(xr[2]-xr[1])*.1
Xp<-runif(nx,a-xf,b+xf)
#Xp<-runif(nx,a-5,b+5)
#Xp<-runif(nx,a+5,b+5)
#Xp<-runif(nx,a+10,b+10)
Yp<-runif(ny,a,b)

Arcs<-ArcsCSMI(Xp,Yp,t,c)
Arcs
summary(Arcs)
plot(Arcs)

S<-Arcs$S
E<-Arcs$E

ArcsCSMI(Xp,Yp,t,c)

ArcsCSMI(Xp,Yp+10,t,c)

jit<-.1
yjit<-runif(nx,-jit,jit)

Xlim<-range(a,b,Xp,Yp)
xd<-Xlim[2]-Xlim[1]

plot(cbind(a,0),main="arcs of CS-PCD for points (jittered along y-axis)\\n in middle intervals ", xlab=" ", ylab=" ",
     xlim=Xlim+xd*c(-.05,.05),ylim=3*c(-jit,jit),pch=".")
abline(h=0,lty=1)
points(Xp, yjit,pch=".",cex=3)
abline(v=Yp,lty=2)
arrows(S, yjit, E, yjit, length = .05, col= 4)

t<-2
c<-.4
a<-0; b<-10;
nx<-40; ny<-10
Xp<-runif(nx,a,b)
Yp<-runif(ny,a,b)

ArcsCSMI(Xp,Yp,t,c)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IncMatCS1D}}, \code{\link{IncMatPEtri}}, and \code{\link{IncMatPEMT}}
}
