% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IncMatCSMT}
\alias{IncMatCSMT}
\title{Incidence matrix for Central Similarity Proximity Catch Digraphs (CS-PCDs) - multiple triangle case}
\usage{
IncMatCSMT(Xp, Yp, t, M = c(1, 1, 1))
}
\arguments{
\item{Xp}{A set of 2D points which constitute the vertices of the CS-PCD}

\item{Yp}{A set of 2D points which constitute the vertices of the Delaunay triangles}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{M}{A 3D point in barycentric coordinates which serves as a center in the interior of each Delaunay
triangle, default for M is c(1,1,1) which is the center of mass of each triangle}
}
\value{
Incidence matrix for the CS-PCD with vertices being 2D data set, Xp.
CS proximity regions are constructed with respect to the Delaunay triangles and M-edge regions.
}
\description{
Returns the incidence matrix of Central Similarity Proximity Catch Digraph (CS-PCD) whose vertices are the
data points in Xp in the multiple triangle case. CS proximity regions are defined with respect to the
Delaunay triangles based on Yp points with expansion parameter t>0 and edge regions in each triangle are
based on the center M=(alpha,beta,gamma) in barycentric coordinates in the interior of each Delaunay
triangle (default for M is c(1,1,1) which is the center of mass of the triangle).
Each Delaunay triangle is first converted to an (unscaled) basic triangle so that M will be the same
type of center for each Delaunay triangle (this conversion is not necessary when M is CM).
Convex hull of Yp is partitioned by the Delaunay triangles based on Yp points
(i.e., multiple triangles are the set of these Delaunay triangles whose union constitutes the
convex hull of Yp points). For the incidence matrix loops are allowed,
so the diagonal entries are all equal to 1.
See (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-CS,ceyhan:test2014;textual}{pcds}) for more on CS-PCDs.
Also see (\insertCite{okabe:2000,ceyhan:comp-geo-2010,renka:1996;textual}{pcds}) for more on Delaunay triangulation and the tripack algorithm.
}
\examples{
nx<-100
ny<-10

#set.seed(1)
Xp<-cbind(runif(nx,0,1),runif(nx,0,1))
#Xp<-cbind(runif(nx,0.5,1.5),runif(nx,0,1))
#Xp<-cbind(runif(nx,1,2),runif(nx,0,1))
Yp<-cbind(runif(ny,0,1),runif(ny,0,1))
#Xp<-c(.5,.5)

M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)
t<-1.5
#t<-2

ArcsCSMT(Xp,Yp,t,M)

Arcs<-ArcsCSMT(Xp,Yp,t,M)
Arcs
summary(Arcs)
plot(Arcs)

S<-Arcs$S
E<-Arcs$E
DT<-tri.mesh(Yp[,1],Yp[,2],duplicate="remove")

Xlim<-range(Xp[,1],Yp[,1])
Ylim<-range(Xp[,2],Yp[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Xp,main=" ", xlab=" ", ylab=" ",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05),pch=".",cex=3)
plot.tri(DT, add=T, do.points=T)
arrows(S[,1], S[,2], E[,1], E[,2], length = 0.1, col= 4)

ArcsCSMT(Xp,Yp,t,M)

ArcsCSMT(Xp,rbind(Yp,Yp),t,M)

dat.fr<-data.frame(a=Xp)
ArcsCSMT(dat.fr,Yp,t,M)

dat.fr<-data.frame(a=Yp)
ArcsCSMT(Xp,dat.fr,t,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IncMatCStri}}, \code{\link{IncMatCSTe}}, \code{\link{IncMatASMT}},
and \code{\link{IncMatPEMT}}
}
