% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IncMatCSTe}
\alias{IncMatCSTe}
\title{Incidence matrix for Central Similarity Proximity Catch Digraphs (CS-PCDs) - standard
equilateral triangle case}
\usage{
IncMatCSTe(dat, t, M = c(1, 1, 1))
}
\arguments{
\item{dat}{A set of 2D points which constitute the vertices of the CS-PCD}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the standard equilateral triangle Te; default is M=c(1,1,1) i.e.
the center of mass of Te}
}
\value{
Incidence matrix for the CS-PCD with vertices being 2D data set, dat and CS proximity
regions are defined in the standard equilateral triangle Te with M-edge regions.
}
\description{
Returns the incidence matrix for the CS-PCD whose vertices are the given 2D numerical data set, dat,
in the standard equilateral triangle Te=T(v=1,v=2,v=3)=T((0,0),(1,0),(1/2,sqrt(3)/2)).
CS proximity region is defined with respect to the standard equilateral triangle
Te=T(v=1,v=2,v=3)=T((0,0),(1,0),(1/2,sqrt(3)/2)) and edge regions are based on the center M=(m1,m2)
in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the interior of Te;
default is M=c(1,1,1) i.e. the center of mass of Te.
 Loops are allowed, so the diagonal entries are all equal to 1.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-CS,ceyhan:test2014;textual}{pcds}).
}
\examples{
A<-c(0,0); B<-c(1,0); C<-c(.5,sqrt(3.)/2);
Te<-rbind(A,B,C);
n<-10

#set.seed(1)
dat<-runifTe(n)$gen.points

M<-as.numeric(runifTe(1)$g)
#M<-c(.6,.2)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)
t<-1

IndNCSTe(dat[1,],dat[2,],t,M)
IndNCSTe(dat[1,],dat[1,],t,M)

IndNCSTe(dat[1,],dat[2,],t=4,M)
IndNCSTe(dat[2,],dat[5,],t,M)
IndNCSTe(c(.2,.5),dat[2,],t,M)

IndNCSTe(c(.2,.5),c(.2,.5),t,M)

#or try
re<-reTeCM(dat[1,])$re
IndNCSTe(dat[1,],dat[2,],t,M,re=re)

IndNCSTe(dat[1,],dat[2,],t,M)


}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IncMatCStri}}, \code{\link{IncMatCSMT}} and \code{\link{IncMatPETe}}
}
