% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IncMatCStri}
\alias{IncMatCStri}
\title{Incidence matrix for Central Similarity Proximity Catch Digraphs (CS-PCDs) - one triangle case}
\usage{
IncMatCStri(dat, tri, t, M = c(1, 1, 1))
}
\arguments{
\item{dat}{A set of 2D points which constitute the vertices of CS-PCD}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri;
default is M=c(1,1,1) i.e. the center of mass of tri}
}
\value{
Incidence matrix for the CS-PCD with vertices being 2D data set, dat,
in the triangle tri with edge regions based on center M
}
\description{
Returns the incidence matrix for the CS-PCD whose vertices are the given 2D numerical data set, dat,
in the triangle tri=T(v=1,v=2,v=3). CS proximity regions are constructed with respect to triangle tri
with expansion parameter t>0 and edge regions are based on the center M=(m1,m2) in Cartesian coordinates
or M=(alpha,beta,gamma) in barycentric coordinates in the interior of the triangle tri;
default is M=c(1,1,1) i.e. the center of mass of tri.
Loops are allowed, so the diagonal entries are all equal to 1.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-CS,ceyhan:test2014;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
n<-10

#set.seed(1)
dat<-runif.tri(n,Tr)$g
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,2))
#dat<-c(1.4,1)

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.0)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)
t<-1.5
#t<-2

ArcsCStri(dat,Tr,t,M)

Arcs<-ArcsCStri(dat,Tr,t,M)
Arcs
summary(Arcs)
plot(Arcs)

S<-Arcs$S
E<-Arcs$E

Xlim<-range(Tr[,1],dat[,1])
Ylim<-range(Tr[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

Ds<-cp2e.tri(Tr,M)

if (dimension(M)==3) {M<-bary2cart(M,Tr)} #need to run this when M is given in barycentric coordinates

plot(Tr,pch=".",xlab="",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Tr)
points(dat,pch=1,col=1)
L<-rbind(M,M,M); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)
arrows(S[,1], S[,2], E[,1], E[,2], length = 0.1, col= 4)

txt<-rbind(Tr,M,Ds)
xc<-txt[,1]#+c(-.02,.03,.02,.03,.04,-.03,-.01)
yc<-txt[,2]#+c(.02,.02,.03,.06,.03,.05,-.07)
txt.str<-c("A","B","C","M","D1","D2","D3")
text(xc,yc,txt.str)

t<-2
ArcsCStri(dat,Tr,t,M)

dat.fr<-data.frame(a=dat)
ArcsCStri(dat.fr,Tr,t,M)

dat.fr<-data.frame(a=Tr)
ArcsCStri(dat,dat.fr,t,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IncMatCSMT}}, \code{\link{IncMatPEtri}}, and  \code{\link{IncMatAStri}}
}
