% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IncMatPEtetra}
\alias{IncMatPEtetra}
\title{Incidence matrix for Proportional Edge Proximity Catch Digraphs (PE-PCDs) - one tetrahedron case}
\usage{
IncMatPEtetra(dat, th, r, M = "CM")
}
\arguments{
\item{dat}{A set of 3D points which constitute the vertices of PE-PCD}

\item{th}{Four 3D points, stacked row-wise, each row representing a vertex of the tetrahedron}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{M}{The center to be used in the construction of the vertex regions in the tetrahedron, th.
Currently it only takes "CC" for circumcenter and "CM" for center of mass; default="CM"}
}
\value{
Incidence matrix for the PE-PCD with vertices being 3D data set, dat,
in the tetrahedron th with vertex regions based on circumcenter or center of mass
}
\description{
Returns the incidence matrix for the PE-PCD whose vertices are the given 3D numerical data set, dat,
in the tetrahedron th=T(v=1,v=2,v=3,v=4). PE proximity regions are constructed with respect to tetrahedron
th with expansion parameter r>=1 and vertex regions are based on the center M which is circumcenter ("CC")
or center of mass ("CM") of th with default="CM".
Loops are allowed, so the diagonal entries are all equal to 1.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010;textual}{pcds}).
}
\examples{
A<-c(0,0,0); B<-c(1,0,0); C<-c(1/2,sqrt(3)/2,0); D<-c(1/2,sqrt(3)/6,sqrt(6)/3)
#A<-runif(3); B<-runif(3); C<-runif(3); D<-runif(3)
#A<-runif(3,1,100); B<-runif(3,1,100); C<-runif(3,1,100); D<-runif(3,1,100)
tetra<-rbind(A,B,C,D)
n<-20
dat<-runif.tetra(n,tetra)$g

M<-"CM"
#M<-"CC"
r<-1.5

IndNPEtetra(dat[1,],dat[2,],r,tetra) #uses the default M="CM"
IndNPEtetra(dat[1,],dat[2,],r,tetra,M)

IndNPEtetra(dat[1,],dat[1,],r,tetra,M)

IndNPEtetra(c(.4,.4,.4),c(.5,.5,.5),r,tetra,M)

#or try
RV<-rv.tetraCC(dat[1,],tetra)$rv
IndNPEtetra(dat[1,],dat[5,],r,tetra,M,rv=RV)

IndNPEtetra(dat[1,],c(-1,-1,-1),r,tetra,M,rv=NULL)

IndNPEtetra(c(-1,-1,-1),dat[1,],r,tetra,M,rv=NULL)
IndNPEtetra(c(-1,-1,-1),c(-1,-1,-1),r,tetra,M)

P1<-c(.1,.1,.1)
P2<-c(.5,.5,.5)
IndNPEtetra(P1,P2,r,tetra,M)


}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IncMatPEtri}}, \code{\link{IncMatPE1D}}, and  \code{\link{IncMatPEMT}}
}
