% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IncMatPEtri}
\alias{IncMatPEtri}
\title{Incidence matrix for Proportional Edge Proximity Catch Digraphs (PE-PCDs) - one triangle case}
\usage{
IncMatPEtri(dat, tri, r, M = c(1, 1, 1))
}
\arguments{
\item{dat}{A set of 2D points which constitute the vertices of PE-PCD}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >=1}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri or the circumcenter of tri;
default is M=c(1,1,1) i.e. the center of mass of tri}
}
\value{
Incidence matrix for the PE-PCD with vertices being 2D data set, dat,
in the triangle tri with vertex regions based on center M
}
\description{
Returns the incidence matrix for the PE-PCD whose vertices are the given 2D numerical data set, dat,
in the triangle tri=T(v=1,v=2,v=3). PE proximity regions are constructed with respect to triangle tri
with expansion parameter r>=1 and vertex regions are based on the center M=(m1,m2) in Cartesian coordinates
or M=(alpha,beta,gamma) in barycentric coordinates in the interior of the triangle tri;
default is M=c(1,1,1) i.e. the center of mass of tri.
Loops are allowed, so the diagonal entries are all equal to 1.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-PE,ceyhan:dom-num-NPE-Spat2011;textual}{pcds}).
}
\examples{
nx<-100
ny<-10

#set.seed(1)
Xp<-cbind(runif(nx),runif(nx))
#Xp<-cbind(runif(nx,.5,1.5),runif(nx))
#Xp<-cbind(runif(nx,1,2),runif(nx))
Yp<-cbind(runif(ny),runif(ny))

plotDeltri(Xp,Yp,xlab="",ylab="")

NumArcsPEMT(Xp,Yp,r=1.25)
TSArcDensPEMT(Xp,Yp,r=1.25)
TSArcDensPEMT(Xp,Yp,r=1.25,ch=TRUE)

NumArcsPEMT(Xp,Yp,r=1.5)
TSArcDensPEMT(Xp,Yp,r=1.5,alt="l")
TSArcDensPEMT(Xp,Yp,r=1.5,ch=TRUE,alt="l")

NumArcsPEMT(Xp,Yp,r=2)
TSArcDensPEMT(Xp,Yp,r=2)
TSArcDensPEMT(Xp,Yp,r=2,ch=TRUE)

TSArcDensPEMT(Xp1,Yp1,r)

TSArcDensPEMT(Xp1,rbind(Yp1,Yp1),r)

dat.fr<-data.frame(a=Xp1)
TSArcDensPEMT(dat.fr,Yp1,r)

dat.fr<-data.frame(a=Yp1)
TSArcDensPEMT(Xp1,dat.fr,r)


}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IncMatPEMT}}, \code{\link{IncMatCStri}}, and  \code{\link{IncMatAStri}}
}
