% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{ind.int.set}
\alias{ind.int.set}
\title{Indices of the intervals where the 1D point(s) reside}
\usage{
ind.int.set(dat, Yp)
}
\arguments{
\item{dat}{A set of 1D points for which the indices of intervals are to be determined}

\item{Yp}{A set of 1D points from which intervals are constructed}
}
\value{
The vector of indices of the intervals in which points in the 1D data set, dat, reside
}
\description{
Returns the indices of intervals for all the points in 1D data set, Dt, as a vector.
Intervals are based on Yp and left end interval is labeled as 0, the next interval as 1, and so on.
}
\examples{
c<-.5

Pg2PE1D.asy(c)

Pg2PE1D.asy(c=1/1.5)
Pg2PE1D(r=1.5,c=1/1.5,n=10)
Pg2PE1D(r=1.5,c=1/1.5,n=100)

}
