% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IndASdomUBtri}
\alias{IndASdomUBtri}
\title{Indicator for an upper bound for the domination number of Arc Slice Proximity Catch Digraph
(AS-PCD) by the exact algorithm - one triangle case}
\usage{
IndASdomUBtri(Dt, k, tri, M = "CC")
}
\arguments{
\item{Dt}{A set of 2D points which constitute the vertices of the AS-PCD}

\item{k}{A positive integer to be tested for an upper bound for the domination number of AS-PCDs}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{M}{"CC" which stand for circumcenter of the triangle tri or a 2D point in Cartesian coordinates or
a 3D point in barycentric coordinates which serves as a center in the interior of tri;
default is M="CC" i.e. the circumcenter of tri}
}
\value{
A list with the elements
\item{domUB}{The suggested upper bound (to be checked) for the domination number of AS-PCD.
It is prespecified as k in the function arguments.}
\item{IndUBdom}{The indicator for the upper bound for domination number of AS-PCD being the
specified value k or not. It returns 1 if the upper bound is k, and 0 otherwise.}
\item{ind.domset}{The vertices (i.e., data points) in the dominating set of size k if it exists,
otherwise it yields NULL.}
}
\description{
Returns I(domination number of AS-PCD whose vertices are the data points Dt is less than or equal to k),
that is, returns 1 if the domination number of CS-PCD is less than the prespecified value k, returns 0
otherwise. It also provides the vertices (i.e. data points) in a dominating set of size k of AS-PCD.
AS proximity regions are constructed with respect to the triangle tri and
vertex regions are based on the center M="CC" for circumcenter of tri;
or M=(m1,m2) in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the
interior of the triangle tri; default is M="CC" i.e. circumcenter of tri. The vertices of triangle, tri,
are labeled as 1,2,3 according to the row number the vertex is recorded in tri.
Loops are allowed in the digraph.
It takes a long time for large number of vertices (i.e., large number of row numbers).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
n<-10

#set.seed(1)
dat<-runif.tri(n,Tr)$gen.points
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.2)
#M<-circ.cent.tri(Tr); #M<-"CC"; #M<-"CM"
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

S<-rbind(dat[1,],dat[2,])
IndNAStri.domset(S,dat,Tr,M)

S<-rbind(dat[1,],dat[2,],dat[3,],dat[5,])
IndNAStri.domset(S,dat,Tr,M)

S<-rbind(c(.1,.1),c(.3,.4),c(.5,.3))
IndNAStri.domset(S,dat,Tr,M)

IndNAStri.domset(c(.2,.5),dat,Tr,M)
IndNAStri.domset(c(.2,.5),c(.2,.5),Tr,M)
IndNAStri.domset(dat[5,],dat[2,],Tr,M)

S<-rbind(dat[1,],dat[2,],dat[3,],dat[5,],c(.2,.5))
IndNAStri.domset(S,dat[3,],Tr,M)

IndNAStri.domset(dat,dat,Tr,M)

P<-c(.4,.2)
S<-dat[c(1,3,4),]
IndNAStri.domset(dat,P,Tr,M)
IndNAStri.domset(S,P,Tr,M)
IndNAStri.domset(S,dat,Tr,M)

IndNAStri.domset(rbind(S,S),dat,Tr,M)

dat.fr<-data.frame(a=dat)
IndNAStri.domset(S,dat.fr,Tr,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IndCSdomUBtri}}, \code{\link{IndCSdomUBTe}}, \code{\link{IndUBdom}},
and \code{\link{dom.exact}}
}
