% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IndCSdomUBTe}
\alias{IndCSdomUBTe}
\title{The indicator for k being an upper bound for the domination number of Central Similarity Proximity
Catch Digraph (CS-PCD) by the exact algorithm - standard equilateral triangle case}
\usage{
IndCSdomUBTe(Dt, k, t, M = c(1, 1, 1))
}
\arguments{
\item{Dt}{A set of 2D points which constitute the vertices of CS-PCD}

\item{k}{A positive integer representing an upper bound for the domination number of CS-PCD}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region in the
standard equilateral triangle Te=T((0,0),(1,0),(1/2,sqrt(3)/2))}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the standard equilateral triangle Te; default is M=c(1,1,1) i.e.
the center of mass of Te}
}
\value{
A list with two elements
\item{domUB}{The upper bound k (to be checked) for the domination number of CS-PCD. It is prespecified
as k in the function arguments.}
\item{IndUBdom}{The indicator for the upper bound for domination number of CS-PCD being the
specified value k or not. It returns 1 if the upper bound is k, and 0 otherwise.}
\item{ind.domset}{The vertices (i.e., data points) in the dominating set of size k if it exists,
otherwise it is NULL.}
}
\description{
Returns I(domination number of CS-PCD whose vertices are the data points Dt is less than or equal to k),
that is, returns 1 if the domination number of CS-PCD is less than the prespecified value k, returns 0
otherwise. It also provides the vertices (i.e. data points) in a dominating set of size k of CS-PCD.
CS proximity region is constructed with respect to the standard equilateral triangle Te=T(A,B,C)=T((0,0),(1,0),(1/2,sqrt(3)/2)) with
expansion parameter t>0 and edge regions are based on the center M=(m1,m2)
in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the interior of Te;
default is M=c(1,1,1) i.e. the center of mass of Te (which is equivalent to the circumcenter of Te).
Edges of Te, AB, BC, AC, are also labeled as 3, 1, and 2, respectively.
Loops are allowed in the digraph.
It takes a long time for large number of vertices (i.e., large number of row numbers).
See also (\insertCite{ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
n<-10

#set.seed(1)
dat<-runif.tri(n,Tr)$gen.points
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.0)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)
t<-.5

S<-rbind(dat[1,],dat[2,])
IndNCStri.domset(S,dat,t,Tr,M)

S<-rbind(dat[1,],dat[2,],dat[3,],dat[5,])
IndNCStri.domset(S,dat,t,Tr,M)

S<-rbind(c(.1,.1),c(.3,.4),c(.5,.3))
IndNCStri.domset(S,dat,t,Tr,M)

IndNCStri.domset(c(.2,.5),dat,t,Tr,M)
IndNCStri.domset(c(.2,.5),c(.2,.5),t,Tr,M)
IndNCStri.domset(dat[5,],dat[2,],t,Tr,M)

S<-rbind(dat[1,],dat[2,],dat[3,],dat[5,],c(.2,.5))
IndNCStri.domset(S,dat[3,],t,Tr,M)

IndNCStri.domset(dat,dat,t,Tr,M)

P<-c(.4,.2)
S<-dat[c(1,3,4),]
IndNCStri.domset(dat,P,t,Tr,M)
IndNCStri.domset(S,P,t,Tr,M)
IndNCStri.domset(S,dat,t,Tr,M)

IndNCStri.domset(rbind(S,S),dat,t,Tr,M)

dat.fr<-data.frame(a=dat)
IndNCStri.domset(S,dat.fr,t,Tr,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IndCSdomUBtri}}, \code{\link{IndUBdom}}, \code{\link{IndASdomUBtri}}, and
\code{\link{dom.exact}}
}
