% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IndCSdomUBtri}
\alias{IndCSdomUBtri}
\title{Indicator for an upper bound for the domination number of Central Similarity Proximity Catch Digraph
(CS-PCD) by the exact algorithm - one triangle case}
\usage{
IndCSdomUBtri(Dt, k, t, tri, M = c(1, 1, 1))
}
\arguments{
\item{Dt}{A set of 2D points which constitute the vertices of CS-PCD}

\item{k}{A positive integer to be tested for an upper bound for the domination number of CS-PCDs}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region in the
triangle tri}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates which serves as a
center in the interior of the triangle tri; default is M=c(1,1,1) i.e. the center of mass of tri}
}
\value{
A list with two elements
\item{domUB}{The upper bound k (to be checked) for the domination number of CS-PCD. It is prespecified
as k in the function arguments.}
\item{IndUBdom}{The indicator for the upper bound for domination number of CS-PCD being the
specified value k or not. It returns 1 if the upper bound is k, and 0 otherwise.}
\item{ind.domset}{The vertices (i.e., data points) in the dominating set of size k if it exists,
otherwise it is NULL.}
}
\description{
Returns I(domination number of CS-PCD whose vertices are the data points Dt is less than or equal to k),
that is, returns 1 if the domination number of CS-PCD is less than the prespecified value k, returns 0
otherwise. It also provides the vertices (i.e. data points) in a dominating set of size k of CS-PCD.
CS proximity region is constructed with respect to the triangle tri=T(A,B,C) with
expansion parameter t>0 and edge regions are based on the center M=(m1,m2)
in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the interior of tri;
default is M=c(1,1,1) i.e. the center of mass of tri.
Edges of tri, AB, BC, AC, are also labeled as 3, 1, and 2, respectively.
Loops are allowed in the digraph.
It takes a long time for large number of vertices (i.e., large number of row numbers).
See also (\insertCite{ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(0,0); B<-c(1,0); C<-c(.5,sqrt(3.)/2);
Te<-rbind(A,B,C);
n<-10

#set.seed(1)
dat<-runifTe(n)$gen.points
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

M<-as.numeric(runifTe(1)$g)
#M<-c(.6,.2)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)
t<-.5

IndCSdomUBTe(dat,1,t,M)

for (k in 1:n)
  print(c(k,IndCSdomUBTe(dat,k,t,M)))

IndCSdomUBTe(dat,k=4,t,M)

P<-c(.4,.2)
IndCSdomUBTe(P,1,t,M)

IndCSdomUBTe(rbind(dat,dat),3,t,M)

dat.fr<-data.frame(a=dat)
IndCSdomUBTe(dat.fr,1,t,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IndCSdomUBTe}}, \code{\link{IndUBdom}}, \code{\link{IndASdomUBtri}},
and \code{\link{dom.exact}}
}
