% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IndNAStriSet}
\alias{IndNAStriSet}
\title{The indicator for the presence of an arc from a point in set S to the point pt for
Arc Slice Proximity Catch Digraphs (AS-PCDs) - one triangle case}
\usage{
IndNAStriSet(S, pt, tri, M = "CC")
}
\arguments{
\item{S}{A set of 2D points whose AS proximity regions are considered}

\item{pt}{A 2D point. The function determines whether pt is inside the union of AS proximity
regions of points in S or not.}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{M}{"CC" which stand for circumcenter of the triangle tri or a 2D point in Cartesian coordinates or
a 3D point in barycentric coordinates which serves as a center in the interior of tri;
default is M="CC" i.e. the circumcenter of tri}
}
\value{
I(pt in U_{x in S}NAS(x,r)), that is, returns 1 if pt is in S or inside NAS(x) for at least
one x in S, returns 0 otherwise, where AS proximity region is constructed in tri
}
\description{
Returns I(pt in NAS(x) for some x in S), that is, returns 1 if pt is in U_{x in S}NAS(x),
returns 0 otherwise, where NAS(x) is the AS proximity region for point x.
AS proximity regions are constructed with respect to the triangle, tri=T(A,B,C)=(rv=1,rv=2,rv=3),
and vertices of tri are also labeled as 1,2, and 3, respectively.
Vertex regions are based on the center M="CC" for circumcenter of tri;
or M=(m1,m2) in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the
interior of the triangle tri; default is M="CC" i.e. circumcenter of tri.
If pt is not in S and either pt or all points in S are outside tri, it returns 0,
but if pt is in S, then it always returns 1 (i.e., loops are allowed).
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);

#set.seed(1)
n<-20
dat<-runif.tri(n,Tr)$g
#dat<-cbind(runif(n,1,2),runif(n,0,2))
#dat<-cbind(runif(n),runif(n))

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.2)
#M<-circ.cent.tri(Tr); #M<-"CC"; #M<-"CM"
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

NumArcsAStri(dat,Tr,M)
ASarcdens.tri(dat,Tr,M)
ASarcdens.tri(dat,Tr,M,tri.cor = F)

ASarcdens.tri(dat,Tr,M)

dat.fr<-data.frame(a=dat)
ASarcdens.tri(dat.fr,Tr,M)

dat.fr<-data.frame(a=Tr)
ASarcdens.tri(dat,dat.fr,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{Ind.NASTe.set}}, \code{\link{IndNAStri}}, \code{\link{Ind.NAS.Te}},
\code{\link{IndNAStriSet}}, and \code{\link{IndNCStriSet}}
}
