% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IndNCSTe.domset}
\alias{IndNCSTe.domset}
\title{The indicator for the set of points S being a dominating set or not for Central Similarity Proximity
Catch Digraphs (CS-PCDs) - standard equilateral triangle case}
\usage{
IndNCSTe.domset(S, Dt, t, M = c(1, 1, 1))
}
\arguments{
\item{S}{A set of 2D points whose CS proximity regions are considered}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region in the
standard equilateral triangle Te=T((0,0),(1,0),(1/2,sqrt(3)/2))}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the standard equilateral triangle Te; default is M=c(1,1,1) i.e.
the center of mass of Te}

\item{pt}{A 2D point. The function determines whether pt is inside the union of CS proximity
regions of points in S or not.}
}
\value{
I(S a dominating set for CS-PCD), that is, returns 1 if S is a dominating set for CS-PCD,
returns 0 otherwise, where CS proximity region is constructed in the standard equilateral triangle Te
}
\description{
Returns I(S a dominating set of CS-PCD whose vertices are the data points Dt), that is,
returns 1 if S is a dominating set for CS-PCD, returns 0 otherwise. CS proximity region is constructed
with respect to the standard equilateral triangle Te=T(A,B,C)=T((0,0),(1,0),(1/2,sqrt(3)/2)) with
expansion parameter t>0 and edge regions are based on the center M=(m1,m2)
in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the interior of Te;
default is M=c(1,1,1) i.e. the center of mass of Te (which is equivalent to the circumcenter of Te).
Edges of Te, AB, BC, AC, are also labeled as 3, 1, and 2, respectively.
See also (\insertCite{ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
n<-10

#set.seed(1)
dat<-runif.tri(n,Tr)$gen.points
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))
S<-rbind(dat[1,],dat[2,])
#S<-c(1.5,1)

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.0)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)
t<-.5

IndNCStriSet(S,dat[3,],t,Tr,M)
IndNCStriSet(S,dat[3,],t=1,Tr,M)
IndNCStriSet(S,dat[3,],t=1.5,Tr,M)

S<-rbind(dat[1,],dat[2,],dat[3,],dat[5,])
IndNCStriSet(S,dat[3,],t,Tr,M)

IndNCStriSet(S,dat[6,],t,Tr,M)
IndNCStriSet(S,dat[6,],t=.25,Tr,M)

S<-rbind(c(.1,.1),c(.3,.4),c(.5,.3))
IndNCStriSet(S,dat[3,],t,Tr,M)

IndNCStriSet(c(.2,.5),dat[2,],t,Tr,M)
IndNCStriSet(dat,c(.2,.5),t,Tr,M)
IndNCStriSet(dat,dat[2,],t,Tr,M)
IndNCStriSet(c(.2,.5),c(.2,.5),t,Tr,M)
IndNCStriSet(dat[5,],dat[2,],t,Tr,M)

S<-rbind(dat[1,],dat[2,],dat[3,],dat[5,],c(.2,.5))
IndNCStriSet(S,dat[3,],t,Tr,M)

P<-c(.4,.2)
S<-dat[c(1,3,4),]
IndNCStriSet(dat,P,t,Tr,M)

IndNCStriSet(rbind(S,S),P,t,Tr,M)

dat.fr<-data.frame(a=S)
IndNCStriSet(dat.fr,P,t,Tr,M)

}
\seealso{
\code{\link{IndNCStri.domset}} and \code{\link{IndNPETe.domset}}
}
