% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IndNCSTe}
\alias{IndNCSTe}
\title{The indicator for the presence of an arc from a point to another for Central Similarity Proximity Catch
Digraphs (CS-PCDs) - standard equilateral triangle case}
\usage{
IndNCSTe(pt1, pt2, t, M = c(1, 1, 1), re = NULL)
}
\arguments{
\item{pt1}{A 2D point whose CS proximity region is constructed.}

\item{pt2}{A 2D point. The function determines whether pt2 is inside the CS proximity region of
pt1 or not.}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the standard equilateral triangle Te; default is M=c(1,1,1) i.e.
the center of mass of Te}

\item{re}{The index of the edge region in Te containing the point, either 1, 2, 3 or NULL
(default is NULL)}
}
\value{
I(pt2 is in NCS(pt1,t)) for pt1, that is, returns 1 if pt2 is in NCS(pt1,t), returns 0 otherwise
}
\description{
Returns I(pt2 is in NCS(pt1,t)) for points pt1 and pt2, that is, returns 1 if pt2 is in NCS(pt1,t),
returns 0 otherwise, where NCS(x,t) is the CS proximity region for point x with expansion parameter t>01.
CS proximity region is defined with respect to the standard equilateral triangle
Te=T(v=1,v=2,v=3)=T((0,0),(1,0),(1/2,sqrt(3)/2)) and edge regions are based on the center M=(m1,m2)
in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the interior of Te;
default is M=c(1,1,1) i.e. the center of mass of Te.
re is the index of the edge region pt1 resides, with default=NULL.
If pt1 and pt2 are distint and either of them are outside Te, it returns 0,
but if they are identical, then it returns 1 regardless of its location (i.e., it allows loops).
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-CS,ceyhan:test2014;textual}{pcds}).
}
\seealso{
\code{\link{IndNCStri}} and \code{\link{IndNPETe}}
}
