% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IndNCSTeSet}
\alias{IndNCSTeSet}
\title{The indicator for the presence of an arc from a point in set S to the point pt for Central Similarity
Proximity Catch Digraphs (CS-PCDs) - standard equilateral triangle case}
\usage{
IndNCSTeSet(S, pt, t, M = c(1, 1, 1))
}
\arguments{
\item{S}{A set of 2D points. Presence of an arc from a point in S to point pt is checked
by the function.}

\item{pt}{A 2D point. Presence of an arc from a point in S to point pt is checked
by the function.}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region in the
standard equilateral triangle Te=T((0,0),(1,0),(1/2,sqrt(3)/2))}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the standard equilateral triangle Te; default is M=c(1,1,1)
i.e. the center of mass of Te}
}
\value{
I(pt is in U_{x in S} NCS(x,t)), that is, returns 1 if pt is in S or inside NCS(x,t) for at least
one x in S, returns 0 otherwise. CS proximity region is constructed with respect to the standard
equilateral triangle Te=T(A,B,C)=T((0,0),(1,0),(1/2,sqrt(3)/2)) with M-edge regions.
}
\description{
Returns I(pt in NCS(x,t) for some x in S), that is, returns 1 if pt is in U_{x in S}NCS(x,t),
returns 0 otherwise, CS proximity region is constructed with respect to the standard equilateral triangle
Te=T(A,B,C)=T((0,0),(1,0),(1/2,sqrt(3)/2)) with the expansion parameter t>0 and edge regions are based
on center M=(m1,m2) in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the
interior of Te; default is M=c(1,1,1) i.e. the center of mass of Te (which is equivalent to circumcenter of Te).
Edges of Te, AB, BC, AC, are also labeled as edges 3, 1, and 2, respectively.
If pt is not in S and either pt or all points in S are outside Te, it returns 0,
but if pt is in S, then it always returns 1 regardless of its location (i.e., loops are allowed).
See also (\insertCite{ceyhan:mcap2012;textual}{pcds}).
}
\examples{
nx<-100
ny<-10

#set.seed(1)
Xp<-cbind(runif(nx,0,1),runif(nx,0,1))
#Xp<-cbind(runif(nx,0.5,1.5),runif(nx,0,1))
#Xp<-cbind(runif(nx,1,2),runif(nx,0,1))
#Xp<-c(.5,.5); Xp<-matrix(Xp,ncol=2)
Yp<-cbind(runif(ny,0,1),runif(ny,0,1))

M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
t<-1.5
#t<-2

Xlim<-range(Xp[,1],Yp[,1])
Ylim<-range(Xp[,2],Yp[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot.regsCSMT(Xp,Yp,t,M,xlab="",ylab="",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IndNCStriSet}}, \code{\link{IndNCSTe}}, \code{\link{IndNCStri}}, and
\code{\link{IndNPETeSet}}
}
