% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IndNCStri.domset}
\alias{IndNCStri.domset}
\title{The indicator for the set of points S being a dominating set or not for Central Similarity Proximity
Catch Digraphs (CS-PCDs) - one triangle case}
\usage{
IndNCStri.domset(S, Dt, t, tri, M = c(1, 1, 1))
}
\arguments{
\item{S}{A set of 2D points}

\item{Dt}{A set of 2D points}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region
constructed in the triangle tri}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri;
default is M=c(1,1,1) i.e. the center of mass of tri}

\item{S}{A set of 2D points which is to be tested for being a dominating set for the CS-PCDs}

\item{Dt}{A set of 2D points which constitute the vertices of the CS-PCD}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region
constructed in the triangle tri}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}
}
\value{
I(S a dominating set for CS-PCD), that is, returns 1 if S is a dominating set for CS-PCD whose
vertices are the data points in Dt; returns 0 otherwise, where CS proximity region is constructed in
the triangle tri
}
\description{
Returns I(S a dominating set of CS-PCD whose vertices are the data set Dt), that is,
returns 1 if S is a dominating set for CS-PCD, returns 0 otherwise. CS proximity region is constructed with
respect to the triangle tri with the expansion parameter t>0 and edge regions are based
on the center M=(m1,m2) in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates
in the interior of the triangle tri; default is M=c(1,1,1) i.e. the center of mass of tri.
The triangle tri has edges AB, BC, AC which are also labeled as edges 3, 1, and 2, respectively.
See also (\insertCite{ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(0,0); B<-c(1,0); C<-c(.5,sqrt(3.)/2);
Te<-rbind(A,B,C);
n<-10

#set.seed(1)
dat<-runifTe(n)$gen.points
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

M<-as.numeric(runifTe(1)$g)
#M<-c(.6,.2)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)
t<-.5

S<-rbind(dat[1,],dat[2,])
IndNCSTe.domset(S,dat,t,M)

S<-rbind(dat[1,],dat[2,],dat[3,],dat[5,])
IndNCSTe.domset(S,dat,t,M)

S<-rbind(c(.1,.1),c(.3,.4),c(.5,.3))
IndNCSTe.domset(S,dat,t,M)

IndNCSTe.domset(c(.2,.5),dat,t,M)
IndNCSTe.domset(c(.2,.5),c(.2,.5),t,M)
IndNCSTe.domset(dat[5,],dat[2,],t,M)

S<-rbind(dat[1,],dat[2,],dat[3,],dat[5,],c(.2,.5))
IndNCSTe.domset(S,dat[3,],t,M)

IndNCSTe.domset(dat,dat,t,M)

P<-c(.4,.2)
S<-dat[c(1,3,4),]
IndNCSTe.domset(dat,P,t,M)
IndNCSTe.domset(S,P,t,M)
IndNCSTe.domset(S,dat,t,M)

IndNCSTe.domset(rbind(S,S),dat,t,M)

dat.fr<-data.frame(a=dat)
IndNCSTe.domset(S,dat.fr,t,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IndNCSTe.domset}}, \code{\link{IndNPEtri.domset}}
and \code{\link{IndNAStri.domset}}
}
