% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IndNPEbastri}
\alias{IndNPEbastri}
\title{The indicator for the presence of an arc from a point to another for Proportional Edge Proximity Catch
Digraphs (PE-PCDs) - basic triangle case}
\usage{
IndNPEbastri(pt1, pt2, r, c1, c2, M = c(1, 1, 1), rv = NULL)
}
\arguments{
\item{pt1}{A 2D point whose PE proximity region is constructed.}

\item{pt2}{A 2D point. The function determines whether pt2 is inside the PE proximity region of
pt1 or not.}

\item{c1, c2}{Positive real numbers which constitute the vertex of the basic triangle
adjacent to the shorter edges; c1 must be in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the basic triangle or circumcenter of Tb;
default is M=c(1,1,1) i.e. the center of mass of Tb}

\item{rv}{The index of the vertex region in Tb containing the point, either 1, 2, 3 or NULL
(default is NULL)}
}
\value{
I(pt2 is in NPE(pt1,r)) for points pt1 and pt2, that is, returns 1 if pt2 is in NPE(pt1,r),
returns 0 otherwise
}
\description{
Returns I(pt2 is in NPE(pt1,r)) for points pt1 and pt2, that is, returns 1 if pt2 is in NPE(pt1,r),
returns 0 otherwise, where NPE(x,r) is the PE proximity region for point x with expansion parameter r>=1.
PE proximity region is defined with respect to the basic triangle Tb=T((0,0),(1,0),(c1,c2))
where c1 is in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1
and vertex regions are based on the center, M=(m1,m2) in Cartesian coordinates or M=(alpha,beta,gamma) in
barycentric coordinates in the interior of the basic triangle Tb or based on circumcenter of Tb;
default is M=c(1,1,1) i.e. the center of mass of Tb.
rv is the index of the vertex region pt1 resides, with default=NULL.
If pt1 and pt2 are distint and either of them are outside Tb, it returns 0,
but if they are identical, then it returns 1 regardless of its location (i.e., it allows loops).
Any given triangle can be mapped to the basic triangle
by a combination of rigid body motions (i.e., translation, rotation and reflection) and scaling,
preserving uniformity of the points in the original triangle. Hence basic triangle is useful for simulation
studies under the uniformness hypothesis.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:arc-density-PE;textual}{pcds}).
}
\examples{
c1<-.4; c2<-.6
A<-c(0,0); B<-c(1,0); C<-c(c1,c2);
Tb<-rbind(A,B,C);

M<-as.numeric(runif.bastri(1,c1,c2)$g)
#M<-c(.6,.2)
#M<-circ.cent.tri(Tb)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(.3,.3)
r<-2

P1<-as.numeric(runif.bastri(1,c1,c2)$g)
#P1<-c(.4,.2)
NPEbastri(P1,r,c1,c2,M)

P2<-c(1.8,.5)
NPEbastri(P2,r,c1,c2,M)

P3<-c(1.7,.6)
NPEbastri(P3,r,c1,c2,M)

M<-c(1.3,1.3)
r<-2

P1<-c(1.4,1.2)
P2<-c(1.5,1.26)
NPEbastri(P1,r,c1,c2,M)
NPEbastri(P2,r,c1,c2,M)

#or try
Rv<-rv.tri.cent(P1,c1,c2,M)$rv
NPEbastri(P1,r,c1,c2,M,Rv)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IndNPEtri}} and \code{\link{IndNPETe}}
}
