% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IndNPEint}
\alias{IndNPEint}
\title{The indicator for the presence of an arc from a point to another for
Proportional Edge Proximity Catch Digraphs (PE-PCDs) - one interval case}
\usage{
IndNPEint(x1, x2, r, c = 0.5, int)
}
\arguments{
\item{x1}{A 1D point for which the proximity region is constructed}

\item{x2}{A 1D point for which it is checked whether it resides in the proximity region
of x1 or not}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >=1}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a), and default=.5}

\item{int}{A vector of two real numbers representing an interval}
}
\value{
I(x2 in NPE(x1,r,c)) for x2, that is, returns 1 if x2 in NPE(x1,r,c), returns 0 otherwise
}
\description{
Returns I(x2 in NPE(x1,r,c)) for points x1 and x2, that is, returns 1 if x2 is in NPE(x1,r,c),
returns 0 otherwise, where NPE(x,r,c) is the PE proximity region for point x with expansion parameter r>=1
and centrality parameter c in (0,1) and PE proximity region is constructed with respect to the
interval (a,b). This function works whether x1 and x2 are inside or outside the interval int.
Vertex regions for middle intervals are based on the center associated with the centrality parameter c in (0,1).
If x1 and x2 are identical, then it returns 1 regardless of its location
(i.e., loops are allowed in the digraph).
See also (\insertCite{ceyhan:metrika-2012;textual}{pcds}).
}
\examples{
c<-.4
r<-2
a<-0; b<-10; int<-c(a,b)

NPEint(7,r,c,int)
NPEint(17,r,c,int)
NPEint(1,r,c,int)
NPEint(-1,r,c,int)

NPEint(3,r,c,int)
NPEint(4,r,c,int)
NPEint(a,r,c,int)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IndNPEmid1D}}, \code{\link{IndNPEend1D}} and \code{\link{IndNCSint}}
}
