% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IndNPETe}
\alias{IndNPETe}
\title{The indicator for the presence of an arc from a point to another for Proportional Edge Proximity Catch
Digraphs (PE-PCDs) - standard equilateral triangle case}
\usage{
IndNPETe(pt1, pt2, r, M = c(1, 1, 1), rv = NULL)
}
\arguments{
\item{pt1}{A 2D point whose PE proximity region is constructed.}

\item{pt2}{A 2D point. The function determines whether pt2 is inside the PE proximity region of
pt1 or not.}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the standard equilateral triangle Te; default is M=c(1,1,1) i.e.
the center of mass of Te}

\item{rv}{The index of the vertex region in Te containing the point, either 1, 2, 3 or NULL
(default is NULL)}
}
\value{
I(pt2 is in NPE(pt1,r)) for points pt1 and pt2, that is, returns 1 if pt2 is in NPE(pt1,r),
returns 0 otherwise
}
\description{
Returns I(pt2 is in NPE(pt1,r)) for points pt1 and pt2, that is, returns 1 if pt2 is in NPE(pt1,r),
returns 0 otherwise, where NPE(x,r) is the PE proximity region for point x with expansion parameter r>=1.
PE proximity region is defined with respect to the standard equilateral triangle
Te=T(v=1,v=2,v=3)=T((0,0),(1,0),(1/2,sqrt(3)/2)) and vertex regions are based on the center M=(m1,m2)
in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the interior of Te;
default is M=c(1,1,1) i.e. the center of mass of Te.
rv is the index of the vertex region pt1 resides, with default=NULL.
If pt1 and pt2 are distint and either of them are outside Te, it returns 0,
but if they are identical, then it returns 1 regardless of its location (i.e., it allows loops).
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:arc-density-CS;textual}{pcds}).
}
\examples{
A<-c(0,0); B<-c(1,0); C<-c(.5,sqrt(3.)/2);
Te<-rbind(A,B,C)
n<-20

#set.seed(1)
dat<-runifTe(n)$gen.points
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

M<-as.numeric(runifTe(1)$g)
#M<-c(.6,.2)
#M<-circ.cent.tri(Te)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

rvTe.cent(dat[1,],M)
rvTe.cent(c(.7,.2),M)
rvTe.cent(c(0,1),M)

Rv<-vector()
for (i in 1:n)
  Rv<-c(Rv,rvTe.cent(dat[i,],M)$rv)
Rv

Ds<-cp2e.tri(Te,M)

Xlim<-range(Te[,1],dat[,1])
Ylim<-range(Te[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

if (dimension(M)==3) {M<-bary2cart(M,Te)} #need to run this when M is given in barycentric coordinates

plot(Te,asp=1,pch=".",xlab="",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Te)
points(dat,pch=".",col=1)
L<-rbind(M,M,M); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)

txt<-rbind(Te,M)
xc<-txt[,1]+c(-.02,.03,.02,0)
yc<-txt[,2]+c(.02,.02,.03,.05)
txt.str<-c("A","B","C","M")
text(xc,yc,txt.str)

text(dat,labels=factor(Rv))

rvTe.cent(c(.7,.2),M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IndNPEtri}}, \code{\link{IndNPEbastri}}, and \code{\link{IndNCSTe}}
}
