% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IndNPETeSet}
\alias{IndNPETeSet}
\title{The indicator for the presence of an arc from a point in set S to the point pt for Proportional Edge
Proximity Catch Digraphs (PE-PCDs) - standard equilateral triangle case}
\usage{
IndNPETeSet(S, pt, r, M = c(1, 1, 1))
}
\arguments{
\item{S}{A set of 2D points. Presence of an arc from a point in S to point pt is checked
by the function.}

\item{pt}{A 2D point. Presence of an arc from a point in S to point pt is checked
by the function.}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region in the
standard equilateral triangle Te=T((0,0),(1,0),(1/2,sqrt(3)/2)); must be >=1}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the standard equilateral triangle Te; default is M=c(1,1,1)
i.e. the center of mass of Te}
}
\value{
I(pt is in U_{x in S} NPE(x,r)), that is, returns 1 if pt is in S or inside NPE(x,r) for at least
one x in S, returns 0 otherwise. PE proximity region is constructed with respect to the standard
equilateral triangle Te=T(A,B,C)=T((0,0),(1,0),(1/2,sqrt(3)/2)) with M-vertex regions
}
\description{
Returns I(pt in NPE(x,r) for some x in S), that is, returns 1 if pt is in U_{x in S}NPE(x,r),
returns 0 otherwise, PE proximity region is constructed with respect to the standard equilateral triangle
Te=T(A,B,C)=T((0,0),(1,0),(1/2,sqrt(3)/2)) with the expansion parameter r>=1 and vertex regions are based
on center M=(m1,m2) in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the
interior of Te; default is M=c(1,1,1) i.e. the center of mass of Te (which is equivalent to the circumcenter
for Te). Vertices of Te are also labeled as 1, 2, and 3, respectively.
If pt is not in S and either pt or all points in S are outside Te, it returns 0,
but if pt is in S, then it always returns 1 regardless of its location (i.e., loops are allowed).
}
\examples{
A<-c(0,0); B<-c(1,0); C<-c(1/2,sqrt(3)/2)
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C)
n<-20
dat<-runif.tri(n,Tr)$g

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-(A+B+C)/3
#M<-circ.cent.tri(Tr)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)
r<-1.4

PEdomtri(dat,Tr,r,M)
IM<-IncMatPEtri(dat,Tr,r,M)
dom.greedy(IM)
dom.exact(IM)

gr.gam<-dom.greedy(IM)
gr.gam
dat[gr.gam$i,]

PEdomtri(rbind(dat,c(5,5)),Tr,r,M)

PEdomtri(dat,Tr,r,M=c(.4,.4))

PEdomtri(rbind(dat,c(5,5)),Tr,r,M=c(.4,.4))

P1<-c(.5,.5)
PEdomtri(P1,Tr,r,Cent)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IndNPEtriSet}}, \code{\link{IndNPETe}}, \code{\link{IndNPEtri}}, and
\code{\link{IndNCSTeSet}}
}
