% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IndNPEtri}
\alias{IndNPEtri}
\title{The indicator for the presence of an arc from a point to another for Proportional Edge Proximity
Catch Digraphs (PE-PCDs) - one triangle case}
\usage{
IndNPEtri(pt1, pt2, r, tri, M = c(1, 1, 1), rv = NULL)
}
\arguments{
\item{pt1}{A 2D point whose PE proximity region is constructed.}

\item{pt2}{A 2D point. The function determines whether pt2 is inside the PE proximity region of
pt1 or not.}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >=1}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri or the circumcenter of tri;
default is M=c(1,1,1) i.e. the center of mass of tri}

\item{rv}{Index of the M-vertex region containing the point, either 1, 2, 3 or NULL (default is NULL)}
}
\value{
I(pt2 is in NPE(pt1,r)) for points pt1 and pt2, that is, returns 1 if pt2 is in NPE(pt1,r),
returns 0 otherwise
}
\description{
Returns I(pt2 is in NPE(pt1,r)) for points pt1 and pt2, that is, returns 1 if pt2 is in NPE(pt1,r),
returns 0 otherwise, where NPE(x,r) is the PE proximity region for point x with the expansion parameter r>=1.
PE proximity region is constructed with respect to the triangle tri and
vertex regions are based on the center, M=(m1,m2) in Cartesian coordinates or
M=(alpha,beta,gamma) in barycentric coordinates in the interior of tri
or based on the circumcenter of tri; default is M=c(1,1,1) i.e. the center of mass of tri.
rv is the index of the vertex region pt1 resides, with default=NULL.
If pt1 and pt2 are distint and either of them are outside tri, it returns 0,
but if they are identical, then it returns 1 regardless of its location (i.e., it allows loops).
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-PE,ceyhan:dom-num-NPE-Spat2011;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.0)
#M<-circ.cent.tri(Tb)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)
r<-1.5

n<-10
#set.seed(1)
dat<-runif.tri(n,Tr)$g

NPEtri(dat[7,],r,Tr,M)

P1<-as.numeric(runif.tri(1,Tr)$g)
#P1<-c(.4,.2)
NPEtri(P1,r,Tr,M)

P2<-c(1.8,.5)
NPEtri(P2,r,Tr,M)

P3<-c(1.7,.6)
NPEtri(P3,r,Tr,M)

M<-c(1.3,1.3)
r<-2

P1<-c(1.4,1.2)
P2<-c(1.5,1.26)
NPEtri(P1,r,Tr,M)
NPEtri(P2,r,Tr,M)

#or try
Rv<-rv.tri.cent(P1,Tr,M)$rv
NPEtri(P1,r,Tr,M,Rv)

dat.fr<-data.frame(a=Tr)
NPEtri(P2,r,dat.fr,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IndNPEbastri}}, \code{\link{IndNPETe}}, \code{\link{IndNAStri}}, and
\code{\link{IndNCStri}}
}
