% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{int.2lines}
\alias{int.2lines}
\title{The point of intersection of two lines}
\usage{
int.2lines(p1, q1, p2, q2)
}
\arguments{
\item{p1, q1}{2D points that determine the first straight line (i.e., through which the first straight line is passing)}

\item{p2, q2}{2D points that determine the second straight line (i.e., through which the second straight line is passing)}
}
\value{
The coordinates of the point of intersection of the two lines, first passing through points \code{p1} and \code{q1}
and second passing through points \code{p2} and \code{q2}.
}
\description{
Returns the intersection of two lines, first line passing through points p1 and q1
and second line passing through points p2 and q2. The points are chosen so that lines are well defined.
}
\examples{
A<-c(-1.22,-2.33); B<-c(2.55,3.75)
#A<-runif(2); B<-runif(2)
#A<-runif(2,1,100); B<-runif(2,1,100)
#A<-c(1,1); B<-c(1,2) #lines are vertical
#A<-c(1,1); B<-c(1,-2) #lines are vertical
xr<-range(A,B);
xf<-(xr[2]-xr[1])*.1 #how far to go at the lower and upper ends in the x-coordinate
x<-seq(xr[1]-xf,xr[2]+xf,l=100)

lnAB<-Line(A,B,x)
lnAB
summary(lnAB)
plot(lnAB)

line(A,B) #this takes A as the x points and B as the y points and fits the line

y<-lnAB$y
Xlim<-range(x,A,B)
if (!is.na(y[1])) {Ylim<-range(y,A,B)} else {Ylim<-range(A,B)}
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]
pf<-c(xd,-yd)*.025

#plot of the line joining A and B
plot(rbind(A,B),pch=1,xlab="x",ylab="y",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
if (!is.na(y[1])) {lines(x,y,lty=1)} else {abline(v=A[1])}
text(rbind(A+pf,B+pf),c("A","B"))
int<-round(lnAB$intercept,2)#intercept
sl<-round(lnAB$slope,2)  #slope
text(rbind((A+B)/2+pf*3),ifelse(is.na(int),paste("x=",A[1]),ifelse(sl==0,paste("y=",int),ifelse(sl==1,ifelse(sign(int)<0,paste("y=x",int),paste("y=x+",int)),
                                                                                                ifelse(sign(int)<0,paste("y=",sl,"x",int),paste("y=",sl,"x+",int))))))

Line(c(1,2),c(2,3),3)
}
\seealso{
\code{\link{int.circ.line}} and \code{\link{dp2l}}
}
