% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{is.in.data}
\alias{is.in.data}
\title{Check a point belong to a given data set.}
\usage{
is.in.data(p, Dt)
}
\arguments{
\item{p}{A 2D point for which the function checks membership to the data set Dt}

\item{Dt}{A set of 2D points representing the set of data points}
}
\value{
TRUE if \code{p} belongs to the data set \code{Dt}.
}
\description{
returns TRUE if the point p of any dimension is inside the data set Dt of the same dimension as p;
otherwise returns FALSE.
}
\examples{
asyvarPE(1.2)

rseq<-seq(1.01,5,by=.05)
lrseq<-length(rseq)

avar<-vector()
for (i in 1:lrseq)
{
  avar<-c(avar,asyvarPE(rseq[i]))
}

par(mar=c(5,5,4,2))
plot(rseq, avar,type="l",xlab="r",ylab=expression(paste(sigma^2,"(r)")),lty=1,xlim=range(rseq))

}
