% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{is.point}
\alias{is.point}
\title{Check the argument is a point is of a given dimension.}
\usage{
is.point(p, dim = 2)
}
\arguments{
\item{p}{A vector to be checked to see it is a point of dimension dim or not.}

\item{dim}{A positive integer representing the dimension of the argument p.}
}
\value{
TRUE if \code{p} is a vector of dimension \code{dim}.
}
\description{
Returns TRUE if the arguement \code{p} is a numeric point of dimension \code{dim} (default is \code{dim}=2);
otherwise returns FALSE
}
\seealso{
\code{\link{dimension}}
}
