% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{is.StdEqTri}
\alias{is.StdEqTri}
\title{Check whether a triangle is a standard equilateral triangle}
\usage{
is.StdEqTri(tri)
}
\arguments{
\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}
}
\value{
TRUE if \code{tri} is a standard equilateral triangle, else FALSE.
}
\description{
Checks whether the triangle, tri, is the standard equilateral triangle Te=T((0,0),(1,0),(1/2,sqrt(3)/2))
or not.
}
