% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{Kfr2vTbVRCC}
\alias{Kfr2vTbVRCC}
\title{An object of class "Extrema"

The k furthest points from vertices in each CC-vertex region in a basic triangle}
\usage{
Kfr2vTbVRCC(Dt, c1, c2, k, ch.all.intri = FALSE)
}
\arguments{
\item{Dt}{A set of 2D points}

\item{c1, c2}{Positive real numbers which constitute the vertex of the basic triangle
adjacent to the shorter edges; c1 must be in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1}

\item{k}{a positive integer}

\item{ch.all.intri}{A logical argument}
}
\value{
A list with the elements
\item{txt1}{Vertex labels are A=1, B=2, and C=3 (corresponds to row number in Extrema Points).}
\item{txt2}{A shorter description of the distances as "Distances of k furthest points in the vertex regions to Vertices".}
\item{type}{Type of the extrema points}
\item{desc}{A short description of the extrema points}
\item{mtxt}{The "main" title for the plot of the exrema}
\item{Ext}{The extrema points, here, k furthest points from vertices in each vertex region.}
\item{X}{The input data, Dt, can be a matrix or data frame}
\item{num.points}{The number of data points, i.e., size of Dt}
\item{supp}{Support of the data points, here, it is Tb.}
\item{cent}{The center point used for construction of edge regions.}
\item{ncent}{Name of the center, cent, it is circumcenter "CC" for this function.}
\item{regions}{Vertex regions inside the triangle, Tb, provided as a list.}
\item{region.names}{Names of the vertex regions as "vr=1","vr=2","vr=3"}
\item{region.centers}{Centers of mass of the vertex regions inside Tb.}
\item{dist2ref}{Distances from k furthest points in each vertex region to the corresponding vertex
(each row representing a vertex).}
}
\description{
Returns the k furthest data points among the data set, Dt, in each CC-vertex region from the vertex in the
basic triangle Tb=T(A=(0,0),B=(1,0),C=(c1,c2)).
Any given triangle can be mapped to the basic triangle by a combination of rigid body motions
(i.e., translation, rotation and reflection) and scaling, preserving uniformity of the points in the
original triangle. Hence basic triangle is useful for simulation
studies under the uniformness hypothesis.
ch.all.intri is for checking whether all data points are inside Tb (default is FALSE).
In the extrema, Ext, in the output, the first k entries are the k furthest points from vertex 1,
second k entries are k furthest points are from vertex 2, and
last k entries are the k furthest points from vertex 3
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
n<-20

#set.seed(1)
dat<-runif.tri(n,Tr)$g
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,2))

Ext<-fr2vVRCC(dat,Tr)
Ext
summary(Ext)
plot(Ext)

fr2vVRCC(dat[1,],Tr)

dat2<-rbind(dat,c(.2,.4))
fr2vVRCC(dat2,Tr)
fr2vVRCC(dat2,Tr,ch.all.intri = TRUE)

f2v<-fr2vVRCC(dat,Tr)

CC<-circ.cent.tri(Tr) #the circumcenter
D1<-(B+C)/2; D2<-(A+C)/2; D3<-(A+B)/2;
Ds<-rbind(D1,D2,D3)

Xlim<-range(Tr[,1],dat[,1])
Ylim<-range(Tr[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Tr,xlab="",asp=1,ylab="",pch=".",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Tr)
L<-matrix(rep(CC,3),ncol=3,byrow=T); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)
points(dat)
points(rbind(f2v$Ext),pch=4,col=2)

txt<-rbind(Tr,CC,Ds)
xc<-txt[,1]+c(-.06,.08,.05,.12,-.1,-.1,-.09)
yc<-txt[,2]+c(.02,-.02,.05,.0,.02,.06,-.04)
txt.str<-c("A","B","C","CC","D1","D2","D3")
text(xc,yc,txt.str)

fr2vVRCC(dat,Tr)

fr2vVRCC(c(1.4,1.2),Tr)

dat.fr<-data.frame(a=dat)
fr2vVRCC(dat.fr,Tr)

dat.fr<-data.frame(a=Tr)
fr2vVRCC(dat,dat.fr)


}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{fr2vTbVRCC}}, \code{\link{fr2vVRCC}}, \code{\link{fr2eTeER}},
and \code{\link{Kfr2vVRCC}}
}
