% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{Line}
\alias{Line}
\title{An object of class "Lines"

The line joining two distinct 2D points a and b}
\usage{
Line(a, b, x)
}
\arguments{
\item{a, b}{2D points that determine the straight line (i.e., through which the straight line passes)}

\item{x}{A scalar or a vector of scalars representing the x-coordinates of the line}
}
\value{
A list with the elements
\item{txt}{A description of the line}
\item{mtxt}{The "main" title for the plot of the line}
\item{points}{The input points \code{a} and \code{b} through which the straight line passes
(stacked row-wise, i.e., row 1 is point \code{a} and row 2 is point \code{b}).}
\item{x}{The input scalar or vector which constitutes the x-coordinates of the point(s) of interest
on the line.}
\item{y}{The output scalar or vector which constitutes the y-coordinates of the point(s) of interest
on the line. If \code{x} is a scalar, then \code{y} will be a scalar and if \code{x} is a vector of scalars,
then \code{y} will be a vector of scalars.}
\item{slope}{Slope of the line, Inf is allowed, passing through points \code{a} and \code{b} }
\item{intercept}{Intercept of the line passing through points \code{a} and \code{b}}
\item{equation}{Equation of the line passing through points \code{a} and \code{b}}
}
\description{
Returns the equation, slope, intercept, and y coordinates of the line crossing two distinct 2D points a and b
with x coordinates provided in vector x. This function is different from the line function in the
standard stats package in R in the sense that Line(a,b,x) fits the line passing through points a and b and
returns various quantities (see below) for this line and x is the x-coordinates of the points we want to find
on the Line(a,b,x) while line(a,b) fits the line robustly whose x-coordinates are in a and y-coordinates are
in b. Line(a,b,x) and line(x,Line(A,B,x)$y) would yield the same straight line (i.e., the line with the same
coefficients.)
}
\examples{
A<-c(-1.22,-2.33); B<-c(2.55,3.75)
slope(A,B)

slope(c(1,2),c(2,3))
slope(c(1,2),c(1,3))

}
\seealso{
\code{\link{slope}}, \code{\link{paraline}}, \code{\link{perpline}}, \code{\link{line}}
in the generic stats package and and \code{\link{Line3D}}
}
