% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{Line3D}
\alias{Line3D}
\title{An object of class "Lines3D"

The line crossing 3D point r0 in the direction of vector v (or if v is a point, in direction of v-r0)}
\usage{
Line3D(r0, v, t, dir.vec = TRUE)
}
\arguments{
\item{r0}{A 3D point through which the straight line passes}

\item{v}{A 3D vector which determines the direction of the straight line (i.e., the straight line would be
parallel to this vector) if the dir.vec=T, otherwise it is 3D point and v-r0 determines the direction of the
the straight line}

\item{t}{A scalar or a vector of scalars representing the parameter of the coordinates of the line
(for the form: x=x0 + a*t, y=y0 + b*t, and z=z0 + c*t where r0=(x0,y0,z0) and v=(a,b,c) if dir.vec=T, else
v-r0=(a,b,c))}

\item{dir.vec}{A logical argument about v, if TRUE v is treated as a vector, else v is treated as a point
and so the direction vector is taken to be v-r0}
}
\value{
A list with the elements
\item{txt}{A description of the line}
\item{mtxt}{The "main" title for the plot of the line}
\item{pts}{The input points that determine a line and/or a plane, NULL for this function.}
\item{pnames}{The names of the input points that determine a line and/or a plane, NULL for this function.}
\item{vecs}{The point r0 and the vector v (if dir.vec=T) or the point v (if dir.vec=F). The first row is r0
and the second row is v.}
\item{vec.names}{The names of the point r0 and the vector v (if dir.vec=T) or the point v (if dir.vec=F).}
\item{x,y,z}{The x-, y- and z-coordinates of the point(s) of interest on the line.}
\item{tsq}{The scalar or the vector of the parameter in defining each coordinate of the line for the form:
x=x0 + a*t, y=y0 + b*t, and z=z0 + c*t where r0=(x0,y0,z0) and v=(a,b,c) if dir.vec=T, else v-r0=(a,b,c).}
\item{equation}{Equation of the line passing through point r0 in the direction of the vector v (if dir.vec=T)
else in the direction of v-r0. The line equation is in the form: x=x0 + a*t, y=y0 + b*t, and z=z0 + c*t where
r0=(x0,y0,z0) and v=(a,b,c) if dir.vec=T, else v-r0=(a,b,c).}
}
\description{
Returns the equation, x, y, and z coordinates of the line crossing 3D point r0
in the direction of vector v (of if v is a point, in the direction of v-r0)
with the parameter t being provided in vector t.
}
\examples{
nx<-1000
ny<-10
r<-1.5
#r<-2
#r<-1.25

#set.seed(1)
Xp<-cbind(runif(nx,0,1),runif(nx,0,1))
#Xp<-cbind(runif(nx,.5,1.5),runif(nx))
#Xp<-cbind(runif(nx,1,2),runif(nx))
Yp<-cbind(runif(ny,0,1),runif(ny,0,1))

plotDeltri(Xp,Yp,xlab="",ylab="")

PEdomMTnd(Xp,Yp,r)

TSDomPENor(Xp,Yp,r)
TSDomPENor(Xp,Yp,r)

TSDomPENor(Xp,Yp,1.25,ch=TRUE)

#or try
ndt<-NumDelTri(Yp)
TSDomPENor(Xp,Yp,r,nt=ndt)
#values might differ due to the random of choice of the three centers M1,M2,M3 for the non-degenerate asymptotic
#distribution of the domination number

TSDomPENor(Xp,Yp,r)

TSDomPENor(Xp,rbind(Yp,Yp),r)

dat.fr<-data.frame(a=Xp)
TSDomPENor(dat.fr,Yp,r)

dat.fr<-data.frame(a=Yp)
TSDomPENor(Xp,dat.fr,r)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{line}}, \code{\link{paraline3D}}, and \code{\link{Plane}}
}
