% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{NASbastri}
\alias{NASbastri}
\title{The vertices of the Arc Slice (AS) Proximity Region in the basic triangle}
\usage{
NASbastri(pt, c1, c2, M = "CC", rv = NULL, dec = 4)
}
\arguments{
\item{pt}{A 2D point whose AS proximity region is to be computed}

\item{c1, c2}{Positive real numbers representing the top vertex in basic triangle Tb=T((0,0),(1,0),(c1,c2)),
c1 must be in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1}

\item{M}{"CC" which stand for circumcenter of the triangle Tb or a 2D point in Cartesian coordinates or
a 3D point in barycentric coordinates which serves as a center in the interior of the triangle Tb;
default is M="CC" i.e. the circumcenter of Tb}

\item{rv}{The index of the M-vertex region containing the point, either 1, 2, 3 or NULL (default is NULL)}

\item{dec}{a positive integer the number of decimals (default is 4) to round the barycentric coordinates when checking whether
the end points fall on the boundary of the triangle Tb or not}
}
\value{
A list with the elements
\item{L,R}{The end points of the line segments on the boundary of the AS proximity region.
 Each row in L and R constitute a line segment on the boundary. }
\item{Arc.Slices}{The end points of the arc-slices on the circular parts of the AS proximity region.
Here points in row 1 and row 2 constitute the end points of one arc-slice, points on row 3 and row 4
constitute the end points for the next arc-slice and so on}
}
\description{
Returns the end points of the line segments and arc-slices that constitute the boundary of
AS proximity region for a point in the basic triangle Tb=T((0,0),(1,0),(c1,c2))
where c1 is in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1
and vertex regions are based on the center M="CC" for circumcenter of Tb; or M=(m1,m2) in Cartesian
coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the interior of Tb;
default is M="CC" the circumcenter of Tb.
rv is the index of the vertex region pt resides, with default=NULL.
If pt is outside Tb, it returns NULL for the proximity region.
dec is the number of decimals (default is 4) to round the barycentric coordinates when checking whether
the end points fall on the boundary of the triangle Tb or not (so as not to miss the intersection points
due to precision in the decimals).
Any given triangle can be mapped to the basic triangle
by a combination of rigid body motions (i.e., translation, rotation and reflection) and scaling,
preserving uniformity of the points in the original triangle. Hence basic triangle is useful for simulation
studies under the uniformness hypothesis.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
P1<-c(.3,.2)*100
P2<-c(.6,.3)*100
cent<-c(1.1,1.1)*100
rad<-2*100

int.circ.line(P1,P2,cent,rad)
int.circ.line(P2,P1,cent,rad)
int.circ.line(P1,P1+c(0,1),cent,rad)
int.circ.line(P1+c(0,1),P1,cent,rad)

dp2l(cent,P1,P2)
rad2<-dp2l(cent,P1,P2)$d
int.circ.line(P1,P2,cent,rad2)
int.circ.line(P1,P2,cent,rad=.8)
int.circ.line(P1,P2,cent,rad=.78)

#plot of the line and the circle
A<-c(.3,.2); B<-c(.6,.3); cent<-c(1,1); rad<-2 #dp2l(cent,A,B)$dis, .3
#A<-runif(2); B<-runif(2); cent<-runif(2); rad<-runif(1)
#A<-runif(2,1,100); B<-runif(2,1,100); cent<-runif(2,1,100); rad<-runif(2,1,100);

IPs<-int.circ.line(A,B,cent,rad)

xr<-range(A[1],B[1],cent[1])
xf<-(xr[2]-xr[1])*.1 #how far to go at the lower and upper ends in the x-coordinate
x<-seq(xr[1]-rad-xf,xr[2]+rad+xf,l=100)
lnAB<-Line(A,B,x)
y<-lnAB$y

Xlim<-range(x,cent[1])
Ylim<-range(y,A[2],B[2],cent[2]-rad,cent[2]+rad)
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(rbind(A,B,cent),pch=1,asp=1,xlab="x",ylab="y",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
lines(x,y,lty=1)
circles(cent[1],cent[2],rad)
IP.txt<-c()
if (!is.null(IPs))
{
  for (i in 1:(length(IPs)/2))
    IP.txt<-c(IP.txt,paste("I",i, sep = ""))
}
txt<-rbind(A,B,cent,IPs)
text(txt+cbind(rep(xd*.03,nrow(txt)),rep(-yd*.03,nrow(txt))),c("A","B","M",IP.txt))

}
\seealso{
\code{\link{NAStri}} and \code{\link{IndNASbastri}}
}
