% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{NCStri}
\alias{NCStri}
\title{The vertices of the Central Similarity (CS) Proximity Region in a general triangle}
\usage{
NCStri(pt, t, tri, M, re = NULL)
}
\arguments{
\item{pt}{A 2D point whose CS proximity region is to be computed}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri}

\item{re}{Index of the M-edge region containing the point pt, either 1, 2, 3 or NULL (default is NULL)}
}
\value{
Vertices of the triangular region which constitutes the CS proximity region with expansion parameter
t>0 and center M for a point pt
}
\description{
Returns the vertices of the CS proximity region (which is itself a triangle) for a point in the
triangle tri=T(A,B,C)=(rv=1,rv=2,rv=3). CS proximity region is defined with respect to the triangle tri
with expansion parameter t>0 and edge regions based on center M=(m1,m2) in Cartesian coordinates or
M=(alpha,beta,gamma) in barycentric coordinates in the interior of the triangle tri.
Edge regions are labeled as 1,2,3 rowwise for the corresponding vertices
of the triangle tri. re is the index of the edge region pt resides, with default=NULL.
If pt is outside of tri, it returns NULL for the proximity region.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:arc-density-CS;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);

M<-c(1.6,1.2)

P<-c(.4,.2)
redges.tri.cent(P,Tr,M)

P<-c(1.8,.5)
redges.tri.cent(P,Tr,M)

P<-c(10.5,1.6)
redges.tri.cent(P,Tr,M)

n<-20
#set.seed(1)
dat<-runif.tri(n,Tr)$g
#dat<-cbind(runif(n,1,2),runif(n,0,2))
#dat<-cbind(runif(n),runif(n))

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.2)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

re<-redges.tri.cent(dat,Tr,M)
re

D1<-(B+C)/2; D2<-(A+C)/2; D3<-(A+B)/2;
Ds<-rbind(D1,D2,D3)

Xlim<-range(Tr[,1],dat[,1])
Ylim<-range(Tr[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

if (dimension(M)==3) {M<-bary2cart(M,Tr)} #need to run this when M is given in barycentric coordinates

plot(Tr,pch=".",xlab="",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Tr)
points(dat,pch=".",col=1)
L<-Tr; R<-rbind(M,M,M)
segments(L[,1], L[,2], R[,1], R[,2], lty=2)

xc<-Tr[,1]#+c(-.02,.03,.02)
yc<-Tr[,2]#+c(.02,.02,.04)
txt.str<-c("A","B","C")
text(xc,yc,txt.str)

txt<-rbind(M,Ds)
xc<-txt[,1]#+c(.05,.06,-.05,-.02)
yc<-txt[,2]#+c(.03,.03,.05,-.08)
txt.str<-c("M","re=2","re=3","re=1")
text(xc,yc,txt.str)
text(dat,labels=factor(re$re))

redges.tri.cent(dat,Tr,M)

dat.fr<-data.frame(a=dat)
redges.tri.cent(dat.fr,Tr,M)

dat.fr<-data.frame(a=Tr)
redges.tri.cent(dat,dat.fr,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{NPEtri}}, \code{\link{NAStri}}, and  \code{\link{IndNCStri}}
}
