% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{NPEbastri}
\alias{NPEbastri}
\title{The vertices of the Proportional Edge (PE) Proximity Region in a basic triangle}
\usage{
NPEbastri(pt, r, c1, c2, M = c(1, 1, 1), rv = NULL)
}
\arguments{
\item{pt}{A 2D point whose PE proximity region is to be computed}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >=1}

\item{c1, c2}{Positive real numbers representing the top vertex in basic triangle Tb=T((0,0),(1,0),(c1,c2)),
c1 must be in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the basic triangle Tb or the circumcenter of Tb;
default is M=c(1,1,1) i.e. the center of mass of Tb}

\item{rv}{Index of the M-vertex region containing the point pt, either 1, 2, 3 or NULL (default is NULL)}
}
\value{
Vertices of the triangular region which constitutes the PE proximity region with expansion parameter
r and center M for a point pt
}
\description{
Returns the vertices of the PE proximity region (which is itself a triangle) for a point in the
basic triangle Tb=T((0,0),(1,0),(c1,c2))=(rv=1,rv=2,rv=3). PE proximity region is defined with respect
to the basic triangle Tb with expansion parameter r>=1 and vertex regions based on center M=(m1,m2) in
Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the interior of the basic
triangle Tb or based on the circumcenter of Tb; default is M=c(1,1,1) i.e. the center of mass of Tb.
Vertex regions are labeled as 1,2,3 rowwise for the vertices
of the triangle Tb. rv is the index of the vertex region pt resides, with default=NULL.
If pt is outside of tri, it returns NULL for the proximity region.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
c1<-.4; c2<-.6
A<-c(0,0); B<-c(1,0); C<-c(c1,c2);
Tb<-rbind(A,B,C);
M<-c(.6,.2)

P<-c(.4,.2)
rv.bastri.cent(P,c1,c2,M)

P<-c(1.8,.5)
rv.bastri.cent(P,c1,c2,M)

P<-c(.5,.26)
rv.bastri.cent(P,c1,c2,M)

n<-20
#set.seed(1)
dat<-runif.bastri(n,c1,c2)$g
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

M<-as.numeric(runif.bastri(1,c1,c2)$g)
#M<-c(.6,.2)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

Rv<-vector()
for (i in 1:n)
{ Rv<-c(Rv,rv.bastri.cent(dat[i,],c1,c2,M)$rv)}
Rv

Ds<-cp2e.bastri(c1,c2,M)

Xlim<-range(Tb[,1],dat[,1])
Ylim<-range(Tb[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

if (dimension(M)==3) {M<-bary2cart(M,Tb)} #need to run this when M is given in barycentric coordinates

plot(Tb,pch=".",xlab="",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.1,.1),ylim=Ylim+yd*c(-.05,.05))
polygon(Tb)
points(dat,pch=".",col=1)
L<-rbind(M,M,M); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)

xc<-Tb[,1]+c(-.04,.05,.04)
yc<-Tb[,2]+c(.02,.02,.03)
txt.str<-c("rv=1","rv=2","rv=3")
text(xc,yc,txt.str)

txt<-rbind(M,Ds)
xc<-txt[,1]+c(-.02,.04,-.03,0)
yc<-txt[,2]+c(-.02,.02,.02,-.03)
txt.str<-c("M","D1","D2","D3")
text(xc,yc,txt.str)

text(dat,labels=factor(Rv))

P<-c(.4,.2)
rv.bastri.cent(P,c1,c2,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{NPEtri}}, \code{\link{NAStri}}, \code{\link{NCStri}}
and \code{\link{IndNPEbastri}}
}
