% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{NPEtetra}
\alias{NPEtetra}
\title{The vertices of the Proportional Edge (PE) Proximity Region in a tetrahedron}
\usage{
NPEtetra(pt, r, th, M = "CM", rv = NULL)
}
\arguments{
\item{pt}{A 3D point whose PE proximity region is to be computed}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{th}{Four 3D points, stacked row-wise, each row representing a vertex of the tetrahedron}

\item{M}{The center to be used in the construction of the vertex regions in the tetrahedron, th.
Currently it only takes "CC" for circumcenter and "CM" for center of mass; default="CM"}

\item{rv}{Index of the vertex region containing the point, either 1, 2, 3, 4 (default is NULL)}
}
\value{
Vertices of the tetrahedron which constitutes the PE proximity region with expansion parameter
r and circumcenter (or center of mass) for a point pt in the tetrahedron
}
\description{
Returns the vertices of the PE proximity region (which is itself a tetrahedron) for a point in the
tetrahedron th. PE proximity region is defined with respect to the tetrahedron th
with expansion parameter r>=1 and vertex regions based on the center M which is circumcenter ("CC") or
center of mass ("CM") of th with default="CM".
Vertex regions are labeled as 1,2,3,4 rowwise for the vertices of the tetrahedron th.
rv is the index of the vertex region pt resides, with default=NULL.
If pt is outside of th, it returns NULL for the proximity region.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010;textual}{pcds}).
}
\examples{
A<-c(0,0,0); B<-c(1,0,0); C<-c(1/2,sqrt(3)/2,0); D<-c(1/2,sqrt(3)/6,sqrt(6)/3)
tetra<-rbind(A,B,C,D)
r<-1.5

n<-20
dat<-runif.stdtetra(n)$g
#dat[,1]<-dat[,1]+1

plot.regsPEstdtetra(dat[1,],r)

plot.regsPEstdtetra(dat[5,],r)

plot.regsPEstdtetra(c(.4,.4,.4),r)

plot.regsPEstdtetra(c(.5,.5,.5),r)

plot.regsPEstdtetra(dat[1:3,],r)

P1<-c(.1,.1,.1)
plot.regsPEstdtetra(c("a","b","c"),r)
plot.regsPEstdtetra(rbind(P1,P1),r)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{NPEstdtetra}}, \code{\link{NPEtri}} and \code{\link{NPEint}}
}
