% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{NPEtri}
\alias{NPEtri}
\title{The vertices of the Proportional Edge (PE) Proximity Region in a general triangle}
\usage{
NPEtri(pt, r, tri, M = c(1, 1, 1), rv = NULL)
}
\arguments{
\item{pt}{A 2D point whose PE proximity region is to be computed}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >=1}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri or the circumcenter of tri;
default is M=c(1,1,1) i.e. the center of mass of tri}

\item{rv}{Index of the M-vertex region containing the point pt, either 1, 2, 3 or NULL (default is NULL)}
}
\value{
Vertices of the triangular region which constitutes the PE proximity region with expansion parameter
r and center M for a point pt
}
\description{
Returns the vertices of the PE proximity region (which is itself a triangle) for a point in the
triangle tri=T(A,B,C)=(rv=1,rv=2,rv=3). PE proximity region is defined with respect to the triangle tri
with expansion parameter r>=1 and vertex regions based on center M=(m1,m2) in Cartesian coordinates or
M=(alpha,beta,gamma) in barycentric coordinates in the interior of the triangle tri
or based on the circumcenter of tri; default is M=c(1,1,1) i.e. the center of mass of tri.
Vertex regions are labeled as 1,2,3 rowwise for the vertices
of the triangle tri. rv is the index of the vertex region pt resides, with default=NULL.
If pt is outside of tri, it returns NULL for the proximity region.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-PE,ceyhan:dom-num-NPE-Spat2011;textual}{pcds}).
}
\seealso{
\code{\link{NPEbastri}}, \code{\link{NAStri}}, \code{\link{NCStri}}
and \code{\link{IndNPEtri}}
}
