% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{NumArcsAStri}
\alias{NumArcsAStri}
\title{Number of arcs of Arc Slice Proximity Catch Digraphs (AS-PCDs) - one triangle case}
\usage{
NumArcsAStri(dat, tri, M = "CC")
}
\arguments{
\item{dat}{A set of 2D points which constitute the vertices of the digraph (i.e., AS-PCD)}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{M}{"CC" which stand for circumcenter of the triangle tri or a 2D point in Cartesian coordinates or
a 3D point in barycentric coordinates which serves as a center in the interior of tri;
default is M="CC" i.e. the circumcenter of tri}
}
\value{
The number of arcs of Arc Slice Proximity Catch Digraphs (AS-PCDs)
}
\description{
Returns the number of arcs of Arc Slice Proximity Catch Digraphs (AS-PCDs)
whose vertices are the 2D data set, dat. The data points could be inside or outside a general
triangle tri=T(A,B,C)=(rv=1,rv=2,rv=3), with vertices of tri stacked row-wise.
AS proximity regions are defined with respect to the triangle tri and vertex regions are
based on the center M="CC" for circumcenter of tri;
or M=(m1,m2) in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the
interior of the triangle tri; default is M="CC" i.e. circumcenter of tri.
For the number of arcs, loops are not allowed,
so arcs are only possible for points inside the triangle, tri.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.2)
#M<-circ.cent.tri(Tr); #M<-"CC"; #M<-"CM"
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

P1<-as.numeric(runif.tri(1,Tr)$g)
P2<-as.numeric(runif.tri(1,Tr)$g)
IndNAStri(P1,P2,Tr,M)

P1<-c(1.3,1.2)
P2<-c(1.8,.5)
IndNAStri(P1,P2,Tr,M)
IndNAStri(P1,P1,Tr,M)

#or try
Rv<-rv.triCC(P1,Tr)$rv
IndNAStri(P1,P2,Tr,M,Rv)

P3<-c(1.6,1.4)
IndNAStri(P1,P3,Tr,M)

P4<-c(1.5,1.0)
IndNAStri(P1,P4,Tr,M)

P5<-c(.5,1.0)
IndNAStri(P1,P5,Tr,M)
IndNAStri(P5,P5,Tr,M)

#or try
Rv<-rv.triCC(P5,Tr)$rv
IndNAStri(P5,P5,Tr,M,Rv)

dat.fr<-data.frame(a=Tr)
IndNAStri(P1,P2,dat.fr,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{NumArcsASMT}}, \code{\link{NumArcsPEtri}}, and \code{\link{NumArcsCStri}}
}
