% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{NumArcsCSend1D}
\alias{NumArcsCSend1D}
\title{Number of arcs of Central Similarity Proximity Catch Digraphs (CS-PCDs) - end interval case}
\usage{
NumArcsCSend1D(dat, t, int)
}
\arguments{
\item{dat}{A vector of 1D points which constitute the vertices of the digraph}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{int}{A vector of two real numbers representing an interval}
}
\value{
Number of arcs for the CS-PCD with vertices being 1D data set, dat,
expansion parameter, t, for the end intervals.
}
\description{
Returns the number of arcs of Central Similarity Proximity Catch Digraphs (CS-PCDs) whose
vertices are a 1D numerical data set, dat, outside the interval int=(a,b).
CS proximity region is constructed only with expansion parameter t>0 for points outside the interval (a,b).
End vertex regions are based on the end points of the interval,
i.e., the corresponding end vertex region is an interval as (-Inf,a) or (b,Inf) for the interval (a,b).
For the number of arcs, loops are not allowed, so arcs are only possible for points outside
the interval, int, for this function.
See also (\insertCite{ceyhan:revstat-2016;textual}{pcds}).
}
\examples{
a<-0; b<-10; int<-c(a,b)
t<-2

IndNCSend1D(15,17,t,int)
IndNCSend1D(15,15,t,int)

IndNCSend1D(1.5,17,t,int)
IndNCSend1D(1.5,1.5,t,int)

IndNCSend1D(-15,17,t,int)

IndNCSend1D(-15,-17,t,int)

a<-0; b<-10; int<-c(a,b)
t<-.5

IndNCSend1D(15,17,t,int)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{NumArcsCSmid1D}}, \code{\link{NumArcsPEmid1D}}, and \code{\link{NumArcsPEend1D}}
}
