% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{NumArcsCSint}
\alias{NumArcsCSint}
\title{Number of arcs of  Central Similarity Proximity Catch Digraphs (CS-PCDs) - one interval case}
\usage{
NumArcsCSint(dat, t, c = 0.5, int)
}
\arguments{
\item{dat}{A set of 1D points which constitute the vertices of CS-PCD}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a); default=.5}

\item{int}{A vector of two real numbers representing an interval}
}
\value{
Number of arcs for the CS-PCD whose vertices are the 1D data set, dat,
with expansion parameter, t>0, and centrality parameter, c in (0,1).
}
\description{
Returns the number of arcs of Central Similarity Proximity Catch Digraphs (CS-PCDs)
whose vertices are the 1D data set dat in the one-interval case.
The data points could be inside or outside the interval is int=(a,b). CS proximity region is constructed
with an expansion parameter t>0 and a centrality parameter c in (0,1).
The CS proximity region is constructed for both points inside and outside the interval, hence
the arcs may exist for all points inside or outside the interval.
See also (\insertCite{ceyhan:revstat-2016;textual}{pcds}).
}
\examples{
c<-.4
t<-2
a<-0; b<-10; int<-c(a,b)
#c<-runif(1)
#a<-runif(1); b<-runif(1); int<-range(a,b); a<-int[1]; b<-int[2]

n<-10
xr<-range(a,b)
xf<-(xr[2]-xr[1])*.1
dat<-runif(n,a-xf,b+xf)
#dat<-runif(n,a-5,b+5)

plot.regsCSint(7,t,c,int)

plot.regsCSint(dat,t,c,int)

plot.regsCSint(17,t,c,int)
plot.regsCSint(1,t,c,int)
plot.regsCSint(4,t,c,int)

plot.regsCSint(-7,t,c,int)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{NumArcsCSmid1D}}, \code{\link{NumArcsCSend1D}}, and \code{\link{NumArcsPEint}}
}
