% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{NumArcsCSTe}
\alias{NumArcsCSTe}
\title{Number of arcs of Central Similarity Proximity Catch Digraphs (CS-PCDs) - standard
equilateral triangle case}
\usage{
NumArcsCSTe(dat, t, M = c(1, 1, 1))
}
\arguments{
\item{dat}{A set of 2D points which constitute the vertices of the digraph}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the standard equilateral triangle Te; default is M=c(1,1,1) i.e.
the center of mass of Te}
}
\value{
Number of arcs for the CS-PCD with vertices being 2D data set, dat, in Te
with expansion parameter, t>0, and center of mass CM. CS proximity regions are defined only
for dat points inside Te, i.e. arcs are possible for such points only.
}
\description{
Returns the number of arcs of Central Similarity Proximity Catch Digraphs (CS-PCDs) whose vertices are the
given 2D numerical data set, dat. CS proximity region NCS(x,r) is defined with respect to the standard
equilateral triangle Te=T(v=1,v=2,v=3)=T((0,0),(1,0),(1/2,sqrt(3)/2)) with expansion parameter t>0
and edge regions are based on the center M=(m1,m2)
in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the interior of Te;
default is M=c(1,1,1) i.e. the center of mass of Te.
For the number of arcs, loops are not allowed so
arcs are only possible for points inside Te for this function.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-CS,ceyhan:test2014;textual}{pcds}).
}
\examples{
A<-c(0,0); B<-c(1,0); C<-c(.5,sqrt(3.)/2);
Te<-rbind(A,B,C);
n<-10

#set.seed(1)
dat<-runifTe(n)$gen.points
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

M<-as.numeric(runifTe(1)$g)
#M<-c(.6,.2)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

NumArcsCSTe(dat,t=1.25)

inc.mat<-IncMatCSTe(dat,t=1.25,M)
inc.mat
sum(inc.mat)-n

dom.greedy(inc.mat)
dom.exact(inc.mat) #might take a long time for large n
IndUBdom(inc.mat,1)

inc.mat<-IncMatCSTe(rbind(dat,c(0,1)),t=1.25,M)
inc.mat
sum(inc.mat)-(n+1)

IncMatCSTe(dat,t=1.5,M);

IncMatCSTe(rbind(dat,dat),t=1.5,M)

dat.fr<-data.frame(a=dat)
IncMatCSTe(dat.fr,t=1.5,M);

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{NumArcsCStri}}, \code{\link{NumArcsCSMT}}, and \code{\link{NumArcsPETe}},
}
