% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{NumArcsCStri}
\alias{NumArcsCStri}
\title{Number of arcs of Central Similarity Proximity Catch Digraphs (CS-PCDs) - one triangle case}
\usage{
NumArcsCStri(Dt, tri, t, M = c(1, 1, 1))
}
\arguments{
\item{Dt}{A set of 2D points which constitute the vertices of CS-PCD}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri; default is M=c(1,1,1) i.e.
the center of mass of tri}
}
\value{
Number of arcs for the CS-PCD with vertices being 2D data set, dat, in tri
with expansion parameter, t>0, and center of mass CM. CS proximity regions are defined only
for dat points inside tri, i.e. arcs are possible for such points only.
}
\description{
Returns the number of arcs of Central Similarity Proximity Catch Digraphs (CS-PCDs) whose vertices are the
given 2D numerical data set, dat. CS proximity region NCS(x,r) is defined with respect to the triangle, tri  with expansion parameter t>0
and edge regions are based on the center M=(m1,m2)
in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the interior of tri;
default is M=c(1,1,1) i.e. the center of mass of tri.
For the number of arcs, loops are not allowed so
arcs are only possible for points inside tri for this function.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-CS,ceyhan:test2014;textual}{pcds}).
}
\examples{
A<-c(0,0); B<-c(1,0); C<-c(.5,sqrt(3.)/2);
n<-20

#set.seed(1)
dat<-runifTe(n)$gen.points
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

M<-as.numeric(runifTe(1)$g)
#M<-c(.6,.2)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

NumArcsCSTe(dat,t=.5,M)
NumArcsCSTe(dat,t=3,M)
NumArcsCSTe(dat,t=1.5,M)

NumArcsCSTe(rbind(dat,c(0,1)),t=2,M)
NumArcsCSTe(c(.4,.2),t=.5,M)

NumArcsCSTe(dat,t=1.5,M);

NumArcsCSTe(rbind(dat,dat),t=1.5,M)

dat.fr<-data.frame(a=dat)
NumArcsCSTe(dat.fr,t=1.5,M);

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{NumArcsCSTe}}, \code{\link{NumArcsCSMT}}, \code{\link{NumArcsPEtri}},
and \code{\link{NumArcsAStri}}
}
