% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{NumArcsPEend1D}
\alias{NumArcsPEend1D}
\title{Number of arcs of Proportional Edge Proximity Catch Digraphs (PE-PCDs) - end interval case}
\usage{
NumArcsPEend1D(dat, r, int)
}
\arguments{
\item{dat}{A vector of 1D points which constitute the vertices of the digraph}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >=1}

\item{int}{A vector of two real numbers representing an interval}
}
\value{
Number of arcs for the PE-PCD with vertices being 1D data set, dat,
expansion parameter, r>=1, for the end intervals.
}
\description{
Returns the number of arcs of Proportional Edge Proximity Catch Digraphs (PE-PCDs) whose
vertices are a 1D numerical data set, dat, outside the interval int=(a,b).
PE proximity region is constructed only with expansion parameter r>=1 for points outside the interval (a,b).
End vertex regions are based on the end points of the interval,
i.e., the corresponding vertex region is an interval as (-Inf,a) or (b,Inf) for the interval (a,b).
For the number of arcs, loops are not allowed, so arcs are only possible for points outside
the interval, int, for this function.
See also (\insertCite{ceyhan:metrika-2012;textual}{pcds}).
}
\examples{
a<-0; b<-10; int<-c(a,b)
r<-2

IndNPEend1D(15,17,r,int)
IndNPEend1D(15,15,r,int)

IndNPEend1D(1.5,17,r,int)

IndNPEend1D(-15,17,r,int)

IndNPEend1D(a,17,r,int)
IndNPEend1D(15,17,r=1.1,int)

a<-0; b<-10; int<-c(a,b)
r<-2
IndNPEend1D(15,17,r,int)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{NumArcsPEmid1D}}, \code{\link{NumArcsCSmid1D}}, and \code{\link{NumArcsCSend1D}}
}
