% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{NumArcsPEint}
\alias{NumArcsPEint}
\title{Number of arcs of Proportional Edge Proximity Catch Digraphs (PE-PCDs) - one interval case}
\usage{
NumArcsPEint(dat, r, c = 0.5, int)
}
\arguments{
\item{dat}{A set of 1D points which constitute the vertices of PE-PCD}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >=1}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a); default=.5}

\item{int}{A vector of two real numbers representing an interval}
}
\value{
Number of arcs for the PE-PCD whose vertices are the 1D data set, dat,
with expansion parameter, r>=1, and centrality parameter, c in (0,1).
}
\description{
Returns the number of arcs of Proportional Edge Proximity Catch Digraphs (PE-PCDs)
whose vertices are the 1D data set dat in the one-interval case.
The data points could be inside or outside the interval is int=(a,b). PE proximity region is constructed
with an expansion parameter r>=1 and a centrality parameter c in (0,1).
The PE proximity region is constructed for both points inside and outside the interval, hence
the arcs may exist for all points inside or outside the interval.
See also (\insertCite{ceyhan:metrika-2012;textual}{pcds}).
}
\examples{
c<-.4
r<-2
a<-0; b<-10; int<-c(a,b)
#c<-runif(1)
#a<-runif(1); b<-runif(1); int<-range(a,b); a<-int[1]; b<-int[2]

n<-10
xr<-range(a,b)
xf<-(xr[2]-xr[1])*.1
dat<-runif(n,a-xf,b+xf)
#dat<-runif(n,a-5,b+5)

plot.regsPEint(7,r,c,int)

plot.regsPEint(dat,r,c,int)

plot.regsPEint(17,r,c,int)
plot.regsPEint(1,r,c,int)
plot.regsPEint(4,r,c,int)

plot.regsPEint(-7,r,c,int)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{NumArcsPEmid1D}}, \code{\link{NumArcsPEend1D}}, and \code{\link{NumArcsCSint}}
}
