% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{NumArcsPEMT}
\alias{NumArcsPEMT}
\title{Number of arcs of Proportional Edge Proximity Catch Digraphs (PE-PCDs) - multiple triangle case}
\usage{
NumArcsPEMT(Xp, Yp, r, M = c(1, 1, 1))
}
\arguments{
\item{Xp}{A set of 2D points which constitute the vertices of the PE-PCD}

\item{Yp}{A set of 2D points which constitute the vertices of the Delaunay triangles}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >=1}

\item{M}{A 3D point in barycentric coordinates which serves as a center in the interior of each Delaunay
triangle or circumcenter of each Delaunay triangle (for this argument should be set as M="CC"),
default for M is c(1,1,1) which is the center of mass of each triangle}
}
\value{
A list of elements
\item{num.arcs}{Total number of arcs in all triangles}
\item{num.in.conhull}{Number of Xp points in the convex hull of Yp points}
\item{weight.vec}{The vector of the areas of Delaunay triangles based on Yp points}
}
\description{
Returns the number of arcs of Proportional Edge Proximity Catch Digraph (PE-PCD) whose vertices are
the data points in Xp in the multiple triangle case. PE proximity regions are defined with respect to the
Delaunay triangles based on Yp points with expansion parameter r>=1 and vertex regions in each triangle
is based on the center M=(alpha,beta,gamma) in barycentric coordinates in the interior of each
Delaunay triangle or based on circumcenter of each Delaunay triangle (default for M is c(1,1,1)
which is the center of mass of the triangle).
Each Delaunay triangle is first converted to an (unscaled) basic triangle so that M will be the same
type of center for each Delaunay triangle (this conversion is not necessary when M is CM).
Convex hull of Yp is partitioned by the Delaunay triangles based on Yp points
(i.e., multiple triangles are the set of these Delaunay triangles whose union constitutes the
convex hull of Yp points). For the number of arcs, loops are not allowed so arcs are only possible
for points inside the convex hull of Yp points.
See (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-PE;textual}{pcds}) for more on PE-PCDs.
Also see (\insertCite{okabe:2000,ceyhan:comp-geo-2010,renka:1996;textual}{pcds}) for more on Delaunay triangulation and the tripack algorithm.
}
\examples{
nx<-100
ny<-10

#set.seed(1)
Xp<-cbind(runif(nx),runif(nx))
Yp<-cbind(runif(ny),runif(ny))
plotDeltri(Xp,Yp,xlab="",ylab="",main="X points and Delaunay Triangulation of Y points")

P<-c(.6,.4)
plotDeltri(P,Yp,xlab="",ylab="",main="X points and Delaunay Triangulation of Y points")

plotDeltri(Xp,Yp,xlab="",ylab="")

plotDeltri(Xp,rbind(Yp,Yp),xlab="",ylab="")

dat.fr<-data.frame(a=Xp)
plotDeltri(dat.fr,Yp,xlab="",ylab="")

dat.fr<-data.frame(a=Yp)
plotDeltri(Xp,dat.fr,xlab="",ylab="")

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{NumArcsPEtri}}, \code{\link{NumArcsPETe}}, \code{\link{NumArcsCSMT}},
and \code{\link{NumArcsASMT}}
}
