% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{NumDelTri}
\alias{NumDelTri}
\title{Number of Delaunay triangles based on a 2D data set}
\usage{
NumDelTri(Yp)
}
\arguments{
\item{Yp}{A set of 2D points which constitute the vertices of Delaunay triangles}
}
\value{
Number of Delaunay triangles based on Yp points.
}
\description{
Returns the number of Delaunay triangles based on the 2D set of points Yp.
See (\insertCite{okabe:2000,renka:1996;textual}{pcds}) for more on Delaunay triangulation and the tripack algorithm.
}
\examples{
Pg2PEtri(r=1.5)
Pg2PEtri(r=1.4999999999)

Pg2PEtri(r=1.5) / Pg2PEtri(r=1.4999999999)

rseq<-seq(1.01,1.49999999999,l=100)
lrseq<-length(rseq)

pg2<-vector()
for (i in 1:lrseq)
{
  pg2<-c(pg2,Pg2PEtri(rseq[i]))
}

plot(rseq, pg2,type="l",xlab="r",ylab=expression(paste("P(", gamma, "=2)")),
     lty=1,xlim=range(rseq)+c(0,.01),ylim=c(0,1))
points(rbind(c(1.50,Pg2PEtri(1.50))),pch=".",cex=3)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plotDeltri}}
}
