% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{paraline}
\alias{paraline}
\title{An object of class "Lines"

The line paralel to the line segment joining two distinct 2D points a and b}
\usage{
paraline(p, a, b, x)
}
\arguments{
\item{p}{A 2D point at which the parallel line to line segment joining a and b crosses.}

\item{a, b}{2D points that determine the line segment (the line will be parallel to this line segment)}

\item{x}{A scalar or a vector of scalars representing the x-coordinates of the line parallel to ab and crossing p}
}
\value{
A list with the elements
\item{txt}{Description of the line passing through point p and parallel to line segment joining
\code{a} and \code{b}}
\item{mtxt}{The "main" title for the plot of the line passing through point p and parallel to line segment joining
\code{a} and \code{b}}
\item{points}{The input points p, a, and b (stacked row-wise,
i.e., point \code{p} is in row 1, point \code{a} is in row 2 and point \code{b} is in row 3).
Line parallel to ab crosses p.}
\item{x}{The input vector. It can be a scalar or a vector of scalars,
which constitute the x-coordinates of the point(s) of interest on the line passing through point p and parallel to line segment joining
\code{a} and \code{b}.}
\item{y}{The output scalar or vector which constitutes the y-coordinates of the point(s) of interest
on the line passing through point p and parallel to line segment joining \code{a} and \code{b}.
If \code{x} is a scalar, then \code{y} will be a scalar and if \code{x} is a vector of scalars,
then \code{y} will be a vector of scalars.}
\item{slope}{Slope of the line, Inf is allowed, passing through point p and parallel to line segment joining
\code{a} and \code{b}}
\item{intercept}{Intercept of the line passing through point p and parallel to line segment joining
\code{a} and \code{b}}
\item{equation}{Equation of the line passing through point p and parallel to line segment joining
\code{a} and \code{b}}
}
\description{
Returns the equation, slope, intercept, and y coordinates of the line crossing the point p and
parallel to the line passing through the points a and b with x coordinates are provided in vector x.
}
\examples{
A<-c(0,0); B<-c(1,0); C<-c(.5,sqrt(3.)/2);
Te<-rbind(A,B,C);
CM<-(A+B+C)/3;
D1<-(B+C)/2; D2<-(A+C)/2; D3<-(A+B)/2;
Ds<-rbind(D1,D2,D3)
nx<-100

#Data generation step
#set.seed(1)
Xdt<-runif.1_6thTe(nx)
Xdt
summary(Xdt)
plot(Xdt,asp=1)

Xd<-runif.1_6thTe(nx)$gen.points

#plot of the data with the regions in the equilateral triangle
Xlim<-range(Te[,1])
Ylim<-range(Te[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Te,asp=1,pch=".",xlim=Xlim+xd*c(-.01,.01),ylim=Ylim+yd*c(-.01,.01),xlab=" ",ylab=" ",
     main="first 1/6th of the \\n standard equilateral triangle")
polygon(Te)
L<-Te; R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)
polygon(rbind(A,D3,CM),col=5)
points(Xd)

#letter annotation of the plot
txt<-rbind(A,B,C,CM,D1,D2,D3)
xc<-txt[,1]+c(-.02,.02,.02,.04,.05,-.03,0)
yc<-txt[,2]+c(.02,.02,.02,.03,0,.03,-.03)
txt.str<-c("A","B","C","CM","D1","D2","D3")
text(xc,yc,txt.str)

}
\seealso{
\code{\link{slope}}, \code{\link{Line}}, and \code{\link{perpline}}, \code{\link{line}}
in the generic stats package, \code{\link{paraline3D}}
}
