% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{PEdom.tetra}
\alias{PEdom.tetra}
\title{The domination number of Proportional Edge Proximity Catch Digraph (PE-PCD) - one tetrahedron case}
\usage{
PEdom.tetra(Xp, th, r, M = "CM")
}
\arguments{
\item{Xp}{A set of 3D points which constitute the vertices of the digraph}

\item{th}{Four 3D points, stacked row-wise, each row representing a vertex of the tetrahedron}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{M}{The center to be used in the construction of the vertex regions in the tetrahedron, th.
Currently it only takes "CC" for circumcenter and "CM" for center of mass; default="CM"}
}
\value{
A list with two elements
\item{dom.num}{Domination number of PE-PCD with vertex set=Xp and expansion parameter r>=1 and center M}
\item{mds}{A minimum dominating set for PE-PCD with vertex set=Xp and expansion parameter r>=1 and center M}
}
\description{
Returns the domination number of PE-PCD whose vertices are the data points in Xp.
PE proximity region is defined with respect to the tetrahedron th with expansion parameter r>=1 and
vertex regions are based on the center M which is circumcenter ("CC") or center of mass ("CM") of th
with default="CM".
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010;textual}{pcds}).
}
\seealso{
\code{\link{PEdomtri}}
}
