% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{PEdomtri}
\alias{PEdomtri}
\title{The domination number of Proportional Edge Proximity Catch Digraph (PE-PCD) - one triangle case}
\usage{
PEdomtri(Xp, tri, r, M = c(1, 1, 1))
}
\arguments{
\item{Xp}{A set of 2D points which constitute the vertices of the digraph}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri or the circumcenter of tri,
default is (1,1,1) i.e. the center of mass}
}
\value{
A list with two elements
\item{dom.num}{Domination number of PE-PCD with vertex set=Xp and expansion parameter r>=1 and center M}
\item{mds}{A minimum dominating set for PE-PCD with vertex set=Xp and expansion parameter r>=1 and center M}
}
\description{
Returns the domination number of PE-PCD whose vertices are the data points in Xp.
PE proximity region is defined with respect to the triangle tri with expansion parameter r>=1 and
vertex regions are constructed with center M=(m1,m2) in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates
in the interior of the triangle tri or the circumcenter of tri.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:masa-2007,ceyhan:dom-num-NPE-Spat2011,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
n<-20

#set.seed(1)
dat<-runif.tri(n,Tr)$g
#dat<-cbind(runif(n,1,2),runif(n,0,2))
#dat<-cbind(runif(n),runif(n))

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.0)
#M<-circ.cent.tri(Tr)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)
r<-1.5
#r<-2

Gam2PEtri(dat[1,],dat[2,],dat,Tr,r,M)
Gam2PEtri(dat[1,],dat[4,],dat,Tr,r,M)
Gam2PEtri(dat[4,],dat[4,],dat,Tr,r,M)

Gam2PEtri(dat[1,],c(1,2),dat,Tr,r,M)
Gam2PEtri(dat[1,],c(1,2),dat,Tr,r,M,ch.data.pnts = TRUE)

Gam2PEtri(c(1,2),c(1,3),rbind(c(1,2),c(1,3)),Tr,r,M)

ind.gam2<-vector()
for (i in 1:(n-1))
  for (j in (i+1):n)
  {if (Gam2PEtri(dat[i,],dat[j,],dat,Tr,r,M)==1)
    ind.gam2<-rbind(ind.gam2,c(i,j))}
ind.gam2

#or try
rv1<-rv.tri.cent(dat[1,],Tr,M)$rv;
rv2<-rv.tri.cent(dat[2,],Tr,M)$rv
Gam2PEtri(dat[1,],dat[2,],dat,Tr,r,M,rv1,rv2)

#or try
rv1<-rv.tri.cent(dat[1,],Tr,M)$rv;
Gam2PEtri(dat[1,],dat[2,],dat,Tr,r,M,rv1)

#or try
rv2<-rv.tri.cent(dat[2,],Tr,M)$rv
Gam2PEtri(dat[1,],dat[2,],dat,Tr,r,M,rv2=rv2)

P1<-c(1.4,1)
P2<-c(1.6,1)
Gam2PEtri(P1,P2,dat,Tr,r,M)
Gam2PEtri(dat[1,],dat[2,],dat,Tr,r,M)

dat.fr<-data.frame(a=dat)
Gam2PEtri(P1,P2,dat.fr,Tr,r,M)

dat.fr<-data.frame(a=Tr)
Gam2PEtri(P1,P2,dat,dat.fr,r,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{PEdomMTnd}}, \code{\link{PEdomMT}} and \code{\link{PEdom1D}}
}
