% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{perpline}
\alias{perpline}
\title{An object of class "Lines"

The line passing thru a point and perpendicular to the line segment joining two points}
\usage{
perpline(p, a, b, x)
}
\arguments{
\item{p}{A 2D point at which the perpendicular line to line segment joining a and b crosses.}

\item{a, b}{2D points that determine the line segment (the line will be perpendicular to this line segment)}

\item{x}{A scalar or a vector of scalars representing the x-coordinates of the line perpendicular to
line joing a and b and crossing p}
}
\value{
A list with the elements
\item{txt}{Description of the line passing through point p and perpendicular to line joining
\code{a} and \code{b}}
\item{mtxt}{The "main" title for the plot of the line passing through point p and perpendicular to line joining
\code{a} and \code{b}}
\item{points}{The input points a and b (stacked row-wise,
i.e., row 1 is point \code{a} and row 2 is point \code{b}). Line passing through point p is perpendicular
to line joining \code{a} and \code{b}}
\item{x}{The input vector, can be a scalar or a vector of scalars,
which constitute the x-coordinates of the point(s) of interest on the line passing through point p and perpendicular to line joining
\code{a} and \code{b}}
\item{y}{The output vector which constitutes the y-coordinates of the point(s) of interest on the line passing through point p and perpendicular to line joining
\code{a} and \code{b}. If \code{x} is a scalar, then \code{y} will be a scalar and if \code{x} is a vector of scalars,
then \code{y} will be a vector of scalars.}
\item{slope}{Slope of the line passing through point p and perpendicular to line joining
\code{a} and \code{b}}
\item{intercept}{Intercept of the line passing through point p and perpendicular to line joining
\code{a} and \code{b}}
\item{equation}{Equation of the line passing through point p and perpendicular to line joining
\code{a} and \code{b}}
}
\description{
Returns the equation, slope, intercept, and y coordinates of the line crossing the point p and
perpendicular to the line passing through the points a and b with x coordinates are provided in vector x.
}
\examples{
A<-c(0,0); B<-c(1,0); C<-c(.5,sqrt(3.)/2);
Te<-rbind(A,B,C);
n<-1000
eps<-.2 #try also .25, .1

#set.seed(1)
Xdt<-rascIITe(n,eps)
Xdt
summary(Xdt)
plot(Xdt,asp=1)

Xlim<-range(Te[,1])
Ylim<-range(Te[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

dat<-rascIITe(n,eps)$gen.points
plot(Te,pch=".",xlab="",ylab="",main="Type II association in the \\n standard equilateral triangle",
     xlim=Xlim+xd*c(-.01,.01),ylim=Ylim+yd*c(-.01,.01))
polygon(Te)
points(dat)

# The support for the Type II association alternative
A1<-c(1/2-eps*sqrt(3),sqrt(3)/6-eps); B1<-c(1/2+eps*sqrt(3),sqrt(3)/6-eps); C1<-c(1/2,sqrt(3)/6+2*eps);
supp<-rbind(A1,B1,C1)

plot(Te,asp=1,pch=".",xlab="",ylab="",main="Support of the Type II Association",xlim=Xlim+xd*c(-.01,.01),ylim=Ylim+yd*c(-.01,.01))
polygon(Te,col=5)
polygon(supp,col=0)
#points(dat)

}
\seealso{
\code{\link{slope}}, \code{\link{Line}}, and \code{\link{paraline}}
}
