% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{Pg2PE1D}
\alias{Pg2PE1D}
\title{The probability of domination number=2 for Proportional Edge Proximity Catch Digraphs (PE-PCDs)
- middle interval case}
\usage{
Pg2PE1D(r, c, n)
}
\arguments{
\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a)}

\item{n}{A positive integer representing the size of the uniform data set}
}
\value{
P(domination number=2) for PE-PCD whose vertices are a uniform data set of size n in a finite
interval (a,b)
}
\description{
Returns P(domination number=2) for PE-PCD whose vertices are a uniform data set of size n in a finite interval
(a,b). The PE proximity region NPE(x,r,c) is defined with respect to (a,b) with centrality parameter c in (0,1)
and expansion parameter r>=1.
}
\seealso{
\code{\link{Pg2PEtri}} and  \code{\link{Pg2PE1D.asy}}
}
