% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plot.intervals}
\alias{plot.intervals}
\title{The plot of the subintervals based on Yp points together with Xp points}
\usage{
\method{plot}{intervals}(Xp, Yp, main = "", xlab = "", ylab = "",
  xlim = NULL, ylim = NULL)
}
\arguments{
\item{Xp}{A set of 1D points whose scatter-plot is provided}

\item{Yp}{A set of 1D points which constitute the end points of the intervals which
partition the real line}

\item{xlab, ylab}{Titles for the x and y axes, respectively (default="" for both)}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges (default=NULL for both)}
}
\value{
Plot of the intervals based on Yp points and also scatter plot of Xp points
}
\description{
Plots the Xp points and the intervals based on Yp points
points.
}
\examples{
r<-2
c<-.4
a<-0; b<-10;
nx<-10; ny<-4

#set.seed(1)
Xp<-runif(nx,a,b)
#Xp<-runif(nx,a-5,b+5)
#Xp<-runif(nx,a+5,b+5)
#Xp<-runif(nx,a+10,b+10)
Yp<-runif(ny,a,b)

IM<-IncMatPE1D(Xp,Yp,r,c)
IM

dom.greedy(IM)
IndUBdom(IM,6)
dom.exact(IM)

Arcs<-ArcsPEMI(Xp,Yp,r,c)
Arcs
summary(Arcs)
plot(Arcs)

IncMatPE1D(Xp,Yp+10,r,c)

r<-2
c<-.4
a<-0; b<-10;
nx<-40; ny<-10
Xp<-runif(nx,a,b)
Yp<-runif(ny,a,b)
IncMatPE1D(Xp,Yp,r,c)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plot.regsPEMI}} and \code{\link{plotDeltri}}
}
