% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxRFuncs4PCDClasses.r
\name{plot.PCDs}
\alias{plot.PCDs}
\title{Plot a \code{PCDs} object}
\usage{
\method{plot}{PCDs}(x, Jit = 0.1, ...)
}
\arguments{
\item{x}{Object of class \code{PCDs}.}

\item{Jit}{A positive real number that determines the amount of jitter along the y-axis, default is 0.1, for the
1D case, the vertices of the PCD are jittered according to \eqn{U(-Jit,Jit)} distribution along the y-axis where
\code{Jit} equals to the range of vertices and the interval end points; it is redundant in the 2D case.}

\item{\dots}{Additional parameters for \code{plot}.}
}
\value{
None
}
\description{
Plots the vertices and the arcs of the PCD together with the vertices and boundaries of the partition
cells (i.e., intervals in the 1D case and triangles in the 2D case)
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
Tr<-rbind(A,B,C);
n<-10
dat<-runif.tri(n,Tr)$g
M<-as.numeric(runif.tri(1,Tr)$g)
Arcs<-ArcsAStri(dat,Tr,M)
Arcs
plot(Arcs)

}
\seealso{
\code{\link{print.PCDs}}, \code{\link{summary.PCDs}}, and \code{\link{print.summary.PCDs}}
}
