% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plot.regsAStri}
\alias{plot.regsAStri}
\title{The plot of the Arc Slice (AS) Proximity Regions for a 2D data set - one triangle case}
\usage{
\method{plot}{regsAStri}(Xp, tri, M = "CC", main = "", xlab = "",
  ylab = "", xlim = NULL, ylim = NULL)
}
\arguments{
\item{Xp}{A set of 2D points for which AS proximity regions are constructed}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{M}{"CC" which stand for circumcenter of the triangle tri or a 2D point in Cartesian coordinates or a 3D point in
barycentric coordinates which serves as a center in the interior of the triangle Tb;
default is M="CC" i.e. the circumcenter of tri}

\item{main}{An overall title for the plot (default="")}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges (default=NULL for both)}

\item{asp}{Numeric, giving the aspect ratio for y axis to x-axis y/x (default is NA),
see the official help for asp by typing "? asp"}
}
\value{
Plot of the AS proximity regions for points inside the triangle tri (and only the points outside tri)
}
\description{
Plots the points in and outside of the triangle tri and also the AS proximity regions
for points in data set Xp. AS proximity regions are defined with respect to the triangle tri,
so AS proximity regions are defined only for points inside the triangle tri and
vertex regions are based on the center M="CC" for
circumcenter of tri; or M=(m1,m2) in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric
coordinates in the interior of the triangle tri; default is M="CC" the circumcenter of tri.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
n<-10

#set.seed(1)
dat<-runif.tri(n,Tr)$g
#dat<-cbind(runif(n,1,2),runif(n,0,2))

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.2)
#M<-circ.cent.tri(Tr); #M<-"CC"; #M<-"CM"
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

if (dimension(M)==3) {M<-bary2cart(M,Tr)} #need to run this when M is given in barycentric coordinates

CC<-circ.cent.tri(Tr)
if (isTRUE(all.equal(M,CC)) || identical(M,"CC"))
{cent<-CC
D1<-(B+C)/2; D2<-(A+C)/2; D3<-(A+B)/2;
Ds<-rbind(D1,D2,D3)
cent.name<-"CC"
} else
{cent<-M
cent.name<-"M"
Ds<-cp2e.tri(Tr,M)
D1<-Ds[1,]; D2<-Ds[2,]; D3<-Ds[3,]
}

Xlim<-range(Tr[,1],dat[,1])
Ylim<-range(Tr[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plotArcsAStri(dat,Tr,M,main="arcs for AS-PCD",xlab="",ylab="",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))

#plotArcsAStri(dat,Tr,M,asp=1,main="arcs for AS-PCD",xlab="",ylab="",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
L<-rbind(cent,cent,cent); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)

txt<-rbind(Tr,cent,Ds)
xc<-txt[,1]#+c(-.02,.03,.03,.03,.05,-0.03,-.01)
yc<-txt[,2]#+c(.02,.02,.02,.07,.02,.05,-.06)
txt.str<-c("A","B","C",cent.name,"D1","D2","D3")
text(xc,yc,txt.str)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plot.regsASMT}}, \code{\link{plot.regsPEtri}}, \code{\link{plot.regsPEMT}},
\code{\link{plot.regsCStri}}, and \code{\link{plot.regsCSMT}}
}
