% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plot.regsCSint}
\alias{plot.regsCSint}
\title{The plot of the Central Similarity (CS) Proximity Regions for a general interval
(vertices jittered along y-coordinate) - one interval case}
\usage{
\method{plot}{regsCSint}(dat, t, c = 0.5, int, jit = 0.1, main = "",
  xlab = "", ylab = "", xlim = NULL, ylim = NULL,
  centers = FALSE)
}
\arguments{
\item{dat}{A set of 1D points for which CS proximity regions are to be constructed}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a); default=.5}

\item{int}{A vector of two real numbers representing an interval}

\item{main}{Title of the main heading of the plot}

\item{xlab, ylab}{Titles for the x and y axes, respectively (default="" for both)}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges.}

\item{centers}{A logical argument, if TRUE, plot includes the centers of the intervals
as vertical lines in the plot, else centers of the intervals are not plotted.}

\item{Jit}{A positive real number that determines the amount of jitter along the y-axis, default is .1 and
dat points are jittered according to U(-jit,jit) along the y-axis where jit=range of dat and proximity region
 intervals multiplied by Jit.)}
}
\value{
Plot of the CS proximity regions for 1D points in or outside the interval int
}
\description{
Plots the points in and outside of the interval int and also the CS proximity regions (which are also intervals).
CS proximity regions are constructed with expansion parameter t>0 and centrality parameter c in (0,1).
For better visualization, a uniform jitter from U(-jit,jit) (default for jit=.1 times range of proximity
regions and dat) is added to  the y-direction. centers is a logical argument, if TRUE, plot includes the
centers of the intervals as vertical lines in the plot, else centers of the intervals are not plotted.
See also (\insertCite{ceyhan:revstat-2016;textual}{pcds}).
}
\examples{
c<-.4
t<-2
a<-0; b<-10; int<-c(a,b)

IndNCSint(7,5,t,c,int)
IndNCSint(17,17,t,c,int)
IndNCSint(15,17,t,c,int)
IndNCSint(1,3,t,c,int)

IndNCSint(-17,17,t,c,int)

IndNCSint(3,5,t,c,int)
IndNCSint(3,3,t,c,int)
IndNCSint(4,5,t,c,int)
IndNCSint(a,5,t,c,int)

c<-.4
r<-2
a<-0; b<-10; int<-c(a,b)

IndNCSint(7,5,t,c,int)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plot.regsCSMI}}, \code{\link{plot.regsPEint}}, and \code{\link{plot.regsPE.MI}}
}
