% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plot.regsCSMI}
\alias{plot.regsCSMI}
\title{The plot of the Central Similarity (CS) Proximity Regions (vertices jittered along y-coordinate)
- multiple interval case}
\usage{
\method{plot}{regsCSMI}(Xp, Yp, t, c, jit = 0.1, main = "",
  xlab = "", ylab = "", xlim = NULL, ylim = NULL,
  centers = FALSE)
}
\arguments{
\item{Xp}{A set of 1D points for which CS proximity regions are plotted}

\item{Yp}{A set of 1D points which constitute the end points of the intervals which
partition the real line}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a); default=.5}

\item{main}{An overall title for the plot (default="")}

\item{xlab, ylab}{Titles for the x and y axes, respectively (default="" for both)}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges (default=NULL for both)}

\item{centers}{A logical argument, if TRUE, plot includes the centers of the intervals
as vertical lines in the plot, else centers of the intervals are not plotted.}

\item{Jit}{A positive real number that determines the amount of jitter along the y-axis, default is .1 and
Xp points are jittered according to U(-jit,jit) along the y-axis where jit=range of Xp and Yp and the
proximity regions (intervals) multiplied by Jit.)}
}
\value{
Plot of the CS proximity regions for 1D points located in the middle or end intervals
based on Yp points
}
\description{
Plots the points in and outside of the intervals based on Yp points and also the CS proximity regions
(which are also intervals). CS proximity region is constructed with expansion parameter t>0 and
centrality parameter c in (0,1). For better visualization, a uniform jitter from U(-jit,jit)
(default for jit=.1 times range of Xp and Yp and the proximity regions (intervals)) is added to the
y-direction. centers is a logical argument, if TRUE,
plot includes the centers of the intervals as vertical lines in the plot,
else centers of the intervals are not plotted.
See also (\insertCite{ceyhan:revstat-2016;textual}{pcds}).
}
\examples{
t<-2
c<-.4
a<-0; b<-10;
nx<-10; ny<-4

#set.seed(1)
Xp<-runif(nx,a,b)
#Xp<-runif(nx,a-5,b+5)
#Xp<-runif(nx,a+5,b+5)
#Xp<-runif(nx,a+10,b+10)
Yp<-runif(ny,a,b)

IM<-IncMatCS1D(Xp,Yp,t,c)
IM
dom.greedy(IM)
dom.exact(IM) #might take a long time depending on n
IndUBdom(IM,5)

Arcs<-ArcsCSMI(Xp,Yp,t,c)
Arcs
summary(Arcs)
plot(Arcs)

IncMatCS1D(Xp,Yp+10,t,c)

t<-2
c<-.4
a<-0; b<-10;
nx<-40; ny<-10
Xp<-runif(nx,a,b)
Yp<-runif(ny,a,b)

IncMatCS1D(Xp,Yp,t,c)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plot.regsCSint}} and \code{\link{plot.regsPEMI}}
}
