% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plot.regsCSMT}
\alias{plot.regsCSMT}
\title{The plot of the Central Similarity (CS) Proximity Regions for a 2D data set - multiple triangle case}
\usage{
\method{plot}{regsCSMT}(Xp, Yp, t, M = c(1, 1, 1), asp = NA,
  main = "", xlab = "", ylab = "", xlim = NULL, ylim = NULL)
}
\arguments{
\item{Xp}{A set of 2D points for which CS proximity regions are constructed}

\item{Yp}{A set of 2D points which constitute the vertices of the Delaunay triangles}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri or the circumcenter of tri}

\item{asp}{Numeric, giving the aspect ratio y/x (default is NA), see the official help for asp by
typing "? asp"}

\item{main}{An overall title for the plot (default="")}

\item{xlab, ylab}{Titles for the x and y axes, respectively (default="" for both)  (default="" for both)}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges (default=NULL for both)}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}
}
\value{
Plot of the Xp points, Delaunay triangles based on Yp and also the CS proximity regions
for Xp points inside the convex hull of Yp points
}
\description{
Plots the points in and outside of the Delaunay triangles based on Yp points which partition
the convex hull of Yp points and also plots the CS proximity regions
for Xp points and the Delaunay triangles based on Yp points.
CS proximity regions are contructed with respect to the Delaunay triangles with the expansion parameter t>0.
Edge regions in each triangle is based on the center M=(alpha,beta,gamma)
in barycentric coordinates in the interior of each Delaunay triangle
(default for M is c(1,1,1) which is the center of mass of the triangle).
See (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-CS,ceyhan:test2014;textual}{pcds}) more on the CS proximity regions.
Also see (\insertCite{okabe:2000,ceyhan:comp-geo-2010,renka:1996;textual}{pcds}) for more on Delaunay triangulation and the tripack algorithm.
}
\examples{
nx<-100
ny<-10

#set.seed(1)
Xp<-cbind(runif(nx,0,1),runif(nx,0,1))
#Xp<-cbind(runif(nx,0.5,1.5),runif(nx,0,1))
#Xp<-cbind(runif(nx,1,2),runif(nx,0,1))
#Xp<-c(.5,.5); Xp<-matrix(Xp,ncol=2)
Yp<-cbind(runif(ny,0,1),runif(ny,0,1))

M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
t<-1.5
#t<-2

Xlim<-range(Xp[,1],Yp[,1])
Ylim<-range(Xp[,2],Yp[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plotArcsCSMT(Xp,Yp,t,M,xlab="",ylab="",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plot.regsCStri}}, \code{\link{plot.regsASMT}} and \code{\link{plot.regsPEMT}}
}
