% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plot.regsPEint}
\alias{plot.regsPEint}
\title{The plot of the Proportional Edge (PE) Proximity Regions for a general interval
(vertices jittered along y-coordinate) - one interval case}
\usage{
\method{plot}{regsPEint}(dat, r, c = 0.5, int, Jit = 0.1, main = "",
  xlab = "", ylab = "", xlim = NULL, ylim = NULL,
  centers = FALSE)
}
\arguments{
\item{dat}{A set of 1D points for which PE proximity regions are to be constructed}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a); default=.5}

\item{int}{A vector of two real numbers representing an interval}

\item{Jit}{A positive real number that determines the amount of jitter along the y-axis, default is .1 and
dat points are jittered according to U(-jit,jit) along the y-axis where jit=range of dat and proximity region
 intervals multiplied by Jit.)}

\item{main}{Title of the main heading of the plot}

\item{xlab, ylab}{Titles for the x and y axes, respectively (default="" for both)}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges.}

\item{centers}{A logical argument, if TRUE, plot includes the centers of the intervals
as vertical lines in the plot, else centers of the intervals are not plotted.}
}
\value{
Plot of the PE proximity regions for 1D points in or outside the interval int
}
\description{
Plots the points in and outside of the interval int and also the PE proximity regions (which are also intervals).
PE proximity regions are constructed with expansion parameter r>=1 and centrality parameter c in (0,1).
For better visualization, a uniform jitter from U(-jit,jit) (default for jit=.1 times range of proximity
regions and dat) is added to  the y-direction. centers is a logical argument, if TRUE, plot includes the
centers of the intervals as vertical lines in the plot, else centers of the intervals are not plotted.
See also (\insertCite{ceyhan:metrika-2012;textual}{pcds}).
}
\examples{
c<-.4
r<-2
a<-0; b<-10; int<-c(a,b)

IndNPEint(7,5,r,c,int)
IndNPEint(17,17,r,c,int)
IndNPEint(15,17,r,c,int)
IndNPEint(1,3,r,c,int)

IndNPEint(-17,17,r,c,int)

IndNPEint(3,5,r,c,int)
IndNPEint(3,3,r,c,int)
IndNPEint(4,5,r,c,int)
IndNPEint(a,5,r,c,int)

c<-.4
r<-2
a<-0; b<-10; int<-c(a,b)
IndNPEint(7,5,r,c,int)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plot.regsPEMI}}, \code{\link{plot.regsCSint}}, and \code{\link{plot.regsCSMI}}
}
