% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plot.regsPEMI}
\alias{plot.regsPEMI}
\title{The plot of the Proportional Edge (PE) Proximity Regions (vertices jittered along y-coordinate)
- multiple interval case}
\usage{
\method{plot}{regsPEMI}(Xp, Yp, r, c, Jit = 0.1, main = "",
  xlab = "", ylab = "", xlim = NULL, ylim = NULL,
  centers = FALSE)
}
\arguments{
\item{Xp}{A set of 1D points for which PE proximity regions are plotted}

\item{Yp}{A set of 1D points which constitute the end points of the intervals which
partition the real line}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a); default=.5}

\item{Jit}{A positive real number that determines the amount of jitter along the y-axis, default is .1 and
Xp points are jittered according to U(-jit,jit) along the y-axis where jit=range of Xp and Yp and the
proximity regions (intervals) multiplied by Jit.)}

\item{main}{An overall title for the plot (default="")}

\item{xlab, ylab}{Titles for the x and y axes, respectively (default="" for both)}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges (default=NULL for both)}

\item{centers}{A logical argument, if TRUE, plot includes the centers of the intervals
as vertical lines in the plot, else centers of the intervals are not plotted.}
}
\value{
Plot of the PE proximity regions for 1D points located in the middle or end intervals
based on Yp points
}
\description{
Plots the points in and outside of the intervals based on Yp points and also the PE proximity regions
(i.e., intervals). PE proximity region is constructed with expansion parameter r>=1 and
centrality parameter c in (0,1). For better visualization, a uniform jitter from U(-jit,jit)
(default for jit=.1 times range of Xp and Yp and the proximity regions (intervals)) is added to the
y-direction. centers is a logical argument, if TRUE,
plot includes the centers of the intervals as vertical lines in the plot,
else centers of the intervals are not plotted.
See also (\insertCite{ceyhan:metrika-2012;textual}{pcds}).
}
\examples{
a<-0; b<-10;
#a<-runif(1); b<-runif(1); int<-range(a,b); a<-int[1]; b<-int[2]
nx<-40; ny<-10

#set.seed(1)
Xp<-runif(nx,a,b)
Yp<-runif(ny,a,b)

plot.intervals(Xp,Yp,xlab="",ylab="")
plot.intervals(Xp,Yp+10,xlab="",ylab="")

}
\seealso{
\code{\link{plot.regsPEMI}}, \code{\link{plot.regsCSint}}, and \code{\link{plot.regsCSMI}}
}
