% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plot.regsPEstdtetra}
\alias{plot.regsPEstdtetra}
\title{The plot of the Proportional Edge (PE) Proximity Regions for a 3D data set - standard
regular tetrahedron case}
\usage{
\method{plot}{regsPEstdtetra}(Xp, r, main = "", xlab = "", ylab = "",
  zlab = "", xlim = NULL, ylim = NULL, zlim = NULL)
}
\arguments{
\item{Xp}{A set of 3D points for which PE proximity regions are constructed}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{main}{An overall title for the plot (default="")}

\item{xlab, ylab, zlab}{titles for the x, y and z axes, respectively  (default="" for all)}

\item{xlim, ylim, zlim}{Numeric vectors of length 2, giving the x-, y- and z-coordinate ranges (default=NULL for all)}
}
\value{
Plot of the PE proximity regions for points inside the standard regular tetrahedron th
(and just the points outside th)
}
\description{
Plots the points in and outside of the standard regular tetrahedron
th=T((0,0,0),(1,0,0),(1/2,sqrt(3)/2,0),(1/2,sqrt(3)/6,sqrt(6)/3)) and also the PE proximity regions
for points in data set Xp. PE proximity regions are defined with respect to the standard regular tetrahedron th
with expansion parameter r>=1, so PE proximity regions are defined only for points inside th.
Vertex regions are based on circumcenter (which is equivalent to the center of mass for the standard
regular tetrahedron) of th.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010;textual}{pcds}).
}
\examples{
A<-c(0,0,0); B<-c(1,0,0); C<-c(1/2,sqrt(3)/2,0); D<-c(1/2,sqrt(3)/6,sqrt(6)/3)
tetra<-rbind(A,B,C,D)

n<-20
dat<-runif.stdtetra(n)$g
r<-1.5
IndNPEstdtetra(dat[1,],dat[1,],r)

IndNPEstdtetra(dat[1,],dat[5,],r)

IndNPEstdtetra(c(.4,.4,.4),c(.5,.5,.5),r)

#or try
RV<-rv.tetraCC(dat[1,],tetra)$rv
IndNPEstdtetra(dat[1,],dat[5,],r,rv=RV)

IndNPEstdtetra(dat[1,],c(-1,-1,-1),r,rv=NULL)

IndNPEstdtetra(c(-1,-1,-1),dat[1,],r,rv=NULL)
IndNPEstdtetra(c(-1,-1,-1),c(-1,-1,-1),r)

IndNPEstdtetra(dat[1,],dat[5,],r)

P1<-c(.1,.1,.1)
P2<-c(.5,.5,.5)
IndNPEstdtetra(P1,P2,r)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plot.regsPEMT}}, \code{\link{plot.regsAStri}}, \code{\link{plot.regsASMT}},
\code{\link{plot.regsCStri}}, and \code{\link{plot.regsCSMT}}
}
