% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plot.regsPEtetra}
\alias{plot.regsPEtetra}
\title{The plot of the Proportional Edge (PE) Proximity Regions for a 3D data set - one tetrahedron case}
\usage{
\method{plot}{regsPEtetra}(Xp, r, th, M = "CM", main = "", xlab = "",
  ylab = "", zlab = "", xlim = NULL, ylim = NULL, zlim = NULL)
}
\arguments{
\item{Xp}{A set of 3D points for which PE proximity regions are constructed}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{th}{Four 3D points, stacked row-wise, each row representing a vertex of the tetrahedron}

\item{M}{The center to be used in the construction of the vertex regions in the tetrahedron, th.
Currently it only takes "CC" for circumcenter and "CM" for center of mass; default="CM"}

\item{main}{An overall title for the plot (default="")}

\item{xlab, ylab, zlab}{Titles for the x, y and z axes, respectively  (default="" for all)}

\item{xlim, ylim, zlim}{Numeric vectors of length 2, giving the x-, y- and z-coordinate ranges
(default=NULL for all)}
}
\value{
Plot of the PE proximity regions for points inside the tetrahedron th
(and just the points outside th)
}
\description{
Plots the points in and outside of the tetrahedron th and also the PE proximity regions  (which are also
tetrahedrons) for points inside the tetrahedron th. PE proximity regions are constructed with respect to
tetrahedron th with expansion parameter r>=1 and vertex regions are based on the center M which is
circumcenter ("CC") or center of mass ("CM") of th with default="CM", so PE proximity regions are defined
only for points inside the tetrahedron th.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010;textual}{pcds}).
}
\examples{
A<-c(0,0,0); B<-c(1,0,0); C<-c(1/2,sqrt(3)/2,0); D<-c(1/2,sqrt(3)/6,sqrt(6)/3)
#A<-runif(3); B<-runif(3); C<-runif(3); D<-runif(3)
#A<-runif(3,1,100); B<-runif(3,1,100); C<-runif(3,1,100); D<-runif(3,1,100)
tetra<-rbind(A,B,C,D)
n<-10
dat<-runif.tetra(n,tetra)$g
#dat<-c(.5,.5,.5)

M<-"CM"
#M<-"CC"
r<-1.5

IM<-IncMatPEtetra(dat,tetra,r=1.25)#uses the default M="CM"
IM<-IncMatPEtetra(dat,tetra,r=1.25,M)
IM
dom.greedy(IM)
IndUBdom(IM,3)
dom.exact(IM) #this might take a long time for large n

IncMatPEtetra(dat,tetra,r=1.5)
IncMatPEtetra(dat,tetra,r=2)

r<-2
IncMatPEtetra(dat,tetra,r,M)

dat.fr<-data.frame(a=dat)
IncMatPEtetra(dat.fr,tetra,r,M)

dat.fr<-data.frame(a=tetra)
IncMatPEtetra(dat,dat.fr,r,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plot.regsPEstdtetra}}, \code{\link{plot.regsPEtri}} and \code{\link{plot.regsPEint}}
}
