% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plotArcsASMT}
\alias{plotArcsASMT}
\title{The plot of the arcs of Arc Slice Proximity Catch Digraph (AS-PCD) for a 2D data set -
multiple triangle case}
\usage{
plotArcsASMT(Xp, Yp, M = "CC", asp = NA, main = "", xlab = "",
  ylab = "", xlim = NULL, ylim = NULL)
}
\arguments{
\item{Xp}{A set of 2D points which constitute the vertices of the AS-PCD}

\item{Yp}{A set of 2D points which constitute the vertices of the Delaunay triangulation. The Delaunay
triangles partition the convex hull of Yp points.}

\item{M}{"CC" which stand for circumcenter of each Delaunay triangle tri or 3D point in barycentric
coordinates which serves as a center in the interior of each Delaunay triangle;
default is M="CC" i.e. the circumcenter of each triangle}

\item{asp}{Numeric, giving the aspect ratio for y axis to x-axis y/x (default is NA),
see the official help for asp by typing "? asp"}

\item{main}{An overall title for the plot (default="")}

\item{xlab, ylab}{Titles for the x and y axes, respectively (default="" for both)}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges (default=NULL for both)}
}
\value{
A plot of the arcs of the AS-PCD for a 2D data set Xp where AS proximity regions
are defined with respect to the Delaunay triangles based on Yp points; also plots the Delaunay triangles
based on Yp points
}
\description{
Plots the arcs of AS-PCD whose vertices are the data points in Xp and Delaunay triangles based on Yp points.
AS proximity regions are constructed with respect to the Delaunay triangles based on Yp points, i.e.,
AS proximity regions are defined only for Xp points inside the convex hull of Yp points.
That is, arcs may exist for Xp points only inside the convex hull of Yp points.
Vertex regions are based on the center M="CC"
for circumcenter of each Delaunay triangle or M=(alpha,beta,gamma) in barycentric coordinates in the
interior of each Delaunay triangle; default is M="CC" i.e. circumcenter of each triangle.
See (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}) for more on AS-PCDs.
Also see (\insertCite{okabe:2000,ceyhan:comp-geo-2010,renka:1996;textual}{pcds}) for more on Delaunay triangulation and the tripack algorithm.
}
\examples{
nx<-100
ny<-10

#set.seed(1)
Xp<-cbind(runif(nx,0,1),runif(nx,0,1))
#Xp<-cbind(runif(nx,0.5,1.5),runif(nx,0,1))
#Xp<-cbind(runif(nx,1,2),runif(nx,0,1))
Yp<-cbind(runif(ny,0,1),runif(ny,0,1))
#Xp<-c(.5,.5)

M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-"CC"
#M<-c(1.3,1.3)

ArcsASMT(Xp,Yp,M)

Arcs<-ArcsASMT(Xp,Yp,M)
Arcs
summary(Arcs)
plot(Arcs)

S<-Arcs$S
E<-Arcs$E
DT<-tri.mesh(Yp[,1],Yp[,2],duplicate="remove")

Xlim<-range(Xp[,1],Yp[,1])
Ylim<-range(Xp[,2],Yp[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Xp,main=" ", xlab=" ", ylab=" ",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05),pch=".",cex=3)
plot.tri(DT, add=T, do.points=T)
arrows(S[,1], S[,2], E[,1], E[,2], length = 0.1, col= 4)

ArcsASMT(Xp,rbind(Yp,Yp),M)

dat.fr<-data.frame(a=Xp)
ArcsASMT(dat.fr,Yp,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plotArcsAStri}}, \code{\link{plotArcsPEtri}}, \code{\link{plotArcsPEMT}},
\code{\link{plotArcsCStri}}, and \code{\link{plotArcsCSMT}},
}
