% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plotArcsCStri}
\alias{plotArcsCStri}
\title{The plot of the arcs of Central Similarity Proximity Catch Digraph (CS-PCD) for a
2D data set - one triangle case}
\usage{
plotArcsCStri(Xp, tri, t, M = c(1, 1, 1), asp = NA, main = "",
  xlab = "", ylab = "", xlim = NULL, ylim = NULL)
}
\arguments{
\item{Xp}{A set of 2D points which constitute the vertices of the CS-PCD}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri;
default is M=c(1,1,1) i.e. the center of mass of tri}

\item{asp}{Numeric, giving the aspect ratio y/x (default is NA), see the official help for asp by
typing "? asp"}

\item{main}{An overall title for the plot (default="")}

\item{xlab, ylab}{Titles for the x and y axes, respectively (default="" for both)}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges (default=NULL for both)}
}
\value{
A plot of the arcs of the CS-PCD whose vertices are the points in data set Xp and the triangle tri
}
\description{
Plots the arcs of CS-PCD whose vertices are the data points, Xp and the triangle tri. CS proximity regions
are constructed with respect to the triangle tri with expansion parameter t>0, i.e. arcs may exist only
for Xp points inside the triangle tri. Edge regions are based on center M=(m1,m2) in Cartesian coordinates
or M=(alpha,beta,gamma) in barycentric coordinates in the interior of the triangle tri; default
is M=c(1,1,1) i.e. the center of mass of tri.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-CS,ceyhan:test2014;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
n<-10

#set.seed(1)
dat<-runif.tri(n,Tr)$g
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,2))
#dat<-c(.5,.5)

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.0)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

IM<-IncMatCStri(dat,Tr,t=1.25,M)
IM
dom.greedy(IM)
dom.exact(IM)
IndUBdom(IM,3)

IncMatCStri(dat,Tr,t=1.5,M)
IncMatCStri(dat,Tr,t=2,M)

t<-2
IncMatCStri(dat,Tr,t,M)

dat.fr<-data.frame(a=dat)
IncMatCStri(dat.fr,Tr,t,M)

dat.fr<-data.frame(a=Tr)
IncMatCStri(dat,dat.fr,t,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plotArcsCSMT}}, \code{\link{plotArcsPEtri}} and \code{\link{plotArcsAStri}}
}
