% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plotArcsPE1D}
\alias{plotArcsPE1D}
\title{The plot of the arcs of Proportional Edge Proximity Catch Digraphs (PE-PCDs) for 1D data
(vertices jittered along y-coordinate) - multiple interval case}
\usage{
plotArcsPE1D(Xp, Yp, r, c, Jit = 0.1, main = "", xlab = "",
  ylab = "", xlim = NULL, ylim = NULL, centers = FALSE)
}
\arguments{
\item{Xp}{A vector of 1D points constituting the vertices of the PE-PCD}

\item{Yp}{A vector of 1D points constituting the end points of the intervals}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a)}

\item{Jit}{A positive real number that determines the amount of jitter along the y-axis, default is .1 and
Xp points are jittered according to U(-jit,jit) along the y-axis where jit=range of Xp and Yp multiplied by
Jit.)}

\item{main}{Title of the main heading of the plot}

\item{xlab, ylab}{Titles of the x and y axes in the plot (default="" for both)}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges (default=NULL for both)}

\item{centers}{A logical argument, if TRUE, plot includes the centers of the intervals
as vertical lines in the plot, else centers of the intervals are not plotted.}
}
\value{
A plot of the arcs of PE-PCD whose vertices are the 1D data set Xp in which vertices are jittered
along y-axis for better visualization.
}
\description{
Plots the arcs of PE-PCD whose vertices are the 1D points, Xp. PE proximity regions are constructed with
expansion parameter r>=1 and centrality parameter c in (0,1) and the intervals are based on Yp points (i.e.
the intervalization is based on Yp points). That is, data set Xp
constitutes the vertices of the digraph and Yp determines the end points of the intervals.
For better visualization, a uniform jitter from U(-jit,jit) (default for jit=.1) is added to
the y-direction where jit=range of Xp and Yp multiplied by Jit with default for Jit=.1).
centers is a logical argument, if TRUE, plot includes the centers of the intervals
as vertical lines in the plot, else centers of the intervals are not plotted.
See also (\insertCite{ceyhan:metrika-2012;textual}{pcds}).
}
\examples{
r<-2
c<-.4
a<-0; b<-10; int<-c(a,b)
#c<-runif(1)
#a<-runif(1); b<-runif(1); int<-range(a,b); a<-int[1]; b<-int[2]
nx<-40; ny<-10

#set.seed(1)
xr<-range(a,b)
xf<-(xr[2]-xr[1])*.1
Xp<-runif(nx,a-xf,b+xf)
#Xp<-runif(nx,a+5,b+5)
#Xp<-runif(nx,a+10,b+10)
Yp<-runif(ny,a,b)

Arcs<-ArcsPE1D(Xp,Yp,r,c)
Arcs
summary(Arcs)
plot(Arcs)

S<-Arcs$S
E<-Arcs$E

ArcsPE1D(Xp,Yp,r,c)
ArcsPE1D(Xp,Yp,r=.5,c)

jit<-.1
yjit<-runif(nx,-jit,jit)

Xlim<-range(a,b,Xp,Yp)
xd<-Xlim[2]-Xlim[1]

plot(cbind(a,0),pch=".",main="arcs for points in mid and end intervals ", xlab=" ", ylab=" ",
     xlim=Xlim+xd*c(-.05,.05),ylim=3*c(-jit,jit))
abline(h=0,lty=1)
points(Xp, yjit,pch=".",cex=3)
abline(v=Yp,lty=2)
arrows(S, yjit, E, yjit, length = .05, col= 4)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plot.CSPCD.1D}}
}
