% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plotArcsPEMT}
\alias{plotArcsPEMT}
\title{The plot of the arcs of Proportional Edge Proximity Catch Digraph (PE-PCD) for a
2D data set - multiple triangle case}
\usage{
plotArcsPEMT(Xp, Yp, r, M = c(1, 1, 1), asp = NA, main = "",
  xlab = "", ylab = "", xlim = NULL, ylim = NULL)
}
\arguments{
\item{Xp}{A set of 2D points which constitute the vertices of the PE-PCD}

\item{Yp}{A set of 2D points which constitute the vertices of the Delaunay triangles}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >=1}

\item{M}{A 3D point in barycentric coordinates which serves as a center in the interior of each Delaunay
triangle or circumcenter of each Delaunay triangle (for this argument should be set as M="CC"),
default for M is c(1,1,1) which is the center of mass of each triangle}

\item{asp}{Numeric, giving the aspect ratio y/x (default is NA), see the official help for asp by typing "? asp"}

\item{main}{An overall title for the plot (default="")}

\item{xlab, ylab}{Titles for the x and y axes, respectively (default="" for both)}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges (default=NULL for both)}
}
\value{
A plot of the arcs of the PE-PCD whose vertices are the points in data set Xp and the Delaunay
triangles based on Yp points
}
\description{
Plots the arcs of Proportional Edge Proximity Catch Digraph (PE-PCD) whose vertices are the data
points in Xp in the multiple triangle case and the Delauany triangles based on Yp points.
PE proximity regions are defined with respect to the Delaunay triangles based on Yp points with
expansion parameter r>=1 and vertex regions in each triangle are based on the center M=(alpha,beta,gamma)
in barycentric coordinates in the interior of each Delaunay triangle or based on circumcenter of
each Delaunay triangle (default for M is c(1,1,1) which is the center of mass of the triangle).
Each Delaunay triangle is first converted to an (unscaled) basic triangle so that M will be the same
type of center for each Delaunay triangle (this conversion is not necessary when M is CM).
Convex hull of Yp is partitioned by the Delaunay triangles based on Yp points
(i.e., multiple triangles are the set of these Delaunay triangles whose union constitutes the
convex hull of Yp points). Loops are not allowed so arcs are only possible
for points inside the convex hull of Yp points.
See (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-PE,ceyhan:dom-num-NPE-Spat2011;textual}{pcds}) for more
on the PE-PCDs.
Also see (\insertCite{okabe:2000,ceyhan:comp-geo-2010,renka:1996;textual}{pcds}) for more on Delaunay triangulation and the tripack algorithm.
}
\examples{
nx<-100
ny<-10

#set.seed(1)
Xp<-cbind(runif(nx,0,1),runif(nx,0,1))
#Xp<-cbind(runif(nx,0.5,1.5),runif(nx,0,1))
#Xp<-cbind(runif(nx,1,2),runif(nx,0,1))
Yp<-cbind(runif(ny,0,1),runif(ny,0,1))

DT<-tri.mesh(Yp[,1],Yp[,2],duplicate="remove")

Xlim<-range(Xp[,1],Yp[,1])
Ylim<-range(Xp[,2],Yp[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

Xch<-XinCHY(Xp,Yp)

plot(Xp,main=" ", xlab=" ", ylab=" ",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05),pch=".",cex=3)
convex.hull(DT,plot.it = T, add = T) # or try polygon(Yp[ch$i,])
points(Xch,pch=4,col="red")

XinCHY(Xp,Yp)

XinCHY(Xp,rbind(Yp,Yp))

dat.fr<-data.frame(a=Xp)
XinCHY(dat.fr,Yp)

dat.fr<-data.frame(a=Yp)
XinCHY(Xp,dat.fr)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plotArcsPEtri}}, \code{\link{plotArcsASMT}}, and \code{\link{plotArcsCSMT}}
}
