% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plotArcsPEtri}
\alias{plotArcsPEtri}
\title{The plot of the arcs of Proportional Edge Proximity Catch Digraph (PE-PCD) for a
2D data set - one triangle case}
\usage{
plotArcsPEtri(Xp, tri, r, M = c(1, 1, 1), asp = NA, main = "",
  xlab = "", ylab = "", xlim = NULL, ylim = NULL)
}
\arguments{
\item{Xp}{A set of 2D points which constitute the vertices of the PE-PCD}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri or the circumcenter of tri;
default is M=c(1,1,1) i.e. the center of mass of tri}

\item{asp}{Numeric, giving the aspect ratio y/x (default is NA), see the official help for asp by
typing "? asp"}

\item{main}{An overall title for the plot (default="")}

\item{xlab, ylab}{Titles for the x and y axes, respectively (default="" for both)}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges (default=NULL for both)}
}
\value{
A plot of the arcs of the PE-PCD whose vertices are the points in data set Xp and the triangle tri
}
\description{
Plots the arcs of PE-PCD whose vertices are the data points, Xp and the triangle tri. PE proximity regions
are constructed with respect to the triangle tri with expansion parameter r>=1, i.e. arcs may exist only
for Xp points inside the triangle tri. Vertex regions are based on center M=(m1,m2) in Cartesian coordinates
or M=(alpha,beta,gamma) in barycentric coordinates in the interior of the triangle tri
or based on the circumcenter of tri; default is M=c(1,1,1) i.e. the center of mass of tri.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-PE,ceyhan:dom-num-NPE-Spat2011;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
n<-10

#set.seed(1)
dat<-runif.tri(n,Tr)$g
#dat<-cbind(runif(n,1,2),runif(n,0,2))
#dat<-cbind(runif(n),runif(n))
#dat<-c(1.4,1)

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.0)
#M<-circ.cent.tri(Tr)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)
r<-1.5
#r<-2

ArcsPEtri(dat,Tr,r,M)

Arcs<-ArcsPEtri(dat,Tr,r,M)
Arcs
summary(Arcs)
plot(Arcs)

S<-Arcs$S
E<-Arcs$E

Xlim<-range(Tr[,1],dat[,1],M[1])
Ylim<-range(Tr[,2],dat[,2],M[2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

Ds<-cp2e.tri(Tr,M)

if (dimension(M)==3) {M<-bary2cart(M,Tr)} #need to run this when M is given in barycentric coordinates

plot(Tr,pch=".",xlab="",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Tr)
points(dat,pch=1,col=1)
L<-rbind(M,M,M); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)
arrows(S[,1], S[,2], E[,1], E[,2], length = 0.1, col= 4)

txt<-rbind(Tr,M,Ds)
xc<-txt[,1]#+c(-.02,.03,.02,.03,.04,-.03,-.01)
yc<-txt[,2]#+c(.02,.02,.03,.06,.04,.05,-.07)
txt.str<-c("A","B","C","M","D1","D2","D3")
text(xc,yc,txt.str)

r<-2
ArcsPEtri(dat,Tr,r,M)

dat.fr<-data.frame(a=dat)
ArcsPEtri(dat.fr,Tr,r,M)

dat.fr<-data.frame(a=Tr)
ArcsPEtri(dat,dat.fr,r,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plotArcsPEMT}}, \code{\link{plotArcsAStri}} and \code{\link{plotArcsASMT}},
}
