% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{PlotIntervals}
\alias{PlotIntervals}
\title{The plot of the subintervals based on \code{Yp} points together with \code{Xp} points}
\usage{
PlotIntervals(
  Xp,
  Yp,
  main = "",
  xlab = "",
  ylab = "",
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{Xp}{A set of 1D points whose scatter-plot is provided.}

\item{Yp}{A set of 1D points which constitute the end points of the intervals which
partition the real line.}

\item{main}{An overall title for the plot (default="")}

\item{xlab, ylab}{Titles for the x and y axes, respectively (default="" for both).}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges (default=NULL for both)}

\item{...}{Additional plot parameters}
}
\value{
Plot of the intervals based on \code{Yp} points and also scatter plot of \code{Xp} points
}
\description{
Plots the \code{Xp} points and the intervals based on \code{Yp} points
points.
}
\examples{
a<-0; b<-10;
#a<-runif(1); b<-runif(1); int<-range(a,b); a<-int[1]; b<-int[2]
nx<-40; ny<-10

#set.seed(1)
Xp<-runif(nx,a,b)
Yp<-runif(ny,a,b)

PlotIntervals(Xp,Yp,xlab="",ylab="")
PlotIntervals(Xp,Yp+10,xlab="",ylab="")

}
\seealso{
\code{\link{PlotRegsPEMI}} and \code{\link{plotDeltri}}
}
