% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plotPEarcsTri}
\alias{plotPEarcsTri}
\title{The plot of the arcs of Proportional Edge Proximity Catch Digraph (PE-PCD) for a
2D data set - one triangle case}
\usage{
plotPEarcsTri(
  Xp,
  tri,
  r,
  M = c(1, 1, 1),
  asp = NA,
  main = "",
  xlab = "",
  ylab = "",
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{Xp}{A set of 2D points which constitute the vertices of the PE-PCD}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be \eqn{\ge 1}}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle \code{tri} or the circumcenter of \code{tri};
default is \eqn{M=(1,1,1)} i.e. the center of mass of \code{tri}}

\item{asp}{a numeric value, giving the aspect ratio y/x (default is \code{NA}), see the official help for \code{asp} by
typing "? asp"}

\item{main}{An overall title for the plot (default="")}

\item{xlab, ylab}{Titles for the x and y axes, respectively (default="" for both)}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges (default=\code{NULL} for both)}

\item{\dots}{Additional \code{plot} parameters}
}
\value{
A plot of the arcs of the PE-PCD whose vertices are the points in data set \code{Xp} and the triangle \code{tri}
}
\description{
Plots the arcs of PE-PCD whose vertices are the data points, \code{Xp} and the triangle \code{tri}. PE proximity regions
are constructed with respect to the triangle \code{tri} with expansion parameter \eqn{r \ge 1}, i.e. arcs may exist only
for \code{Xp} points inside the triangle \code{tri}.

Vertex regions are based on center \eqn{M=(m_1,m_2)} in Cartesian coordinates
or \eqn{M=(\alpha,\beta,\gamma)} in barycentric coordinates in the interior of the triangle \code{tri}
or based on the circumcenter of \code{tri}; default is \eqn{M=(1,1,1)} i.e. the center of mass of \code{tri}.

See also (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-PE,ceyhan:dom-num-NPE-Spat2011;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#try also the following
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);

Tr<-rbind(A,B,C);
n<-10 #try also n<-20

set.seed(1)
dat<-runif.tri(n,Tr)$g
#try also the following
#dat<-cbind(runif(n,1,2),runif(n,0,2))
#dat<-cbind(runif(n),runif(n))
#dat<-c(.5,.5)

M<-as.numeric(runif.tri(1,Tr)$g)
#try also the following
#M<-c(1.6,1.0)
#M<-circ.cent.tri(Tr)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

r<-1.5 #try also #r<-2

ifelse(isTRUE(all.equal(M,circ.cent.tri(Tr))),
Ds<-rbind((B+C)/2,(A+C)/2,(A+B)/2),Ds<-cp2e.tri(Tr,M))

dat<-matrix(dat,ncol=2)

Xlim<-range(Tr[,1],dat[,1],M[1])
Ylim<-range(Tr[,2],dat[,2],M[2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

if (dimension(M)==3) {M<-bary2cart(M,Tr)}
#need to run this when M is given in barycentric coordinates

plotPEarcsTri(dat,Tr,r,M,main="arcs of PE-PCD with r=1.5",
xlab="",ylab="",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
if (dimension(M)==3) {M<-bary2cart(M,Tr)}
#need to run this when M is given in barycentric coordinates
L<-rbind(M,M,M); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)

txt<-rbind(Tr,M,Ds)
xc<-txt[,1]
yc<-txt[,2]
txt.str<-c("A","B","C","M","D1","D2","D3")
text(xc,yc,txt.str)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plotPEarcsMT}}, \code{\link{plotASarcsTri}} and \code{\link{plotASarcsMT}}
}
