% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{PlotRegsPEMI}
\alias{PlotRegsPEMI}
\title{The plot of the Proportional Edge (PE) Proximity Regions (vertices jittered along y-coordinate)
- multiple interval case}
\usage{
PlotRegsPEMI(
  Xp,
  Yp,
  r,
  c,
  Jit = 0.1,
  main = "",
  xlab = "",
  ylab = "",
  xlim = NULL,
  ylim = NULL,
  centers = FALSE,
  ...
)
}
\arguments{
\item{Xp}{A set of 1D points for which PE proximity regions are plotted}

\item{Yp}{A set of 1D points which constitute the end points of the intervals which
partition the real line}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be \eqn{\ge 1}}

\item{c}{A positive real number in \eqn{(0,1)} parameterizing the center inside \eqn{int=(a,b)}.
For the interval, \eqn{int=(a,b)}, the parameterized center is \eqn{M_c=a+c(b-a)}; default=.5}

\item{Jit}{A positive real number that determines the amount of jitter along the y-axis, default is .1 and
\code{Xp} points are jittered according to \eqn{U(-Jit,Jit)} along the y-axis where \code{Jit} equals to the range of \code{Xp} and \code{Yp} and the
proximity regions (intervals) multiplied by \code{Jit}.)}

\item{main}{An overall title for the plot (default="")}

\item{xlab, ylab}{Titles for the x and y axes, respectively (default="" for both)}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges (default=NULL for both)}

\item{centers}{A logical argument, if TRUE, plot includes the centers of the intervals
as vertical lines in the plot, else centers of the intervals are not plotted (default is FALSE).}

\item{...}{Additional plot parameters}
}
\value{
Plot of the PE proximity regions for 1D points located in the middle or end intervals
based on \code{Yp} points
}
\description{
Plots the points in and outside of the intervals based on \code{Yp} points and also the PE proximity regions
(i.e., intervals). PE proximity region is constructed with expansion parameter \eqn{r \ge 1} and
centrality parameter \code{c} in \eqn{(0,1)}. For better visualization, a uniform jitter from \eqn{U(-Jit,Jit)}
(default for \eqn{Jit=.1} times range of \code{Xp} and \code{Yp} and the proximity regions (intervals)) is added to the
y-direction. \code{centers} is a logical argument, if TRUE,
plot includes the centers of the intervals as vertical lines in the plot,
else centers of the intervals are not plotted.
See also (\insertCite{ceyhan:metrika-2012;textual}{pcds}).
}
\examples{
r<-2
c<-.4
a<-0; b<-10;
#c<-runif(1)
#a<-runif(1); b<-runif(1); int<-range(a,b); a<-int[1]; b<-int[2]
nx<-40; ny<-10

#set.seed(1)
xr<-range(a,b)
xf<-(xr[2]-xr[1])*.1
Xp<-runif(nx,a-xf,b+xf)
#Xp<-runif(nx,a-5,b+5)
#Xp<-runif(nx,a+5,b+5)
#Xp<-runif(nx,a+10,b+10)
Yp<-runif(ny,a,b)

PlotRegsPEMI(Xp,Yp,r,c,xlab="",ylab="")
PlotRegsPEMI(Xp,Yp+10,r,c,xlab="",ylab="")

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{PlotRegsPEMI}}, \code{\link{PlotRegsCSint}}, and \code{\link{PlotRegsCSMI}}
}
