% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFunctions.R
\name{print.Lines3D}
\alias{print.Lines3D}
\title{Print a \code{Lines3D} \code{object}}
\usage{
\method{print}{Lines3D}(x, ...)
}
\arguments{
\item{x}{A \code{Lines3D} \code{object}.}

\item{\dots}{Additional arguments for the S3 method \code{'print'}.}
}
\value{
The \code{call} of the \code{object} of class \code{"Lines3D"},
the \code{coefficients} of the line (in the form: \code{x=x0 + A*t}, \code{y=y0 + B*t}, and \code{z=z0 + C*t}),
and the initial point together with the direction vector.
}
\description{
Prints the \code{call} of the \code{object} of class \code{"Lines3D"},
the \code{coefficients} of the line (in the form: \code{x=x0 + A*t}, \code{y=y0 + B*t}, and \code{z=z0 + C*t}),
and the initial point together with the direction vector.
}
\examples{
\dontrun{
P<-c(1,10,3); Q<-c(1,1,3);
vecs<-rbind(P,Q)
Line3D(P,Q,.1)
Line3D(P,Q,.1,dir.vec=FALSE)

tr<-range(vecs);
tf<-(tr[2]-tr[1])*.1 #how far to go at the lower and upper ends in the x-coordinate
tsq<-seq(-tf*10-tf,tf*10+tf,l=3) #try also l=10, 20 or 100

lnPQ3D<-Line3D(P,Q,tsq)
lnPQ3D
print(lnPQ3D)
}

}
\seealso{
\code{\link{summary.Lines3D}}, \code{\link{print.summary.Lines3D}}, and \code{\link{plot.Lines3D}}
}
