% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxRFuncs4PCDClasses.r
\name{print.Patterns}
\alias{print.Patterns}
\title{Print a \code{Patterns} object}
\usage{
\method{print}{Patterns}(x, ...)
}
\arguments{
\item{x}{A \code{Patterns} object.}

\item{\dots}{Additional arguments for the S3 method '\code{print}'.}
}
\value{
The \code{call} of the object of class '\code{Patterns}'
and also the \code{type} (or description) of the pattern).
}
\description{
Prints the \code{call} of the object of class '\code{Patterns}'
and also the \code{type} (or description) of the pattern).
}
\examples{
nx<-20; #try also 100 and 1000
ny<-3; #try also 1
e<-.15;
Y<-cbind(runif(ny),runif(ny)) #with default bounding box (i.e., unit square)

Xdt<-rseg.disc(nx,Y,e)
Xdt
print(Xdt)

}
\seealso{
\code{\link{summary.Patterns}}, \code{\link{print.summary.Patterns}}, and \code{\link{plot.Patterns}}
}
