% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxRFuncs4PCDClasses.r
\name{print.summary.Extrema}
\alias{print.summary.Extrema}
\title{Print a summary of a \code{Extrema} object}
\usage{
\method{print}{summary.Extrema}(x, ...)
}
\arguments{
\item{x}{object of class "\code{summary.Extrema}", generated by \code{summary.Extrema}.}

\item{\dots}{Additional parameters for \code{print}.}
}
\value{
None
}
\description{
Prints some information about the \code{object}.
}
\seealso{
\code{\link{print.Extrema}}, \code{\link{summary.Extrema}}, and \code{\link{plot.Extrema}}
}
