% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxRFuncs4PCDClasses.r
\name{print.summary.Lines}
\alias{print.summary.Lines}
\title{Print a summary of a \code{Lines} object}
\usage{
\method{print}{summary.Lines}(x, ...)
}
\arguments{
\item{x}{object of class "\code{summary.Lines}", generated by \code{summary.Lines}.}

\item{\dots}{Additional parameters for \code{print}.}
}
\value{
None
}
\description{
Prints some information about the \code{object}.
}
\seealso{
\code{\link{print.Lines}}, \code{\link{summary.Lines}}, and \code{\link{plot.Lines}}
}
