% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFunctions.R
\name{print.summary.NumArcs}
\alias{print.summary.NumArcs}
\title{Print a summary of a \code{NumArcs} \code{object}}
\usage{
\method{print}{summary.NumArcs}(x, ...)
}
\arguments{
\item{x}{An \code{object} of class \code{"summary.NumArcs"}, generated by \code{summary.NumArcs}.}

\item{\dots}{Additional parameters for \code{print}.}
}
\value{
None
}
\description{
Prints some information about the \code{object}.
}
\seealso{
\code{\link{print.NumArcs}}, \code{\link{summary.NumArcs}},
and \code{\link{plot.NumArcs}}
}
