% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxRFuncs4PCDClasses.r
\name{print.summary.Patterns}
\alias{print.summary.Patterns}
\title{Print a summary of a \code{Patterns} object}
\usage{
\method{print}{summary.Patterns}(x, ...)
}
\arguments{
\item{x}{object of class "\code{summary.Patterns}", generated by \code{summary.Patterns}.}

\item{\dots}{Additional parameters for \code{print}.}
}
\value{
None
}
\description{
Prints some information about the \code{object}.
}
\seealso{
\code{\link{print.Patterns}}, \code{\link{summary.Patterns}}, and \code{\link{plot.Patterns}}
}
