% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFunctions.R
\name{print.summary.PCDs}
\alias{print.summary.PCDs}
\title{Print a summary of a \code{PCDs} \code{object}}
\usage{
\method{print}{summary.PCDs}(x, ...)
}
\arguments{
\item{x}{An \code{object} of class \code{"summary.PCDs"}, generated by \code{summary.PCDs}.}

\item{\dots}{Additional parameters for \code{print}.}
}
\value{
None
}
\description{
Prints some information about the \code{object}.
}
\seealso{
\code{\link{print.PCDs}}, \code{\link{summary.PCDs}}, and \code{\link{plot.PCDs}}
}
