% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFunctions.R
\name{print.summary.Planes}
\alias{print.summary.Planes}
\title{Print a summary of a \code{Planes} \code{object}}
\usage{
\method{print}{summary.Planes}(x, ...)
}
\arguments{
\item{x}{An \code{object} of class \code{"summary.Planes"},
generated by \code{summary.Planes}.}

\item{\dots}{Additional parameters for \code{print}.}
}
\value{
None
}
\description{
Prints some information about the \code{object}.
}
\seealso{
\code{\link{print.Planes}},
\code{\link{summary.Planes}},
and \code{\link{plot.Planes}}
}
