% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFunctions.R
\name{print.summary.TriLines}
\alias{print.summary.TriLines}
\title{Print a summary of a \code{TriLines} \code{object}}
\usage{
\method{print}{summary.TriLines}(x, ...)
}
\arguments{
\item{x}{An \code{object} of class \code{"summary.TriLines"},
generated by \code{summary.TriLines}.}

\item{\dots}{Additional parameters for \code{print}.}
}
\value{
None
}
\description{
Prints some information about the \code{object}
}
\seealso{
\code{\link{print.TriLines}},
\code{\link{summary.TriLines}},
and \code{\link{plot.TriLines}}
}
