% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFunctions.R
\name{print.summary.Uniform}
\alias{print.summary.Uniform}
\title{Print a summary of a \code{Uniform} \code{object}}
\usage{
\method{print}{summary.Uniform}(x, ...)
}
\arguments{
\item{x}{An \code{object} of class \code{"summary.Uniform"}, generated by \code{summary.Uniform}.}

\item{\dots}{Additional parameters for \code{print}.}
}
\value{
None
}
\description{
Prints some information about the \code{object}.
}
\seealso{
\code{\link{print.Uniform}}, \code{\link{summary.Uniform}}, and \code{\link{plot.Uniform}}
}
