% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{radii}
\alias{radii}
\title{The radii of points from one class with respect to points from the other class}
\usage{
radii(x, y)
}
\arguments{
\item{X}{A set of d-dimensional points for which the radii are computed. Radius of an X point equals to the
distance to the closest Y point}

\item{Y}{A set of d-dimensional points representing the reference points for the balls. That is, radius
of an X point is defined as the minimum distance to the Y points.}
}
\value{
A list with three elements
\item{rad}{A vector whose entries are the radius values for the X points. Radius of an X point equals to
the distance to the closest Y point}
\item{index.of.clYp}{A vector of indices of the closest Y points to the X points. The i-th entry in this
vector is the index of the closest Y point to i-th X point.}
\item{closest.Yp}{A vector of the closest Y points to the X points. The i-th entry in this
vector or i-th row in the matrix is the closest Y point to i-th X point.}
}
\description{
Returns the radii of the balls centered at X points where radius of an X point equals to the minimum distance
to Y points (i.e. distance to the closest Y point). That is, for each X point radius=\min_{y in Y}(d(x,y)).
X and Y points must be of the same dimension.
}
\examples{
A<-c(0,0); B<-c(1,0); C<-c(1/2,sqrt(3)/2);
Te<-rbind(A,B,C) #+10^(-16)
is.StdEqTri(Te)

is.StdEqTri(Te)

Te<-rbind(B,C,A)
is.StdEqTri(Te)

Tr<-rbind(A,B,-C)
is.StdEqTri(Tr)

A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
Tr<-rbind(A,B,C);
is.StdEqTri(Tr)

A<-c(0,0); B<-c(1,0); C<-c(.5,0.8660254);
Te<-rbind(A,B,C)
is.StdEqTri(Te)

is.StdEqTri(rbind(A,A,B))

dat.fr<-data.frame(a=Te)
is.StdEqTri(dat.fr)

}
\seealso{
\code{\link{radius}}
}
