% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{radius}
\alias{radius}
\title{The radius of a point from one class with respect to points from the other class}
\usage{
radius(pt, Y)
}
\arguments{
\item{pt}{A d-dimensional point for which radius is computed. Radius of pt equals to the
distance to the closest Y point to pt.}

\item{Y}{A set of d-dimensional points representing the reference points for the balls. That is, radius
of the point pt is defined as the minimum distance to the Y points.}
}
\value{
A list with three elements
\item{rad}{Radius value for the point, pt defined as \min_{y in Y}(d(pt,y))}
\item{index.of.clYpnt}{Index of the closest Y points to the point pt}
\item{closest.Ypnt}{The closest Y point to the point pt}
}
\description{
Returns the radius for the ball centered at point pt with radius=min distance to Y points.
That is, for the point pt radius=\min_{y in Y}(d(pt,y)) (i.e. distance form pt to the closest Y point).
The point pt and Y points must be of same dimension.
}
\examples{
nx<-10
ny<-5
X<-cbind(runif(nx),runif(nx))
Y<-cbind(runif(ny),runif(ny))
Rad<-radii(X,Y)
Rad
rd<-Rad$rad

Xlim<-range(X[,1]-rd,X[,1]+rd,Y[,1])
Ylim<-range(X[,2]-rd,X[,2]+rd,Y[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(rbind(Y),asp=1,pch=16,col=2,xlab="",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
points(rbind(X))
circles(X[,1],X[,2],Rad$rad,lty=1,lwd=1,col=4)

###
nx<-5
ny<-1
X<-cbind(runif(nx),runif(nx))
Y<-matrix(c(runif(ny),runif(ny)),ncol=2)
Rad<-radii(X,Y)
Rad
radii(Y,X)

rd<-Rad$rad

Xlim<-range(X[,1]-rd,X[,1]+rd,Y[,1])
Ylim<-range(X[,2]-rd,X[,2]+rd,Y[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(rbind(Y),asp=1,pch=16,col=2,xlab="",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
points(rbind(X))
circles(X[,1],X[,2],Rad$rad,lty=1,lwd=1,col=4)

radii(c(1,2),c(2,3))

###
nx<-10
ny<-5
X<-runif(nx)
Y<-runif(ny)
radii(X,Y) #this does not work, as X and Y are treated as vectors (i.e. points)

X<-as.matrix(X)
Y<-as.matrix(Y)
radii(X,Y) #this works as X and Y are treated as 1D data sets

###
nx<-10
ny<-5
X<-cbind(runif(nx),runif(nx),runif(nx))
Y<-cbind(runif(ny),runif(ny),runif(ny))
radii(X,Y)

dat.fr<-data.frame(a=X)
radii(dat.fr,Y)
radii(Y,dat.fr)

}
\seealso{
\code{\link{radii}}
}
