% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rasc.disc}
\alias{rasc.disc}
\title{An object of class "Patterns"

Generation of points associated (in a radial or circular fashion) with a given set of points}
\usage{
rasc.disc(n, Y, e, a1 = min(Y[, 1]), a2 = max(Y[, 1]), b1 = min(Y[,
  2]), b2 = max(Y[, 2]))
}
\arguments{
\item{n}{A positive integer representing the number of points to be generated}

\item{Y}{A set of 2D points representing the reference points. The generated points are associated
(in a circular/radial fashion) with these points.}

\item{e}{A positive real number representing the radius of the balls centered at Y points. Only these balls
are allowed for the generated points (i.e., generated points would be in the union of these balls)}

\item{a1, a2}{Real numbers representing the range of x-coordinates in the region
(default is the range of x-coordinates of the Y points)}

\item{b1, }{b2 Real numbers representing the range of y-coordinates in the region
(default is the range of y-coordinates of the Y points)}
}
\value{
A list with the elements
\item{type}{The type of the point pattern}
\item{mtxt}{The "main" title for the plot of the point pattern}
\item{parameters}{Radial attraction parameter of the association pattern}
\item{ref.points}{The input set of attraction points \code{Y},
i.e., points with which generated points are associated.}
\item{gen.points}{The output set of generated points associated with \code{Y} points}
\item{tri.Y}{Logical output for triangulation based on Y points should be implemented or not.
if TRUE triangulation based on Y points is to be implemented (default is set to FALSE).}
\item{desc.pat}{Description of the point pattern}
\item{num.points}{The vector of two numbers, which are the number of generated points and the number of attraction (i.e., Y) points.}
\item{xlimit}{The possible range of the x-coordinates of the generated points.}
\item{ylimit}{The possible range of the y-coordinates of the generated points.}
}
\description{
Generates n 2D points uniformly in (a1-e,a1+e)x(a1-e,a1+e) \cap \cup_i B(y_i,e) where Y=(y_1,y_2,...,y_{ny}) with
ny being number of Y points for various values of e under the association pattern
and B(y_i,e) is the ball centered at y_i with radius e.
e must be positive and very large values of e provide patterns close to CSR.
a1 is defaulted to the minimum of the x-coordinates of the Y points,
a2 is defaulted to the maximum of the x-coordinates of the Y points,
b1 is defaulted to the minimum of the y-coordinates of the Y points,
b2 is defaulted to the maximum of the y-coordinates of the Y points.
This function is also very similar to rasc.matern, where rasc.disc needs the study window to be specified,
while rasc.matern does not.
}
\examples{
nx<-1000;
ny<-10; #try 1 also
e<-.15;
#with default bounding box (i.e., unit square)
Y<-cbind(runif(ny),runif(ny))

Xdt<-rseg.disc(nx,Y,e)
Xdt
summary(Xdt)
plot(Xdt,asp=1)

e<-.15;
#e<--.1; #this provides a CSR realization
#e<-.15;

#with default bounding box (i.e., unit square)
Y<-cbind(runif(ny),runif(ny))
Xdt<-rseg.disc(nx,Y,e)$gen.points
Xlim<-range(Xdt[,1],Y[,1]);
Ylim<-range(Xdt[,2],Y[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Y,asp=1,pch=16,col=2,lwd=2, xlab="x",ylab="y",main="segregation of two classes",
     xlim=Xlim+xd*c(-.01,.01),ylim=Ylim+yd*c(-.01,.01))
points(Xdt)

#with a rectangular bounding box
a1<-0; a2<-10;
b1<-0; b2<-5;
e<-1.5;
#e<-1.5;
Y<-cbind(runif(ny,a1,a2),runif(ny,b1,b2))
#Y<-cbind(runif(ny,a1,a2/2),runif(ny,b1,b2/2))

Xdt<-rseg.disc(nx,Y,e,a1,a2,b1,b2)$gen.points
Xlim<-range(Xdt[,1],Y[,1]); Ylim<-range(Xdt[,2],Y[,2])

plot(Y,pch=16,asp=1,col=2,lwd=2, xlab="x",ylab="y",main="segregation of two classes",
     xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
points(Xdt)

}
\seealso{
\code{\link{rseg.disc}}, \code{\link{rascTe}}, \code{\link{rascIITe}},
\code{\link{rasc.matern}}, and \code{\link{rascMT}}
}
