% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rascIITe}
\alias{rascIITe}
\title{An object of class "Patterns"

Generation of points associated (in a Type II fashion) with the edges of Te}
\usage{
rascIITe(k, eps)
}
\arguments{
\item{k}{A positive integer representing the number of points to be generated}

\item{eps}{A positive real number representing the parameter of type II association (where sqrt(3)/6-eps
is the distance from the interior triangle distance from the interior triangle to Te.)}
}
\value{
A list with the elements
\item{type}{The type of the point pattern}
\item{mtxt}{The "main" title for the plot of the point pattern}
\item{parameters}{The attraction parameter, eps, of the association pattern, where sqrt(3)/6-eps
is the distance from the interior triangle to Te}
\item{ref.points}{The input set of points \code{Y};
reference points, i.e., points with which generated points are associated (i.e., vertices of Te).}
\item{gen.points}{The output set of generated points associated with \code{Y} points (i.e., edges of Te).}
\item{tri.Y}{Logical output for triangulation based on Y points should be implemented or not.
if TRUE triangulation based on Y points is to be implemented (default is set to FALSE).}
\item{desc.pat}{Description of the point pattern}
\item{num.points}{The vector of two numbers, which are the number of generated points and
the number of reference (i.e., Y) points, which is 3 here.}
\item{xlimit,ylimit}{The ranges of the x- and y-coordinates of the reference points, which are the
vertices of Te here}
}
\description{
Generates k points uniformly in the standard equilateral triangle Te=T((0,0),(1,0),(1/2,sqrt(3)/2))
under the type II association alternative for eps in (0,sqrt(3)/6=0.2886751]
In the type II association, the annular allowed regions around the edges are determined by
the parameter eps where sqrt(3)/6-eps is the distance from the interior triangle
(i.e., forbidden region for association) to Te (see examples for a sample plot.)
}
\examples{
A<-c(0,0); B<-c(1,0); C<-c(.5,sqrt(3.)/2);
Te<-rbind(A,B,C);
n<-1000
eps<-.25 #try also .15, .5, .75

#set.seed(1)
Xdt<-rascTe(n,eps)
Xdt
summary(Xdt)
plot(Xdt,asp=1)

Xlim<-range(Te[,1])
Ylim<-range(Te[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

dat<-rascTe(n,eps)$gen.points
plot(Te,pch=".",xlab="",ylab="",main="Type I association in the \\n standard equilateral triangle",
     xlim=Xlim+xd*c(-.01,.01),ylim=Ylim+yd*c(-.01,.01))
polygon(Te)
points(dat)

# The support for the Type I association alternative
sr<-(sqrt(3)/3-eps)/(sqrt(3)/2)
C1<-C+sr*(A-C); C2<-C+sr*(B-C)
A1<-A+sr*(B-A); A2<-A+sr*(C-A)
B1<-B+sr*(A-B); B2<-B+sr*(C-B)
supp<-rbind(A1,B1,B2,C2,C1,A2)

plot(Te,asp=1,pch=".",xlab="",ylab="",main="Support of the Type I Association",xlim=Xlim+xd*c(-.01,.01),ylim=Ylim+yd*c(-.01,.01))
if (sr<=.5)
{
  polygon(Te,col=5)
  polygon(supp,col=0)
} else
{
  polygon(Te,col=0,lwd=2.5)
  polygon(rbind(A,A1,A2),col=5,border=NA)
  polygon(rbind(B,B1,B2),col=5,border=NA)
  polygon(rbind(C,C1,C2),col=5,border=NA)
}
#points(dat)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{rseg.disc}}, \code{\link{rasc.disc}},  \code{\link{rsegIITe}},
and \code{\link{rsegMT}}
}
