% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rascMT}
\alias{rascMT}
\title{An object of class "Patterns"

Generation of points associated (in a Type I fashion) with a given set of points}
\usage{
rascMT(n, Yp, delta, DTmesh = NULL, DTr = NULL)
}
\arguments{
\item{n}{A positive integer representing the number of points to be generated}

\item{Yp}{A set of 2D points from which Delaunay triangulation is constructed}

\item{delta}{A positive real number in (0,4/9). delta is the parameter of association (that is, only
delta*100 %  area around each vertex in each Delaunay triangle is allowed for point generation).}

\item{DTmesh}{Delaunay triangulation of Yp, default is NULL, which is computed via tri.mesh function
in tripack package. tri.mesh function yields the triangulation nodes with their neigbours, and
creates a triangulation object}

\item{DTr}{Delaunay triangles based on Yp, default is NULL, which is computed via tri.mesh function
in tripack package. triangles function yields a triangulation data structure from the triangulation object created
by tri.mesh.}
}
\description{
Generates n points uniformly in the support for Type I segregation in the convex hull of set of points, Yp.
delta is the parameter of association (that is, only delta*100 % area around each vertex in each Delaunay
triangle is allowed for point generation). delta corresponds to eps in the standard equilateral triangle
Te as delta=4eps^2/3 (see rsegTe function). DTmesh must be the Delaunay triangulation of Yp
and DTr must be the corresponding Delaunay triangles (both DTmesh and DTr are NULL by default).
If NULL, DTmesh is computed via tri.mesh and DTr is computed via triangles function in tripack package.
tri.mesh function yields the triangulation nodes with their neigbours, and creates a triangulation object,
and triangles function yields a triangulation data structure from the triangulation object created
by tri.mesh (the first three colums are the vertex indices of the Delaunay triangles).
See (\insertCite{ceyhan:arc-density-PE,ceyhan:arc-density-CS,ceyhan:dom-num-NPE-Spat2011;textual}{pcds}) for more on the
association pattern.
Also see (\insertCite{okabe:2000,ceyhan:comp-geo-2010,renka:1996;textual}{pcds}) for more on Delaunay triangulation and the tripack algorithm.
}
\examples{
nx<-1000 #number of X points (target) #try nx<-1, 10, 100 also
ny<-10 #number of Y points (nontarget)

#set.seed(1)
Yp<-cbind(runif(ny),runif(ny))
del<-.4

Xdt<-rsegMT(nx,Yp,del)
Xdt
summary(Xdt)
plot(Xdt)

Yp<-cbind(runif(ny),runif(ny))
del<-.3 #try .5, .75, .85
dat<-rsegMT(nx,Yp,del) #data under CSR in the convex hull of Ypoints
#or use
DTY<-tri.mesh(Yp[,1],Yp[,2],duplicate="remove") #Delaunay triangulation based on Y points
dat<-rsegMT(nx,Yp,del,DTY) #data under CSR in the convex hull of Ypoints

#or use
DTY<-tri.mesh(Yp[,1],Yp[,2],duplicate="remove") #Delaunay triangulation based on Y points
TRY<-triangles(DTY)[,1:3];
dat<-rsegMT(nx,Yp,del,DTr=TRY) #data under CSR in the convex hull of Ypoints

#or use
DTY<-tri.mesh(Yp[,1],Yp[,2],duplicate="remove") #Delaunay triangulation based on Y points
TRY<-triangles(DTY)[,1:3];
dat<-rsegMT(nx,Yp,del,DTY,TRY)$gen.points #data under CSR in the convex hull of Ypoints

Xlim<-range(Yp[,1])
Ylim<-range(Yp[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

#plot of the data in the convex hull of Y points together with the Delaunay triangulation
DTY<-tri.mesh(Yp[,1],Yp[,2],duplicate="remove") #Delaunay triangulation based on Y points

#par(pty="s")
plot(dat,main=" ", xlab=" ", ylab=" ",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05),type="n")
plot.tri(DTY, add=TRUE, do.points=TRUE,col="blue")
points(dat,pch=".",cex=3)

Yp<-rbind(c(.3,.2),c(.4,.5),c(.14,.15))
rsegMT(nx,Yp,del)

dat.fr<-data.frame(a=Yp)
rsegMT(nx,dat.fr,del)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{rasc.disc}}, \code{\link{rascTe}}, \code{\link{rascIITe}},
and \code{\link{rsegMT}}
}
