% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rascTe}
\alias{rascTe}
\title{An object of class "Patterns"

Generation of points associated (in a Type I fashion) with the vertices of Te}
\usage{
rascTe(k, eps)
}
\arguments{
\item{k}{A positive integer representing the number of points to be generated}

\item{eps}{A positive real number representing the parameter of type I association (where sqrt(3)/3-eps
serves as the height of the triangular support regions around the vertices)}
}
\value{
A list with the elements
\item{type}{The type of the point pattern}
\item{mtxt}{The "main" title for the plot of the point pattern}
\item{parameters}{The attraction parameter of the association pattern, eps,
where sqrt(3)/3-eps serves as the height of the triangular support regions around the vertices}
\item{ref.points}{The input set of points \code{Y};
reference points, i.e., points with which generated points are associated (i.e., vertices of Te).}
\item{gen.points}{The output set of generated points associated with \code{Y} points (i.e., vertices of Te).}
\item{tri.Y}{Logical output for triangulation based on Y points should be implemented or not.
if TRUE triangulation based on Y points is to be implemented (default is set to FALSE).}
\item{desc.pat}{Description of the point pattern.}
\item{num.points}{The vector of two numbers,
which are the number of generated points and the number of reference (i.e., Y) points.}
\item{xlimit,ylimit}{The ranges of the x- and y-coordinates of the reference points, which are the
vertices of Te here}
}
\description{
Generates k points uniformly in the standard equilateral triangle Te=T((0,0),(1,0),(1/2,sqrt(3)/2))
under the type I association alternative for eps in (0,sqrt(3)/3=0.5773503].
The allowed triangular regions around the vertices are determined by the parameter eps.
In the type I association, the triangular support regions around the vertices are determined by
the parameter eps where sqrt(3)/3-eps serves as the height of these triangles
(see examples for a sample plot.)
See also (\insertCite{ceyhan:arc-density-PE,ceyhan:arc-density-CS,ceyhan:dom-num-NPE-Spat2011;textual}{pcds}).
}
\examples{
A<-c(0,0); B<-c(1,0); C<-c(.5,sqrt(3.)/2);
Te<-rbind(A,B,C);
n<-1000
eps<-.15 #try also .2

#set.seed(1)
Xdt<-rsegIITe(n,eps)
Xdt
summary(Xdt)
plot(Xdt,asp=1)

Xlim<-range(Te[,1])
Ylim<-range(Te[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

dat<-rsegIITe(n,eps)$gen.points

plot(Te,pch=".",xlab="",ylab="",main="Type II segregation in the \\n standard equilateral triangle",
     xlim=Xlim+xd*c(-.01,.01),ylim=Ylim+yd*c(-.01,.01))
polygon(Te)
points(dat)
C1<-c(1/2,sqrt(3)/2-2*eps);
supp<-rbind(A1,B1,C1)
# The support for the Type II segregation alternative
A1<-c(eps*sqrt(3),eps); B1<-c(1-eps*sqrt(3),eps);

plot(Te,asp=1,pch=".",xlab="",ylab="",main="Support of the Type II Segregation",
     xlim=Xlim+xd*c(-.01,.01),ylim=Ylim+yd*c(-.01,.01))
  polygon(Te)
  polygon(supp,col=5)
#points(dat)

}
\seealso{
\code{\link{rseg.disc}}, \code{\link{rasc.disc}},  \code{\link{rsegIITe}},
and \code{\link{rsegMT}}
}
