% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{re.bastri.cent}
\alias{re.bastri.cent}
\title{The index of the edge region in a basic triangle that contains a point}
\usage{
re.bastri.cent(pt, c1, c2, M)
}
\arguments{
\item{pt}{A 2D point for which M-edge region it resides in is to be determined in the
basic triangle Tb}

\item{c1, c2}{Positive real numbers which constitute the upper vertex of the basic triangle (i.e., the vertex
adjacent to the shorter edges of Tb); c1 must be in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the basic triangle Tb}
}
\value{
A list with three elements

\item{re}{Index of the edge whose region contains point, pt.}
\item{tri}{The vertices of the triangle, where row labels are A=(0,0), B=(1,0), and C=(c1,c2)
with edges are labeled as 3 for edge AB, 1 for edge BC, and 2 for edge AC.}
\item{txt}{Description of the edge labels.}

\item{re}{Index of the M-edge region that contains point, pt in the basic triangle Tb.}
\item{tri}{The vertices of the triangle, where row labels are A,B, and C
with edges are labeled as 3 for edge AB, 1 for edge BC, and 2 for edge AC.}
\item{txt}{Description of the edge labels}
}
\description{
Returns the index of the edge whose region contains point, pt, in
the basic triangle Tb=T(A=(0,0),B=(1,0),C=(c1,c2)) and edge regions based on center
M=(m1,m2) in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates
in the interior of the basic triangle Tb.
Edges are labeled as 3 for edge AB, 1 for edge BC, and 2 for edge AC.
If the point, pt, is not inside tri, then the function yields NA as output.
Edge region 1 is the triangle T(B,C,M), edge region 2 is T(A,C,M), and edge region 3 is T(A,B,M).
In the basic triangle Tb=T(A=(0,0),B=(1,0),C=(c1,c2)) c1 is in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1.
Any given triangle can be mapped to the basic triangle by a combination of rigid body motions
(i.e., translation, rotation and reflection) and scaling, preserving uniformity of the points in the
original triangle. Hence basic triangle is useful for simulation
studies under the uniformness hypothesis.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012,ceyhan:arc-density-CS;textual}{pcds}).
}
\examples{
c1<-.4; c2<-.6
P<-c(.4,.2)
re.bastriCM(P,c1,c2)

c1<-.5; c2<-.8
P<-c(.4,.2)
re.bastriCM(P,c1,c2)

P<-c(.8,.2)
re.bastriCM(P,c1,c2)

P<-c(1.8,.2)
re.bastriCM(P,c1,c2)

c1<-.4; c2<-.6
A<-c(0,0);B<-c(1,0);C<-c(c1,c2);
Tb<-rbind(A,B,C)
CM<-(A+B+C)/3

re.bastriCM(A,c1,c2)
re.bastriCM(B,c1,c2)
re.bastriCM(C,c1,c2)
re.bastriCM(CM,c1,c2)

n<-20
dat<-runif.bastri(n,c1,c2)$g
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

re<-vector()
for (i in 1:n)
  re<-c(re,re.bastriCM(dat[i,],c1,c2)$re)
re

Xlim<-range(Tb[,1],dat[,1])
Ylim<-range(Tb[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Tb,xlab="",ylab="",axes=TRUE,pch=".",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
points(dat,pch=".")
polygon(Tb)
L<-Tb; R<-matrix(rep(CM,3),ncol=3,byrow=T)
segments(L[,1], L[,2], R[,1], R[,2], lty=2)
text(dat,labels=factor(re))

txt<-rbind(Tb,CM)
xc<-txt[,1]+c(-.03,.03,.02,0)
yc<-txt[,2]+c(.02,.02,.02,-.04)
txt.str<-c("A","B","C","CM")
text(xc,yc,txt.str)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{re.triCM}}, \code{\link{re.tri.cent}},
\code{\link{re.bastri.cent}}, \code{\link{reTeCM}}, and \code{\link{redge.triCM}}
}
