% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{re.triCM}
\alias{re.triCM}
\title{The index of the CM-edge region in a triangle that contains the point}
\usage{
re.triCM(pt, tri)
}
\arguments{
\item{pt}{A 2D point for which CM-edge region it resides in is to be determined in the triangle
tri}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}
}
\value{
A list with three elements
\item{re}{Index of the CM-edge region that contains point, pt in the triangle tri.}
\item{tri}{The vertices of the triangle, where row labels are A,B, and C
with edges are labeled as 3 for edge AB, 1 for edge BC, and 2 for edge AC.}
\item{txt}{Description of the edge labels}
}
\description{
Returns the index of the edge whose region contains point, pt, in
the triangle tri=T(A,B,C) with edge regions based on center of mass CM=(A+B+C)/3.
Edges are labeled as 3 for edge AB, 1 for edge BC, and 2 for edge AC.
If the point, pt, is not inside tri, then the function yields NA as output.
Edge region 1 is the triangle T(B,C,CM), edge region 2 is T(A,C,CM), and edge region 3 is T(A,B,CM).
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012,ceyhan:arc-density-CS;textual}{pcds}).
}
\examples{
A<-c(0,0); B<-c(1,0); C<-c(0.5,.8);
#A<-runif(2); B<-runif(2); C<-runif(2)
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
area.polygon(Tr)

A<-c(0,0); B<-c(1,0); C<-c(.7,.6); D<-c(0.3,.8);
h1<-rbind(A,B,C,D); #h1<-rbind(A,B,D,C); #h1<-rbind(A,C,B,D)
#h1<-rbind(A,D,C,B);
area.polygon(h1)

Xlim<-range(h1[,1])
Ylim<-range(h1[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(h1,xlab="",ylab="",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(h1)
xc<-rbind(A,B,C,D)[,1]+c(-.03,.03,.02,-.01)
yc<-rbind(A,B,C,D)[,2]+c(.02,.02,.02,.04)
txt.str<-c("A","B","C","D")
text(xc,yc,txt.str)

#when the triangle is degenerate
B<-A+2*(C-A);
T2<-rbind(A,B,C)
area.polygon(T2)

dat.fr<-data.frame(a=Tr)
area.polygon(dat.fr)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{re.tri.cent}}, \code{\link{re.triCM}}, \code{\link{re.bastriCM}},
\code{\link{re.bastri.cent}}, \code{\link{reTeCM}}, and \code{\link{redge.triCM}}
}
