% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{redges.triCM}
\alias{redges.triCM}
\title{The indices of the CM-edge regions in a triangle that contains the points in a give data set}
\usage{
redges.triCM(Dt, tri)
}
\arguments{
\item{Dt}{A set of 2D points representing the set of data points for which indices of the edge regions
containing them are to be determined}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}
}
\value{
A list with the elements
\item{rv}{Indices (i.e., a vector of indices) of the edges whose region contains points in Dt
in the triangle tri}
\item{tri}{The vertices of the triangle, where row number corresponds to the vertex index in rv.}
\item{txt}{Description of the edge labels as "Edge labels are AB=1, BC=2, and AC=3".}
}
\description{
Returns the indices of the edges whose regions contain the points in data set Dt in
a triangle tri=(A,B,C) and edge regions are based on the center of mass CM of tri.
(see the plots in the example for illustrations). The vertices of the  triangle tri are labeled as
1=A,2=B, and 3=C also according to the row number the vertex is recorded in tri
and the corresponding edges are 1=BC, 2=AC, and 3=AB.
If a point in Dt is not inside tri, then the function yields NA as output for that entry.
The corresponding edge region is the polygon
with the vertex, CM, and vertices other than the non-adjacent vertex, i.e., edge region 1 is the triangle
T(B,CM,C), edge region 2 is T(A,CM,C) and edge region 3 is T(A,B,CM).
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012,ceyhan:arc-density-CS;textual}{pcds}).
}
\seealso{
\code{\link{redges.tri.cent}}, \code{\link{rverts.tri.cent}} and \code{\link{rverts.tri.nd}}
}
