% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rel.six.Te}
\alias{rel.six.Te}
\title{Region index inside the Gamma-1 region.}
\usage{
rel.six.Te(Pt)
}
\arguments{
\item{Pt}{A 2D point.}
}
\value{
rel an integer between 1-6 (inclusive) or NA
}
\description{
Returns the region index of the point Pt for the 6 regions in sandard equilateral triangle
Te=T((0,0),(1,0),(1/2,sqrt(3)/2)), starting with 1 on the first one-sixth of the triangle,
and numbering follows the counter-clockwise direction (see the plot in the examples).
These regions are in the inner hexagon which is the Gamma-1 region for CS-PCD with t=1
if Pt is not in any of the 6 regions the function returns NA.
}
\examples{
n<-20
dat<-runifTe(n)$gen.points
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

Ext<-cl2edgesTe(dat)
Ext
summary(Ext)
plot(Ext,asp=1)

cl2edgesTe(dat[1,])
cl2edgesTe(c(10,10))

ed.clo<-cl2edgesTe(dat)

dat2<-rbind(dat,c(.8,.8))
cl2edgesTe(dat2)

A<-c(0,0); B<-c(1,0); C<-c(0.5,sqrt(3)/2);
Te<-rbind(A,B,C)
CM<-(A+B+C)/3
p1<-(A+B)/2
p2<-(B+C)/2
p3<-(A+C)/2

Xlim<-range(Te[,1],dat[,1])
Ylim<-range(Te[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(A,pch=".",xlab="",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Te)
points(dat,xlab="",ylab="")
points(ed.clo$Ext,pty=2,pch=4,col="red")

txt<-rbind(Te,p1,p2,p3)
xc<-txt[,1]+c(-.03,.03,.03,0,0,0)
yc<-txt[,2]+c(.02,.02,.02,0,0,0)
txt.str<-c("A","B","C","re=1","re=2","re=3")
text(xc,yc,txt.str)

dat.fr<-data.frame(a=dat)
cl2edgesTe(dat.fr)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{runif.1_6thTe}}
}
