% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{reTeCM}
\alias{reTeCM}
\title{The index of the edge region in the standard equilateral triangle that contains a point}
\usage{
reTeCM(pt)
}
\arguments{
\item{pt}{A 2D point for which CM-edge region it resides in is to be determined in the
the standard equilateral triangle Te}
}
\value{
A list with three elements
\item{re}{Index of the CM-edge region that contains point, pt in the standard equilateral triangle Te}
\item{tri}{The vertices of the standard equilateral triangle Te, where row labels are A, B, and C
with edges are labeled as 3 for edge AB, 1 for edge BC, and 2 for edge AC.}
\item{txt}{Description of the edge labels}
}
\description{
Returns the index of the edge whose region contains point, Pt, in
the standard equilateral triangle Te=T(A=(0,0),B=(1,0),C=(1/2,sqrt(3)/2)) with edge regions based on
center of mass CM=(A+B+C)/3.
Edges are labeled as 3 for edge AB, 1 for edge BC, and 2 for edge AC.
If the point, pt, is not inside tri, then the function yields NA as output.
Edge region 1 is the triangle T(B,C,M), edge region 2 is T(A,C,M), and edge region 3 is T(A,B,M).
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012,ceyhan:arc-density-CS;textual}{pcds}).
}
\examples{
c1<-.4; c2<-.6
A<-c(0,0); B<-c(1,0); C<-c(c1,c2);
Tb<-rbind(A,B,C);
M<-c(.6,.2)

P<-c(.4,.2)
re.bastri.cent(P,c1,c2,M)

P<-c(1.4,.2)
re.bastri.cent(P,c1,c2,M)

c1<-.5; c2<-.8
P<-c(.4,.2)
re.bastri.cent(P,c1,c2,M)

P<-c(.8,.2)
re.bastri.cent(P,c1,c2,M)

c1<-.4; c2<-.6
A<-c(0,0);B<-c(1,0);C<-c(c1,c2);
Tb<-rbind(A,B,C)

re.bastri.cent(A,c1,c2,M)
re.bastri.cent(B,c1,c2,M)
re.bastri.cent(C,c1,c2,M)
re.bastri.cent(M,c1,c2,M)

n<-20
dat<-runif.bastri(n,c1,c2)$g
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

M<-as.numeric(runif.bastri(1,c1,c2)$g)
#M<-c(.6,.2)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

re<-vector()
for (i in 1:n)
  re<-c(re,re.bastri.cent(dat[i,],c1,c2,M)$re)
re

Xlim<-range(Tb[,1],dat[,1])
Ylim<-range(Tb[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Tb,xlab="",ylab="",axes=TRUE,pch=".",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
points(dat,pch=".")
polygon(Tb)
L<-Tb; R<-rbind(M,M,M)
segments(L[,1], L[,2], R[,1], R[,2], lty=2)
text(dat,labels=factor(re))

txt<-rbind(Tb,M)
xc<-txt[,1]+c(-.03,.03,.02,0)
yc<-txt[,2]+c(.02,.02,.02,-.03)
txt.str<-c("A","B","C","M")
text(xc,yc,txt.str)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{re.triCM}}, \code{\link{re.tri.cent}},
\code{\link{re.bastriCM}}, \code{\link{re.bastri.cent}}, and \code{\link{redge.triCM}}
}
