% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rsegTe}
\alias{rsegTe}
\title{An object of class "Patterns"

Generation of points segregated (in a Type I fashion) from the vertices of Te}
\usage{
rsegTe(k, eps)
}
\arguments{
\item{k}{A positive integer representing the number of points to be generated}

\item{eps}{A positive real number representing the parameter of type I segregation (which is the
height of the triangular forbidden regions around the vertices)}
}
\value{
A list with the elements
\item{type}{The type of the point pattern}
\item{mtxt}{The "main" title for the plot of the point pattern}
\item{parameters}{The exclusion parameter, eps, of the segregation pattern, which is the height
of the triangular forbidden regions around the vertices }
\item{ref.points}{The input set of points \code{Y};
reference points, i.e., points from which generated points are segregated (i.e., vertices of Te).}
\item{gen.points}{The output set of generated points segregated from \code{Y} points (i.e., vertices of Te).}
\item{tri.Y}{Logical output for triangulation based on Y points should be implemented or not.
if TRUE triangulation based on Y points is to be implemented (default is set to FALSE).}
\item{desc.pat}{Description of the point pattern}
\item{num.points}{The vector of two numbers, which are the number of generated points and
the number of reference (i.e., Y) points, which is 3 here.}
\item{xlimit,ylimit}{The ranges of the x- and y-coordinates of the reference points, which are the
vertices of Te here}
}
\description{
Generates k points uniformly in the standard equilateral triangle Te=T((0,0),(1,0),(1/2,sqrt(3)/2))
under the type I segregation alternative for eps in (0,sqrt(3)/3=0.5773503].
In the type I segregation, the triangular forbidden regions around the vertices are determined by
the parameter eps which serves as the height of these triangles (see examples for a sample plot.)
See also (\insertCite{ceyhan:arc-density-PE,ceyhan:arc-density-CS,ceyhan:dom-num-NPE-Spat2011;textual}{pcds}).
}
\examples{
n<-100
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C)

Xdt<-runif.tri(n,Tr)
Xdt
summary(Xdt)
plot(Xdt)

dat<-runif.tri(n,Tr)$g
Xlim<-range(Tr[,1])
Ylim<-range(Tr[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Tr,pch=".",xlab="",ylab="",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Tr)
points(dat)
xc<-Tr[,1]#+c(-.03,.03,0.03)
yc<-Tr[,2]#+c(.02,.02,.02)
txt.str<-c("A","B","C")
text(xc,yc,txt.str)

}
\seealso{
\code{\link{rseg.disc}}, \code{\link{rasc.disc}},  \code{\link{rsegIITe}},
and \code{\link{rsegMT}}
}
