% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{runif.1_6thTe}
\alias{runif.1_6thTe}
\title{An object of class "Uniform"

Generation of Uniform Points in a part of standard equilateral triangle}
\usage{
runif.1_6thTe(k)
}
\arguments{
\item{k}{a positive integer}
}
\value{
A list with the elements
\item{type}{The type of the point pattern}
\item{mtxt}{The "main" title for the plot of the point pattern}
\item{support}{The vertices of the support of the uniformly generated points}
\item{gen.points}{The output set of uniformly generated points in the first 1/6th of
the standard equilateral triangle.}
\item{out.region}{The outer region for the one-sixth of Te, which is just Te here.}
\item{desc.pat}{Description of the point pattern}
\item{num.points}{The vector of two numbers, which are the number of generated points
and the number of vertices of the support (i.e., Y) points.}
\item{txt4pnts}{Description of the two numbers in num.points.}
\item{xlimit}{The range of the x-coordinates of the generated, support and outer region points}
\item{ylimit}{The range of the y-coordinates of the generated, support and outer region points}
}
\description{
Generates k points uniformly in the first 1/6th of the standard equilateral triangle Te=(A,B,C)
with vertices with A=(0,0);  B=(1,0); C=(.5,sqrt(3)/2) (see the examples below).
The first 1/6th of the standard equilateral triangle is the triangle with vertices
A=(0,0);  (1/2,0); C=(1/2,sqrt(3)/6).
}
\examples{
nx<-1000;
ny<-10;

e<-.15;
#e<-1.1; #closer to CSR than association, as e is large
#e<--.15; #not possible, gives error
#Y points uniform in unit square
Y<-cbind(runif(ny),runif(ny))

Xdt<-rasc.matern(nx,Y,e)
Xdt
summary(Xdt)
plot(Xdt,asp=1)

Xdt<-rasc.matern(nx,Y,e)$gen.points
Xlim<-range(Xdt[,1],Y[,1]);
Ylim<-range(Xdt[,2],Y[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Y,asp=1,xlab="x",ylab="y",main="Association of two classes",
     xlim=Xlim+xd*c(-.01,.01),ylim=Ylim+yd*c(-.01,.01),pch=16,col=2,lwd=2)
points(Xdt)

#Y points uniform in a rectangle
a1<-0; a2<-10;
b1<-0; b2<-5;
e<-1.1;

Y<-cbind(runif(ny,a1,a2),runif(ny,b1,b2))
#Y<-cbind(runif(ny,a1,a2/2),runif(ny,b1,b2/2))

Xdt<-rasc.matern(nx,Y,e)$gen.points
Xlim<-range(Xdt[,1],Y[,1]);
Ylim<-range(Xdt[,2],Y[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Y,asp=1,xlab="x",ylab="y",main="association of two classes",
     xlim=Xlim+xd*c(-.01,.01),ylim=Ylim+yd*c(-.01,.01),pch=16,col=2,lwd=2)
points(Xdt)

e<-.15
Y<-data.frame(yp=cbind(runif(ny),runif(ny)) )
Xdt<-rasc.matern(nx,Y,e)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{runifTe}}, \code{\link{runif.bastri}}, \code{\link{runif.tri}},
and \code{\link{runifMT}}
}
