% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{runifTe}
\alias{runifTe}
\title{An object of class "Uniform"

Generation of Uniform Points in the Standard Equilateral Triangle Te}
\usage{
runifTe(k)
}
\arguments{
\item{k}{A positive integer representing the number of uniform points to be generated in the
standard equilateral triangle Te}
}
\value{
A list with the elements
\item{type}{The type of the pattern from which points are to be generated}
\item{mtxt}{The "main" title for the plot of the point pattern}
\item{tess.points}{The vertices of the support region of the uniformly generated points, it is the
standard equilateral triangle Te for this function}
\item{gen.points}{The output set of generated points uniformly in the standard equilateral triangle Te.}
\item{out.region}{The outer region which contains the support region, NULL for this function.}
\item{desc.pat}{Description of the point pattern from which points are to be generated}
\item{num.points}{The vector of two numbers, which are the number of generated points and the number
of vertices of the support points (here it is 3).}
\item{txt4pnts}{Description of the two numbers in num.points}
\item{xlimit,ylimit}{The ranges of the x- and y-coordinates of the support, Te}
}
\description{
Generates k points uniformly in the standard equilateral triangle Te=(A,B,C)
with vertices A=(0,0), B=(1,0), and C=(.5,sqrt(3)/2).
}
\seealso{
\code{\link{runif.bastri}}, \code{\link{runif.tri}}, and \code{\link{runifMT}}
}
