% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rv.bastriCM}
\alias{rv.bastriCM}
\title{The index of the CM-vertex region in a basic triangle that contains a point}
\usage{
rv.bastriCM(p, c1, c2)
}
\arguments{
\item{p}{A 2D point for which CM-vertex region it resides in is to be determined in the
basic triangle Tb}

\item{c1, c2}{Positive real numbers which constitute the upper vertex of the basic triangle (i.e., the vertex
adjacent to the shorter edges of Tb); c1 must be in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1}
}
\value{
A list with two elements
\item{rv}{Index of the CM-vertex region that contains point, p in the basic triangle Tb}
\item{tri}{The vertices of the triangle, where row number corresponds to the vertex index in rv
with row 1=(0,0), row 2=(1,0), and row 3=(c1,c2).}
}
\description{
Returns the index of the vertex whose region contains point p in
the basic triangle Tb=T((0,0),(1,0),(c1,c2)) where c1 is in [0,1/2], c2>0 and (1-c1)^2+c2^2 <=1
and vertex regions are based on the center of mass CM=((1+c1)/3,c2/3) of Tb.
(see the plots in the example for illustrations). The vertices of the basic triangle Tb are labeled as
1=(0,0),2=(1,0),and 3=(c1,c2) also according to the row number the vertex is recorded in Tb.
If the point, p, is not inside Tb, then the function yields NA as output.
The corresponding vertex region is the polygon with the vertex, CM, and
midpoints of the edges adjacent to the vertex. Any given triangle can be mapped to the basic triangle
by a combination of rigid body motions (i.e., translation, rotation and reflection) and scaling,
preserving uniformity of the points in the original triangle. Hence basic triangle is useful for simulation
studies under the uniformness hypothesis.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012,ceyhan:arc-density-PE;textual}{pcds})
}
\examples{
c1<-.4; c2<-.6
A<-c(0,0); B<-c(1,0); C<-c(c1,c2);

#A<-runif(2); B<-runif(2); C<-runif(2)
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C)

bary2cart(c(.3,.2,.5),Tr)
bary2cart(c(.4,.2,.4),Tr)
bary2cart(c(.5,.2,.3),Tr)
bary2cart(c(6,2,4),Tr)

P<-c(.8,.2,.3)
bary2cart(P,Tr)

P<-c(-.5,.4,.2)
bary2cart(P,Tr)

CM<-(A+B+C)/3; CM
bary2cart(c(1,1,1),Tr)

####
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
Tr<-rbind(A,B,C);
bary2cart(c(1.4,1.2,1),Tr)

bary2cart(c(.8,.2,.6),Tr)

bary2cart(c(1,2,3),Tr)

dat.fr<-data.frame(a=Tr)
bary2cart(c(.8,.2,.3),dat.fr)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{rv.triCM}}, \code{\link{rv.tri.cent}}, \code{\link{rv.triCC}},
\code{\link{rv.bastriCC}}, \code{\link{rv.bastri.cent}}, and \code{\link{rvTeCM}}
}
