% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rv.end.int}
\alias{rv.end.int}
\title{The index of the vertex region in an end-interval that contains a given point}
\usage{
rv.end.int(pt, int)
}
\arguments{
\item{pt}{A 1D point whose end interval region is provided by the function}

\item{int}{A vector of two real numbers representing an interval}
}
\value{
A list with two elements
\item{rv}{Index of the end vertex whose region contains point, pt.}
\item{int}{The vertices of the interval as a vector where position of the vertex corresponds to
the vertex index as int=(rv=1,rv=2).}
}
\description{
Returns the index of the vertex in the interval, int, whose end interval contains the 1D point pt,
that is, it finds the index of the vertex for the point, pt, outside
the interval int=(a,b)=(vertex 1,vertex 2);
vertices of interval are labeled as 1 and 2 according to their order in the interval.
If the point, pt, is inside int, then the function yields NA as output.
The corresponding vertex region is an interval as (-Inf,a) or (b,Inf) for the interval (a,b).
Then if pt<a, then rv=1 and if pt>b, then rv=2. Unlike rv.mid.int, centrality parameter (i.e., center
of the interval is not relevant for rv.end.int.)
See also (\insertCite{ceyhan:metrika-2012,ceyhan:revstat-2016;textual}{pcds}).
}
\examples{
r<-2
c<-.4
a<-0; b<-10;
#c<-runif(1)
#a<-runif(1); b<-runif(1); int<-range(a,b); a<-int[1]; b<-int[2]
nx<-40; ny<-10

#set.seed(1)
Xp<-runif(nx,a,b)
#Xp<-runif(nx,a+5,b+5)
#Xp<-runif(nx,a+10,b+10)
Yp<-runif(ny,a,b)

Arcs<-ArcsPEmid1D(Xp,Yp,r,c)
Arcs
summary(Arcs)
plot(Arcs)

S<-Arcs$S
E<-Arcs$E

ArcsPEmid1D(Xp,Yp,r,c)
ArcsPEmid1D(Xp,Yp+10,r,c)

jit<-.1
yjit<-runif(nx,-jit,jit)

Xlim<-range(Xp,Yp)
xd<-Xlim[2]-Xlim[1]

plot(cbind(a,0),main="arcs of PE-PCD for points (jittered along y-axis)\\n in middle intervals ", xlab=" ", ylab=" ",
     xlim=Xlim+xd*c(-.05,.05),ylim=3*c(-jit,jit),pch=".")
abline(h=0,lty=1)
points(Xp, yjit,pch=".",cex=3)
abline(v=Yp,lty=2)
arrows(S, yjit, E, yjit, length = .05, col= 4)

r<-2
c<-.4
a<-0; b<-10;
nx<-40; ny<-10
Xp<-runif(nx,a,b)
Yp<-runif(ny,a,b)
ArcsPEmid1D(Xp,Yp,r,c)


}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{rv.mid.int}}
}
