% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rv.mid.int}
\alias{rv.mid.int}
\title{The index of the vertex region in a middle interval that contains a given point}
\usage{
rv.mid.int(pt, c, int)
}
\arguments{
\item{pt}{A 1D point. The vertex region pt resides is to be found.}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a)}

\item{int}{A vector of two real numbers representing an interval}
}
\value{
A list with two elements
\item{rv}{Index of the vertex in the interval int whose region contains point, pt.}
\item{int}{The vertices of the interval as a vector where position of the vertex corresponds to
the vertex index as int=(rv=1,rv=2).}
}
\description{
Returns the index of the vertex whose region contains point pt in
the interval int=(a,b)=(vertex 1,vertex 2) with (parameterized) center Mc associated with
the centrality parameter c in (0,1); vertices of interval are labeled as 1 and 2 according to their
order in the interval int. If the point, pt, is not inside int, then the function yields NA as output.
The corresponding vertex region is the interval (a,b) as (a,Mc) and (Mc,b) where Mc=a+c*(b-a).
See also (\insertCite{ceyhan:metrika-2012,ceyhan:revstat-2016;textual}{pcds}).
}
\examples{
n<-10
c<-.4
#c<-runif(1)
x<-runif(n)
centersMc(x,c)

n<-10
c<-.3
#c<-runif(1)
x<-runif(n,0,10)
centersMc(x,c)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{rv.end.int}}
}
