% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rv.tetraCC}
\alias{rv.tetraCC}
\title{The index of the CC-vertex region in a tetrahedron that contains a point}
\usage{
rv.tetraCC(p, th)
}
\arguments{
\item{p}{A 3D point for which CC-vertex region it resides in is to be determined in the
tetrahedron th}

\item{th}{Four 3D points, stacked row-wise, each row representing a vertex of the tetrahedron}
}
\value{
A list with two elements
\item{rv}{Index of the CC-vertex region that contains point, p in the tetrahedron th}
\item{tri}{The vertices of the tetrahedron, where row number corresponds to the vertex index in rv.}
}
\description{
Returns the index of the vertex whose region contains point p in
a tetrahedron th=T(A,B,C,D) and vertex regions are based on the circumcenter CC of th.
(see the plots in the example for illustrations). The vertices of the tetrahedron th are labeled as
1=A, 2=B, 3=C, and 4=C also according to the row number the vertex is recorded in th.
If the point, p, is not inside th, then the function yields NA as output.
The corresponding vertex region is the polygon whose interior points are closest to that vertex.
If th is regular tetrahedron, then CC and CM (center of mass) coincide.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010;textual}{pcds}).
}
\examples{
A<-c(0,0,0); B<-c(1,0,0); C<-c(1/2,sqrt(3)/2,0); D<-c(1/2,sqrt(3)/6,sqrt(6)/3)
#A<-runif(3); B<-runif(3); C<-runif(3); D<-runif(3)
#A<-runif(3,1,100); B<-runif(3,1,100); C<-runif(3,1,100); D<-runif(3,1,100)
tetra<-rbind(A,B,C,D)

CC<-circ.cent.tetra(tetra)
CC

Xlim<-range(tetra[,1],CC[1])
Ylim<-range(tetra[,2],CC[2])
Zlim<-range(tetra[,3],CC[3])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]
zd<-Zlim[2]-Zlim[1]

library(plot3D)
scatter3D(tetra[,1],tetra[,2],tetra[,3], phi =0,theta=40, bty = "g", xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05), zlim=Zlim+zd*c(-.05,.05),
          pch = 20, cex = 1, ticktype = "detailed")
#add the vertices of the tetrahedron
points3D(CC[1],CC[2],CC[3], add=TRUE)
L<-rbind(A,A,A,B,B,C); R<-rbind(B,C,D,C,D,D)
segments3D(L[,1], L[,2], L[,3], R[,1], R[,2],R[,3], add=TRUE,lwd=2)

text3D(tetra[,1],tetra[,2],tetra[,3], labels=c("A","B","C","D"), add=TRUE)

D1<-(A+B)/2; D2<-(A+C)/2; D3<-(A+D)/2; D4<-(B+C)/2; D5<-(B+D)/2; D6<-(C+D)/2;
L<-rbind(D1,D2,D3,D4,D5,D6); R<-
segments3D(L[,1], L[,2], L[,3], R[,1], R[,2],R[,3], add=TRUE,lty=2)

text3D(CC[1],CC[2],CC[3], labels="CC", add=TRUE)

dat.fr<-data.frame(a=tetra)
circ.cent.tetra(dat.fr)

}
\seealso{
\code{\link{rv.tetraCM}} and \code{\link{rv.triCC}}
}
