% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rv.tri.cent}
\alias{rv.tri.cent}
\title{The index of the vertex region in a triangle that contains a given point}
\usage{
rv.tri.cent(p, tri, M)
}
\arguments{
\item{p}{A 2D point for which M-vertex region it resides in is to be determined in the
triangle tri}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle tri}
}
\value{
A list with two elements
\item{rv}{Index of the vertex whose region contains point, p; index of the vertex is the same as the row
number in the triangle, tri}
\item{tri}{The vertices of the triangle, tri, where row number corresponds to the vertex index rv
with rv=1 is row 1, rv=2 is row 2, and rv=3 is row 3.}
}
\description{
Returns the index of the related vertex in the triangle, tri, whose region contains point p.
Vertex regions are based on the general center M=(m1,m2) in Cartesian coordinates or
M=(alpha,beta,gamma) in barycentric coordinates in the interior of the triangle tri.
Vertices of the triangle tri are labeled according to the row number the vertex is recorded.
If the point, p, is not inside tri, then the function yields NA as output.
The corresponding vertex region is the polygon with the vertex, M, and projections from M
to the edges on the lines joining vertices and M. (see the illustration in the examples).
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);

M<-as.numeric(runif.tri(1,Tr)$g)
#M<-c(1.6,1.0)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)

cp2e.tri(Tr,M)
cp2e.tri(Tr,M=c(1,1))

Ds<-cp2e.tri(Tr,M)

Xlim<-range(Tr[,1])
Ylim<-range(Tr[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

if (dimension(M)==3) {M<-bary2cart(M,Tr)} #need to run this when M is given in barycentric coordinates

plot(Tr,pch=".",xlab="",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Tr)
L<-rbind(M,M,M); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)

xc<-Tr[,1]#+c(-.04,.05,.06)
yc<-Tr[,2]#+c(.02,.02,.03)
txt.str<-c("rv=1","rv=2","rv=3")
text(xc,yc,txt.str)

txt<-rbind(M,Ds)
xc<-txt[,1]#+c(-.02,.04,-.04,-.02)
yc<-txt[,2]#+c(-.02,.04,.04,-.06)
txt.str<-c("M","D1","D2","D3")
text(xc,yc,txt.str)

cp2e.tri(Tr,M)

dat.fr<-data.frame(a=Tr)
cp2e.tri(dat.fr,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{rv.triCM}}, \code{\link{rv.triCC}}, \code{\link{rv.bastriCC}},
\code{\link{rv.bastriCM}}, \code{\link{rv.bastri.cent}}, and \code{\link{rvTeCM}}
}
