% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rv.triCC}
\alias{rv.triCC}
\title{The index of the CC-vertex region in a triangle that contains a point}
\usage{
rv.triCC(p, tri)
}
\arguments{
\item{p}{A 2D point for which CC-vertex region it resides in is to be determined in the
triangle tri}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}
}
\value{
A list with two elements
\item{rv}{Index of the CC-vertex region that contains point, p in the triangle tri}
\item{tri}{The vertices of the triangle, where row number corresponds to the vertex index in rv.}
}
\description{
Returns the index of the vertex whose region contains point p in
a triangle tri=(A,B,C) and vertex regions are based on the circumcenter CC of tri.
(see the plots in the example for illustrations). The vertices of the  triangle tri are labeled as
1=A,2=B, and 3=C also according to the row number the vertex is recorded in tri.
If the point, p, is not inside tri, then the function yields NA as output.
The corresponding vertex region is the polygon whose interior points are closest to that vertex.
If tri is equilateral triangle, then CC and CM (center of mass) coincide.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
c1<-.4; c2<-.6;
#c1<-.5; c2<-.5;

P<-c(.3,.2)
rv.bastriCC(P,c1,c2)

P<-c(.6,.2)
rv.bastriCC(P,c1,c2)

P<-c(.5,.4)
rv.bastriCC(P,c1,c2)

P<-c(1.5,.4)
rv.bastriCC(P,c1,c2)

A<-c(0,0);B<-c(1,0);C<-c(c1,c2);
Tb<-rbind(A,B,C)
CC<-circ.cent.bastri(c1,c2) #the circumcenter
D1<-(B+C)/2; D2<-(A+C)/2; D3<-(A+B)/2;
Ds<-rbind(D1,D2,D3)

Xlim<-range(Tb[,1])
Ylim<-range(Tb[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Tb,asp=1,xlab="",ylab="",axes=TRUE,pch=".",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Tb)
L<-matrix(rep(CC,3),ncol=3,byrow=T); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)

txt<-rbind(Tb,CC,Ds)
xc<-txt[,1]+c(-.03,.03,0.02,-.01,.05,-.05,.01)
yc<-txt[,2]+c(.02,.02,.03,.06,.03,.03,-.03)
txt.str<-c("A","B","C","CC","D1","D2","D3")
text(xc,yc,txt.str)

RV1<-(A+D3+CC+D2)/4
RV2<-(B+D3+CC+D1)/4
RV3<-(C+D2+CC+D1)/4

txt<-rbind(RV1,RV2,RV3)
xc<-txt[,1]
yc<-txt[,2]
txt.str<-c("rv=1","rv=2","rv=3")
text(xc,yc,txt.str)

###
n<-20
dat<-runif.bastri(n,c1,c2)$g
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

Rv<-vector()
for (i in 1:n)
  Rv<-c(Rv,rv.bastriCC(dat[i,],c1,c2)$rv)
Rv

Xlim<-range(Tb[,1],dat[,1])
Ylim<-range(Tb[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Tb,asp=1,xlab="",pch=".",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
points(dat,pch=".")
polygon(Tb)
L<-matrix(rep(CC,3),ncol=3,byrow=T); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)
text(dat,labels=factor(Rv))

txt<-rbind(Tb,CC,Ds)
xc<-txt[,1]+c(-.03,.03,0.02,-.01,.05,-.05,.01)
yc<-txt[,2]+c(.02,.02,.03,.06,.03,.03,-.04)
txt.str<-c("A","B","C","CC","D1","D2","D3")
text(xc,yc,txt.str)

P<-c(.3,.2)
rv.bastriCC(P,c1,c2)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{rv.tri.cent}}, \code{\link{rv.triCM}}, \code{\link{rv.bastriCM}},
\code{\link{rv.bastriCC}}, \code{\link{rv.bastri.cent}}, and \code{\link{rvTeCM}}
}
