% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rv.triCM}
\alias{rv.triCM}
\title{The index of the CM-vertex region in a triangle that contains a given point}
\usage{
rv.triCM(p, tri)
}
\arguments{
\item{p}{A 2D point for which CM-vertex region it resides in is to be determined in the triangle
tri}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}
}
\value{
A list with two elements
\item{rv}{Index of the CM-vertex region that contains point, p in the triangle tri.}
\item{tri}{The vertices of the triangle, where row number corresponds to the vertex index in rv.}
}
\description{
Returns the index of the vertex whose region contains point p in
the triangle tri=(y1,y2,y3) with vertex regions are constructed with center of mass CM=(y1+y2+y3)/3
(see the plots in the example for illustrations). The vertices of triangle, tri, are labeled as 1,2,3
according to the row number the vertex is recorded in tri. If the point, p, is not inside tri, then the
function yields NA as output. The corresponding vertex region is the polygon with the vertex, CM, and
midpoints of the edges adjacent to the vertex.
See (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds})
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
ny<-10
Y<-cbind(runif(ny),runif(ny))
radius(A,Y)
radius(B,Y)
radius(C,Y)

radius(B,C)

nx<-10
X<-cbind(runif(nx),runif(nx))
rad<-rep(0,nx)
for (i in 1:nx)
rad[i]<-radius(X[i,],Y)$rad

Xlim<-range(X[,1]-rad,X[,1]+rad,Y[,1])
Ylim<-range(X[,2]-rad,X[,2]+rad,Y[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(rbind(Y),asp=1,pch=16,col=2,xlab="",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
points(rbind(X))
circles(X[,1],X[,2],rad,lty=1,lwd=1,col=4)

dat.fr<-data.frame(a=Y)
radii(A,dat.fr)

}
\seealso{
\code{\link{rv.tri.cent}}, \code{\link{rv.triCC}}, \code{\link{rv.bastriCM}},
\code{\link{rv.bastriCC}}, \code{\link{rv.bastri.cent}}, and \code{\link{rvTeCM}}
}
