% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rverts.tri.cent}
\alias{rverts.tri.cent}
\title{The indices of the vertex regions in a triangle that contains the points in a give data set}
\usage{
rverts.tri.cent(Dt, tri, M)
}
\arguments{
\item{Dt}{A set of 2D points representing the set of data points for which indices of the vertex regions
containing them are to be determined}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the riangle tri or the circumcenter of tri}
}
\value{
A list with two elements
\item{rvs}{Indices of the vertices whose regions contains points in Dt.}
\item{tri}{The vertices of the triangle, where row number corresponds to the vertex index in rv.}
}
\description{
Returns the indices of the vertices whose regions contain the points in data set Dt in
a triangle tri=T(A,B,C). Vertex regions are based on center M=(m1,m2) in Cartesian
coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the interior of the triangle
to the edges on the extension of the lines joining M to the vertices or based on the circumcenter of tri.
Vertices of triangle tri are labeled as 1,2,3 according to the row number the vertex is recorded
if a point in Dt is not inside tri, then the function yields NA as output for that entry.
The corresponding vertex region is the polygon with the vertex, M, and
projection points from M to the edges crossing the vertex (as the output of cp2e.tri(Tr,M))
or CC-vertex region. (see the examples for an illustration).
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:dom-num-NPE-Spat2011,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
r<-1.35
cen<-2

P<-c(1.4,1.0)
rverts.tri.nd(P,Tr,r,cen)

P<-c(1.8,.5)
rverts.tri.nd(P,Tr,r,cen)

P<-c(10.5,1.6)
rverts.tri.nd(P,Tr,r,cen)

n<-20
#set.seed(1)
dat<-runif.tri(n,Tr)$g
#dat<-cbind(runif(n,1,2),runif(n,0,2))
#dat<-cbind(runif(n),runif(n))

rverts.tri.nd(dat,Tr,r,cen)
rverts.tri.nd(rbind(dat,c(2,2)),Tr,r,cen)

rv<-rverts.tri.nd(dat,Tr,r,cen)

M<-cent.nondeg(Tr,r)[cen,];
Ds<-cp2edges.nd(Tr,r,cen)

Xlim<-range(Tr[,1],dat[,1])
Ylim<-range(Tr[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Tr,pch=".",xlab="",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Tr)
points(dat,pch=".",col=1)
L<-rbind(M,M,M); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)

xc<-Tr[,1]#+c(-.03,.05,.05)
yc<-Tr[,2]#+c(-.06,.02,.05)
txt.str<-c("rv=1","rv=2","rv=3")
text(xc,yc,txt.str)

txt<-rbind(M,Ds)
xc<-txt[,1]#+c(.02,.04,-.03,0)
yc<-txt[,2]#+c(.07,.03,.05,-.07)
txt.str<-c("M","D1","D2","D3")
text(xc,yc,txt.str)

text(dat,labels=factor(rv$rv))

P<-c(1.4,1.0)
rverts.tri.nd(P,Tr,r,cen)
rverts.tri.nd(dat,Tr,r,cen)

rverts.tri.nd(rbind(dat,dat),Tr,r,cen)

dat.fr<-data.frame(a=dat)
rverts.tri.nd(dat.fr,Tr,r,1)

dat.fr<-data.frame(a=Tr)
rverts.tri.nd(dat,dat.fr,r,1)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{rverts.triCM}}, \code{\link{rverts.triCC}} and \code{\link{rverts.tri.nd}}
}
