% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rverts.tri.nd}
\alias{rverts.tri.nd}
\title{The indices of the vertex regions in a triangle that contains the points in a give data set}
\usage{
rverts.tri.nd(Dt, tri, r, cent = 1)
}
\arguments{
\item{Dt}{A set of 2D points representing the set of data points for which indices of the vertex regions
containing them are to be determined}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region;
must be in (1,1.5] for this function}

\item{cent}{index of the center (as 1,2,3 corresponding to M1,M2,M3) which gives nondegenerate asymptotic
distribution of the domination number of PE-PCD for uniform data in tri for expansion parameter r>=1 in (1,1.5];
default cent=1.}
}
\value{
A list with two elements
\item{rv}{Indices (i.e., a vector of indices) of the vertices whose region contains points in Dt
in the triangle tri}
\item{tri}{The vertices of the triangle, where row number corresponds to the vertex index in rv.}
}
\description{
Returns the indices of the vertices whose regions contain the points in data set Dt in
a triangle tri=(A,B,C) and vertex regions are based on the center M which yields nondegenerate asymptotic
distribution of the domination number of PE-PCD for uniform data in tri for expansion parameter r>=1 in (1,1.5].
Vertices of triangle tri are labeled as 1,2,3 according to the row number the vertex is recorded
if a point in Dt is not inside tri, then the function yields NA as output for that entry.
The corresponding vertex region is the polygon with the vertex, M, and projection points on the edges.
The center label cent values 1,2,3 correspond to the vertices M1,M2,and M3; with default 1
(see the examples for an illustration).
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:dom-num-NPE-Spat2011,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#c1<-.4; c2<-.6; A<-c(0,0); B<-c(1,0); C<-c(c1,c2);
#A<-c(0,0); B<-c(1,0); C<-c(.5,sqrt(3.)/2);
A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);
r<-1.35

cp2edges.nd(Tr,r,cent=2)
cp2edges.nd(Tr,r,cent=5)
cp2edges.nd(Tr,r=2,cent=2)

T2<-rbind(A,A+2*(C-A),C)
cp2edges.nd(T2,r,cent=2)

Ms<-cent.nondeg(Tr,r)

#Ds<-cp2edges.nd(T,r,cent=1)
Ds<-cp2edges.nd(Tr,r,cent=2)
#Ds<-cp2edges.nd(T,r,cent=3)
D1<-Ds[1,]; D2<-Ds[2,]; D3<-Ds[3,]

Xlim<-range(Tr[,1])
Ylim<-range(Tr[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Tr,pch=".",xlab="",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Tr)
points(Ms,pch=".",col=1)
polygon(Ms,lty=2)

xc<-Tr[,1]#+c(-.02,.03,.02)
yc<-Tr[,2]#+c(.02,.02,.03)
txt.str<-c("A","B","C")
text(xc,yc,txt.str)

txt<-Ms
xc<-txt[,1]#+c(-.03,.04,.01)
yc<-txt[,2]#+c(.02,.02,.06)
txt.str<-c("M1","M2","M3")
text(xc,yc,txt.str)

points(Ds,pch=4,col=2)
L<-Tr; R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2,col=2)
txt<-Ds
xc<-txt[,1]#+c(.05,-.03,0)
yc<-txt[,2]#+c(03,.06,-.07)
txt.str<-c("D1","D2","D3")
text(xc,yc,txt.str)

cp2edges.nd(Tr,r,1)

dat.fr<-data.frame(a=Tr)
cp2edges.nd(dat.fr,r,1)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{rverts.triCM}}, \code{\link{rverts.triCC}} and \code{\link{rverts.tri.cent}}
}
