% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rverts.triCC}
\alias{rverts.triCC}
\title{The indices of the CC-vertex regions in a triangle that contains the points in a give data set}
\usage{
rverts.triCC(Dt, tri)
}
\arguments{
\item{Dt}{A set of 2D points representing the set of data points for which indices of the vertex regions
containing them are to be determined}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}
}
\value{
A list with two elements
\item{rv}{Indices (i.e., a vector of indices) of the vertices whose region contains points in Dt
in the triangle tri}
\item{tri}{The vertices of the triangle, where row number corresponds to the vertex index in rv.}
}
\description{
Returns the indices of the vertices whose regions contain the points in data set Dt in
a triangle tri=(A,B,C) and vertex regions are based on the circumcenter CC of tri.
(see the plots in the example for illustrations). The vertices of the  triangle tri are labeled as
1=A,2=B, and 3=C also according to the row number the vertex is recorded in tri.
If a point in Dt is not inside tri, then the function yields NA as output.
The corresponding vertex region is the polygon whose interior points are closest to that vertex.
If tri is equilateral triangle, then CC and CM (center of mass) coincide.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
#A<-runif(2); B<-runif(2); C<-runif(2);
#A<-runif(2,1,100); B<-runif(2,1,100); C<-runif(2,1,100);
Tr<-rbind(A,B,C);

P<-c(1.3,1.2)
rv.triCC(P,Tr)

P<-c(1.8,.5)
rv.triCC(P,Tr)

P<-c(1.6,1.4)
rv.triCC(P,Tr)

P<-c(.5,.8)
rv.triCC(P,Tr)

CC<-circ.cent.tri(Tr) #the circumcenter
D1<-(B+C)/2; D2<-(A+C)/2; D3<-(A+B)/2;
Ds<-rbind(D1,D2,D3)

Xlim<-range(Tr[,1],CC[1])
Ylim<-range(Tr[,2],CC[2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Tr,asp=1,xlab="",ylab="",pch=".",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Tr)
L<-matrix(rep(CC,3),ncol=3,byrow=T); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)

txt<-rbind(Tr,CC,Ds)
xc<-txt[,1]#+c(-.07,.08,.06,.12,-.1,-.1,-.09)
yc<-txt[,2]#+c(.02,-.02,.03,.0,.02,.06,-.04)
txt.str<-c("A","B","C","CC","D1","D2","D3")
text(xc,yc,txt.str)

RV1<-(A+.5*(D3-A)+A+.5*(D2-A))/2
RV2<-(B+.5*(D3-B)+B+.5*(D1-B))/2
RV3<-(C+.5*(D2-C)+C+.5*(D1-C))/2

txt<-rbind(RV1,RV2,RV3)
xc<-txt[,1]#+c(-.08,.0,-.04)
yc<-txt[,2]#+c(.02,-.2,.2)
txt.str<-c("rv=1","rv=2","rv=3")
text(xc,yc,txt.str)

###
n<-20
dat<-runif.tri(n,Tr)$g
#dat<-cbind(runif(n,1,2),runif(n,0,2))
#dat<-cbind(runif(n),runif(n))

Rv<-vector()
for (i in 1:n)
  Rv<-c(Rv,rv.triCC(dat[i,],Tr)$rv)
Rv

Xlim<-range(Tr[,1],dat[,1])
Ylim<-range(Tr[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Tr,asp=1,xlab="",ylab="",pch=".",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Tr)
points(dat,pch=".")
L<-matrix(rep(CC,3),ncol=3,byrow=T); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)
text(dat,labels=factor(Rv))

txt<-rbind(Tr,CC,Ds)
xc<-txt[,1]#+c(-.07,.08,.06,.12,-.1,-.1,-.09)
yc<-txt[,2]#+c(.02,-.02,.03,.0,.02,.06,-.04)
txt.str<-c("A","B","C","CC","D1","D2","D3")
text(xc,yc,txt.str)

P<-c(1.3,1.2)
rv.triCC(P,Tr)

dat.fr<-data.frame(a=Tr)
rv.triCC(P,dat.fr)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{rverts.triCM}}, \code{\link{rverts.tri.cent}}, and \code{\link{rverts.tri.nd}}
}
