% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rverts.triCM}
\alias{rverts.triCM}
\title{The indices of the CM-vertex regions in a triangle that contains the points in a give data set}
\usage{
rverts.triCM(Dt, tri)
}
\arguments{
\item{Dt}{A set of 2D points representing the set of data points for which indices of the vertex regions
containing them are to be determined}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}
}
\value{
A list with two elements
\item{rv}{Indices (i.e., a vector of indices) of the vertices whose region contains points in Dt
in the triangle tri}
\item{tri}{The vertices of the triangle, where row number corresponds to the vertex index in rv.}
}
\description{
Returns the indices of the vertices whose regions contain the points in data set Dt in
a triangle tri=(A,B,C) and vertex regions are based on the center of mass CM of tri.
(see the plots in the example for illustrations). The vertices of the  triangle tri are labeled as
1=A,2=B, and 3=C also according to the row number the vertex is recorded in tri.
If a point in Dt is not inside tri, then the function yields NA as output for that entry.
The corresponding vertex region is the polygon
with the vertex, CM, and midpoints the edges crossing the vertex.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
c1<-.4; c2<-.6;
A<-c(0,0); B<-c(1,0); C<-c(c1,c2);
Tb<-rbind(A,B,C)
n<-20

#set.seed(1)
dat<-runif.bastri(n,c1,c2)$g
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

M<-as.numeric(runif.bastri(1,c1,c2)$g)
#M<-c(.6,.3)
#M<-circ.cent.tri(Tb)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)
r<-2

Gam2PEbastri(dat[1,],dat[2,],dat,r,c1,c2,M)
Gam2PEbastri(c(1,1),dat[2,],dat,r,c1,c2,M)

Gam2PEbastri(c(1,2),dat[2,],dat,r,c1,c2,M)
Gam2PEbastri(c(1,2),dat[2,],dat,r,c1,c2,M,ch.data.pnts = TRUE)


Gam2PEbastri(c(1,2),c(1,3),rbind(c(1,2),c(1,3)),r,c1,c2,M)
Gam2PEbastri(c(1,2),c(1,3),rbind(c(1,2),c(1,3)),r,c1,c2,M,ch.data.pnts = TRUE)

ind.gam2<-vector()
for (i in 1:(n-1))
  for (j in (i+1):n)
  {if (Gam2PEbastri(dat[i,],dat[j,],dat,r,c1,c2,M)==1)
    ind.gam2<-rbind(ind.gam2,c(i,j))}
ind.gam2

#or try
rv1<-rv.bastri.cent(dat[1,],c1,c2,M)$rv;
rv2<-rv.bastri.cent(dat[2,],c1,c2,M)$rv;
Gam2PEbastri(dat[1,],dat[2,],dat,r,c1,c2,M,rv1,rv2)

#or try
rv1<-rv.bastri.cent(dat[1,],c1,c2,M)$rv;
Gam2PEbastri(dat[1,],dat[2,],dat,r,c1,c2,M,rv1)

#or try
rv2<-rv.bastri.cent(dat[2,],c1,c2,M)$rv;
Gam2PEbastri(dat[1,],dat[2,],dat,r,c1,c2,M,rv2=rv2)

P1<-c(.4,.2)
P2<-c(.6,.2)
Gam2PEbastri(P1,P2,dat,r,c1,c2,M)

Gam2PEbastri(P1,P2,rbind(dat,dat),r,c1,c2,M)
Gam2PEbastri(P1,P2,cbind(dat,dat),r,c1,c2,M)
Gam2PEbastri(P1,P2, rbind(c("a","b"),dat),r,c1,c2,M)

dat.fr<-data.frame(a=dat)
Gam2PEbastri(P1,P2,dat.fr,r,c1,c2,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{rverts.tri.cent}}, \code{\link{rverts.triCC}} and \code{\link{rverts.tri.nd}}
}
