% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rvTe.cent}
\alias{rvTe.cent}
\title{The index of the vertex region in the standard equilateral triangle that contains a given point}
\usage{
rvTe.cent(pt, M)
}
\arguments{
\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the standard equilateral triangle Te}

\item{p}{A 2D point for which M-vertex region it resides in is to be determined in the
standard equilateral triangle Te}
}
\value{
A list with two elements
\item{rv}{Index of the vertex whose region contains point, pt}
\item{tri}{The vertices of the triangle, Te, where row number corresponds to the vertex index in rv
with row 1=(0,0), row 2=(1,0), and row 3=(1/2,sqrt(3)/2).}
}
\description{
Returns the index of the vertex whose region contains point pt in standard equilateral triangle
Te=T((0,0),(1,0),(1/2,sqrt(3)/2)) with vertex regions are constructed with center M=(m1,m2)
in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the interior of Te.
(see the plots in the example for illustrations). The vertices of triangle, Te, are labeled as 1,2,3
according to the row number the vertex is recorded in Te. If the point, pt, is not inside Te, then the
function yields NA as output. The corresponding vertex region is the polygon with the vertex, M, and
projections from M to the edges on the lines joining vertices and M.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
A<-c(0,0); B<-c(1,0); C<-c(.5,sqrt(3.)/2);
Te<-rbind(A,B,C)

n<-20
#set.seed(1)
dat<-runifTe(n)$gen.points
#dat<-cbind(runif(n),runif(n))
#dat<-cbind(runif(n,1,2),runif(n,0,1))

rvTeCM(dat[1,])
rvTeCM(c(.7,.2))
rvTeCM(c(0,1))

Rv<-vector()
for (i in 1:n)
  Rv<-c(Rv,rvTeCM(dat[i,])$rv)
Rv

CM<-(A+B+C)/3
D1<-(B+C)/2; D2<-(A+C)/2; D3<-(A+B)/2;
Ds<-rbind(D1,D2,D3)

Xlim<-range(Te[,1],dat[,1])
Ylim<-range(Te[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(Te,asp=1,pch=".",xlab="",ylab="",axes=TRUE,xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
polygon(Te)
points(dat,pch=".",col=1)
L<-matrix(rep(CM,3),ncol=3,byrow=T); R<-Ds
segments(L[,1], L[,2], R[,1], R[,2], lty=2)

txt<-rbind(Te,CM)
xc<-txt[,1]+c(-.02,.03,.02,0)
yc<-txt[,2]+c(.02,.02,.03,.05)
txt.str<-c("A","B","C","CM")
text(xc,yc,txt.str)

text(dat,labels=factor(Rv))

rvTeCM(c(.7,.2))

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{rvTeCM}}, \code{\link{rv.tri.cent}}, \code{\link{rv.triCC}},
\code{\link{rv.bastriCC}}, \code{\link{rv.triCM}}, and \code{\link{rv.bastri.cent}}
}
