% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{rvTeCM}
\alias{rvTeCM}
\title{The index of the CM-vertex region in the standard equilateral triangle that contains a given point}
\usage{
rvTeCM(pt)
}
\arguments{
\item{p}{A 2D point for which CM-vertex region it resides in is to be determined in the
standard equilateral triangle Te}
}
\value{
A list with two elements
\item{rv}{Index of the vertex whose region contains point, pt}
\item{tri}{The vertices of the triangle, Te, where row number corresponds to the vertex index in rv
with row 1=(0,0), row 2=(1,0), and row 3=(1/2,sqrt(3)/2).}
}
\description{
Returns the index of the vertex whose region contains point pt in standard equilateral triangle
Te=T((0,0),(1,0),(1/2,sqrt(3)/2)) with vertex regions are constructed with center of mass CM
(see the plots in the example for illustrations). The vertices of triangle, Te, are labeled as 1,2,3
according to the row number the vertex is recorded in Te. If the point, pt, is not inside Te, then the
function yields NA as output. The corresponding vertex region is the polygon with the vertex, CM, and
midpoints of the edges adjacent to the vertex.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
c1<-.4; c2<-.6
A<-c(0,0); B<-c(1,0); C<-c(c1,c2);
Tb<-rbind(A,B,C);

M<-as.numeric(runif.bastri(1,c1,c2)$g)
#M<-c(.6,.2)
#M<-circ.cent.tri(Tb)
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)
#M<-c(.3,.3)
r<-2

P1<-as.numeric(runif.bastri(1,c1,c2)$g)
P2<-as.numeric(runif.bastri(1,c1,c2)$g)
IndNPEbastri(P1,P2,r,c1,c2,M)

P1<-c(.4,.2)
P2<-c(.5,.26)
IndNPEbastri(P1,P2,r,c1,c2,M)
IndNPEbastri(P2,P1,r,c1,c2,M)
IndNPEbastri(P1,P1,r,c1,c2,M)

#or try
Rv<-rv.bastri.cent(P1,c1,c2,M)$rv
IndNPEbastri(P1,P2,r,c1,c2,M,Rv)

P1<-c(1.4,.2)
P2<-c(.5,.26)
IndNPEbastri(P1,P2,r,c1,c2,M)
IndNPEbastri(P1,P1,r,c1,c2,M)

#or try
Rv<-rv.bastri.cent(P1,c1,c2,M)$rv
IndNPEbastri(P1,P2,r,c1,c2,M,Rv)
IndNPEbastri(P1,P1,r,c1,c2,M,Rv)

P1<-c(.4,.2)
P2<-c(1.5,.26)
IndNPEbastri(P1,P2,r,c1,c2,M)

P1<-c(.4,.2)
P2<-c(.5,.26)
IndNPEbastri(P1,P2,r,c1,c2,M)

}
\references{
\insertAllCited{}

\insertAllCited{}
}
\seealso{
\code{\link{rv.bastriCM}}, \code{\link{rv.tri.cent}}, \code{\link{rv.triCC}},
\code{\link{rv.bastriCC}}, \code{\link{rv.triCM}}, and \code{\link{rv.bastri.cent}}
}
