% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{seg.tri.supp}
\alias{seg.tri.supp}
\title{The auxiliary triangle to define the support of type I segregation}
\usage{
seg.tri.supp(delta, tri)
}
\arguments{
\item{delta}{a positive real number between 0 and 1}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}
}
\value{
the vertices of the triangle (stacked row-wise) whose intersection with a general triangle
gives the support for type I segregation for the given delta
}
\description{
Returns the triangle whose intersection with a general triangle gives the support for
type I segregation given the delta (i.e., delta*100 % area of a triangle around the vertices is chopped off).
See the plot in the examples.
Caveat: the vertices of this triangle may be outside the triangle, tri, depending on the value of delta
(i.e., for small values of delta)
}
\examples{
nx<-40 #number of X points (target)
ny<-10 #number of Y points (nontarget)

#set.seed(1)
Yp<-cbind(runif(ny),runif(ny))
dat<-runifMT(nx,Yp)$g #data under CSR in the convex hull of Ypoints
#dat<-cbind(runif(nx),runif(nx))

tr.ind<-indices.Del.tri(dat,Yp) #indices of the Delaunay triangles
tr.ind

#or use
DTY<-tri.mesh(Yp[,1],Yp[,2],duplicate="remove") #Delaunay triangulation based on Y points
tr.ind<-indices.Del.tri(dat,Yp,DTY) #indices of the Delaunay triangles
tr.ind

Xlim<-range(Yp[,1],dat[,1])
Ylim<-range(Yp[,2],dat[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

#plot of the data in the convex hull of Y points together with the Delaunay triangulation
#par(pty="s")
plot(dat,main=" ", xlab=" ", ylab=" ",xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05),pch=".")
plot.tri(DTY, add=TRUE, do.points=T,pch=16,col="blue")
text(dat,labels = factor(tr.ind) )

Yp<-rbind(c(.3,.2),c(.4,.5),c(.14,.15))
indices.Del.tri(c(.25,.25),Yp)

dat<-rbind(c(.25,.25),c(.15,.25))
indices.Del.tri(dat,Yp)

dat.fr<-data.frame(a=dat)
indices.Del.tri(dat.fr,Yp)

dat.fr<-data.frame(a=Yp)
indices.Del.tri(c(.25,.25),dat.fr)

}
\seealso{
\code{\link{rsegTe}} and \code{\link{rsegMT}}
}
