% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{slope}
\alias{slope}
\title{The slope of a line}
\usage{
slope(a, b)
}
\arguments{
\item{a, b}{2D points that determine the straight line (i.e., through which the straight line is passing)}
}
\value{
Slope of the line joining 2D points \code{a} and \code{b}
}
\description{
Returns the slope of the line joining two distinct 2D points a and b
}
\examples{
dimension(3)
dim(3)

A<-c(1,2)
dimension(A)
dim(A)

B<-c(2,3)
dimension(rbind(A,B,A))
dimension(cbind(A,B,A))

M<-matrix(runif(20),ncol=5)
dimension(M)
dim(M)

dat.fr<-data.frame(a=A,b=B)
dimension(dat.fr)
dim(dat.fr)

dimension(c("a","b"))
}
\seealso{
\code{\link{Line}}, \code{\link{paraline}}, and \code{\link{perpline}}
}
