% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{TSArcDensPE1D}
\alias{TSArcDensPE1D}
\title{An object of class "htest"

A test of uniformity of 1D data in a given interval based on Proportional Edge Proximity Catch Digraph
(PE-PCD)}
\usage{
TSArcDensPE1D(dat, r, c, int, alternative = c("two.sided", "less",
  "greater"), conf.level = 0.95)
}
\arguments{
\item{dat}{A set or vector of 1D points which constitute the vertices of PE-PCD}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be >= 1}

\item{c}{A positive real number in (0,1) parameterizing the center inside int=(a,b).
For the interval, int=(a,b), the parameterized center is Mc=a+c*(b-a)}

\item{int}{A vector of two real numbers representing an interval}

\item{alternative}{Type of the alternative hypothesis in the test, one of "two.sided", "less", "greater".}

\item{conf.level}{Level of the confidence interval, default is 0.95, for the arc density of PE-PCD based on
the 1D data set dat.}
}
\value{
A list with the elements
\item{statistic}{Test statistic}
\item{p.value}{p-value for the hypothesis test for the corresponding alternative}
\item{conf.int}{Confidence interval for the arc density at the given confidence level conf.level and
depends on the type of alternative.}
\item{estimate}{Estimate of the parameter, i.e., arc density}
\item{null.value}{Hypothesized value for the parameter, i.e., the null arc density, which is usually the
mean arc density under uniform distribution.}
\item{alternative}{Type of the alternative hypothesis in the test, one of "two.sided", "less", "greater"}
\item{method}{Description of the hypothesis test}
\item{data.name}{Name of the data set}
}
\description{
This is an "htest" (i.e., hypothesis test) function which performs a hypothesis test of uniformity of 1D data
in one interval based on the normal approximation of the arc density of the PE-PCD with expansion parameter r>=1
and centrality parameter c in (0,1). The function  yields the test statistic, p-value for the
corresponding alternative, the confidence interval, estimate and null value for the parameter of interest
(which is the arc density), and method and name of the data set used. The null hypothesis is that data is
uniform in a finite interval (i.e. arc density of PE-PCD equals to its expected value under uniform
distribution) and alternative could be two-sided, or left-sided (i.e. data is accumulated around the end
points) or right-sided (i.e. data is accumulated around the mid point or center Mc).
See also (\insertCite{ceyhan:metrika-2012,ceyhan:revstat-2016;textual}{pcds}).
}
\examples{
asyvarPE1D(1.2,.8)

rseq<-seq(1.01,5,by=.05)
cseq<-seq(0.01,.99,by=.05)

lrseq<-length(rseq)
lcseq<-length(cseq)

var.grid<-matrix(0,nrow=lrseq,ncol=lcseq)
for (i in 1:lrseq)
  for (j in 1:lcseq)
  {
    var.grid[i,j]<-asyvarPE1D(rseq[i],cseq[j])
  }

persp(rseq,cseq,var.grid, xlab="r", ylab="c", zlab="var(r,c)", theta = -30, phi = 30, expand = 0.5, col = "lightblue",
      ltheta = 120, shade = 0.05, ticktype = "detailed")

}
\seealso{
\code{\link{TS.ArcDens.CS1D}}
}
