% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{TSDomPEBin.1D}
\alias{TSDomPEBin.1D}
\title{An object of class "htest"

A test of segregation/association based on domination number of Proportional Edge Proximity Catch Digraph
(PE-PCD) for 1D Data - Binomial Approximation}
\usage{
TSDomPEBin.1D(Xp, Yp = NULL, int, c = 0.5, asy.bin = F,
  end.int.cor = F, alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95)
}
\arguments{
\item{Xp}{A set of 1D points which constitute the vertices of the PE-PCD}

\item{Yp}{A set of 1D points which constitute the end points of the subintervals, default is NULL.
When Yp=NULL, the support interval (a,b) is partitioned as Yp=(b-a)*(0:nint)/nint
where nint=round(sqrt(nx),0) and nx is the number of Xp points}

\item{int}{Support interval (a,b) with a<b. Uniformness of Xp points in this interval is tested.}

\item{c}{A positive real number which serves as the centrality parameter in PE proximity region;
must be in (0,1) (default c=.5)}

\item{asy.bin}{A logical argument for the use of aymptotic probability of success for the binomial distribution,
default asy.bin=FALSE. It is an option only when Yp is not provided. When Yp is provided or when Yp is not provided
but asy.bin=T, aymptotic probability of success for the binomial distribution is used. When Yp is not provided
and asy.bin=F, the finite sample aymptotic probability of success for the binomial distribution is used with number
of trials equals to expected number of Xp points per subinterval.}

\item{end.int.cor}{A logical argument for end interval correction, default ch=FALSE,
recommended when both Xp and Yp have the same interval support}

\item{alternative}{Type of the alternative hypothesis in the test, one of "two.sided", "less", "greater".}

\item{conf.level}{Level of the confidence interval, default is 0.95, for the probability of success
(i.e. Pr(domination number=2) for PE-PCD whose vertices are the 1D data set Xp.}
}
\value{
A list with the elements
\item{statistic}{Test statistic}
\item{p.value}{p-value for the hypothesis test for the corresponding alternative}
\item{conf.int}{Confidence interval for Pr(Domination Number=2) at the given level conf.level and
depends on the type of alternative.}
\item{estimate}{a vector with two entries: first is is the estimate of the parameter, i.e.,
Pr(Domination Number=2) and second is the domination number}
\item{null.value}{Hypothesized value for the parameter, i.e., the null value for Pr(Domination Number=2)}
\item{alternative}{Type of the alternative hypothesis in the test, one of "two.sided", "less", "greater"}
\item{method}{Description of the hypothesis test}
\item{data.name}{Name of the data set}
}
\description{
This is an "htest" (i.e., hypothesis test) function which performs a hypothesis test of complete spatial
randomness (CSR) or uniformity of Xp points within the subintervals based on Yp points (both residing in the
interval (a,b)). If Yp=NULL the support interval (a,b) is partitioned as Yp<-(b-a)*(0:nint)/nint
where nint<-round(sqrt(nx),0) and nx is number of Xp points, and the test is for testing the uniformity of Xp
points in the interval (a,b). If Yp points are given, the test is for testing the spatial interaction between
Xp and Yp points. In either case, the null hypothesis is uniformness of Xp points on (a,b).
Yp determines the end points of the intervals (i.e., partition the real line via intervalization)
where end points are the order statistics of Yp points.
The alternatives are segregation (where Xp points cluster away from Yp points i.e. cluster around the centers of the subintervals)
and association (where Xp points cluster around Yp points). The test is based on the (asymptotic) binomial
distribution of the domination number of PE-PCD for uniform 1D data in the subintervals based on Yp points.
The function yields the test statistic, p-value for the corresponding
alternative, the confidence interval, estimate and null value for the parameter of interest (which is
Pr(Domination Number=2)), and method and name of the data set used.
Under the null hypothesis of uniformity of Xp points in the interval based on Yp points, probability of success
(i.e. Pr(Domination Number=2) ) equals to its expected value under the uniform distribution) and
alternative could be two-sided, or left-sided (i.e. data is accumulated around the Yp points, or association)
or right-sided (i.e. data is accumulated around the centers of the subintervals, or segregation).
PE proximity region is constructed with the expansion parameter r>=1 and centrality parameter c which yields
M-vertex regions. More precisely M=a+c*(b-a) for the centrality parameter c and for a given c in (0,1), the
expansion parameter r is taken to be 1/max(c,1-c) which yields non-degenerate asymptotic distribution of the
domination number. The test statistic is based on the binomial distribution, when domination number is scaled
to have value 0 and 1 in the one interval case (i.e., Domination Number - 1 for the one interval case).
That is, the test statistic is based on the domination number for Xp points inside the interval based on Yp points
for the PE-PCD . For this approach to work, Xp must be large for each subinterval, but 5 or more per subinterval
seems to work fine in practice. Probability of success is chosen in the following way for various parameter choices.
asy.bin A logical argument for the use of aymptotic probability of success for the binomial distribution,
default is asy.bin=FALSE. It is an option only when Yp is not provided. When Yp is provided or when Yp is not provided
but asy.bin=T, aymptotic probability of success for the binomial distribution is used. When Yp is not provided
and asy.bin=F, the finite sample aymptotic probability of success for the binomial distribution is used with number
of trials equals to expected number of Xp points per subinterval.
}
\examples{
a<-0; b<-10; int<-c(a,b)
c<-.4
r<-2
#c<-runif(1); r<-rexp(1)+1
#a<-runif(1); b<-runif(1); int<-range(a,b); a<-int[1]; b<-int[2]
nx<-40; ny<-10

#set.seed(1)
Xp<-runif(nx,a,b)
Yp<-runif(ny,a,b)

plot.intervals(Xp,Yp,main="Xp Points and Intervals Based on Yp Points")
plot.regsPEMI(Xp,Yp,r,c,main="PE Proximity Regions for Xp points - Intervalization is by Yp Points")

PEdom1D(Xp,Yp,r,c)

PEdom1D(Xp,Yp,r,c=.25)
PEdom1D(Xp,Yp,r,c=.01)
PEdom1D(Xp,Yp,r=1.25,c)

}
\seealso{
\code{\link{TSDomPEBin}} and \code{\link{PEdom1D}}
}
