% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{XinCHY}
\alias{XinCHY}
\title{Points from one class inside the convex hull of the points from the other class}
\usage{
XinCHY(Xp, Yp)
}
\arguments{
\item{Xp}{A set of 2D points which constitute the data set}

\item{Yp}{A set of 2D points which constitute the vertices of the Delaunay triangles}
}
\value{
Xp points inside the convex hull of Yp points
}
\description{
Given two 2D data sets, Xp and Yp, it returns the Xp points inside the convex hull of Yp points.
See (\insertCite{okabe:2000,ceyhan:comp-geo-2010,renka:1996;textual}{pcds}) for more on Delaunay triangulation and the tripack algorithm.
}
\examples{
nx<-100
ny<-10

#set.seed(1)
Xp<-cbind(runif(nx,0,1),runif(nx,0,1))
#Xp<-cbind(runif(nx,0.5,1.5),runif(nx,0,1))
#Xp<-cbind(runif(nx,1,2),runif(nx,0,1))
#Xp<-c(.5,.5)
Yp<-cbind(runif(ny,0,1),runif(ny,0,1))

M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-"CC"
#M<-c(1.3,1.3)
r<-1.5
#r<-2

IM<-IncMatPEMT(Xp,Yp,r,M)
IM
dom.greedy(IM)
#dom.exact(IM) #might take a long time in this brute-force fashion ignoring the
#disconnected nature of the digraph inherent by the geometric construction of it
PEdomMTnd(Xp,Yp,r)

Arcs<-ArcsPEMT(Xp,Yp,r,M)
Arcs
summary(Arcs)
plot(Arcs)

IncMatPEMT(Xp,Yp,r,M)

IncMatPEMT(Xp,rbind(Yp,Yp),r,M)

dat.fr<-data.frame(a=Xp)
IncMatPEMT(dat.fr,Yp,r,M)

dat.fr<-data.frame(a=Yp)
IncMatPEMT(Xp,dat.fr,r,M)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plotDeltri}}
}
