% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IncMatAStri}
\alias{IncMatAStri}
\title{Incidence matrix for Arc Slice Proximity Catch Digraphs (AS-PCDs) - one triangle case}
\usage{
IncMatAStri(dat, tri, M = "CC")
}
\arguments{
\item{dat}{A set of 2D points which constitute the vertices of AS-PCD}

\item{tri}{Three 2D points, stacked row-wise, each row representing a vertex of the triangle}

\item{M}{"CC" which stand for circumcenter of the triangle tri or a 2D point in Cartesian coordinates or
a 3D point in barycentric coordinates which serves as a center in the interior of tri;
default is M="CC" i.e. the circumcenter of tri}
}
\value{
Incidence matrix for the AS-PCD whose vertices are 2D data set, dat,
and AS proximity regions are defined with respect to the triangle tri and
vertex regions based on circumcenter.
}
\description{
Returns the incidence matrix for the AS-PCD whose vertices are the given 2D numerical data set, dat.
AS proximity regions are defined with respect to the triangle tri=T(v=1,v=2,v=3) and
vertex regions based on the center M="CC" for circumcenter of tri;
or M=(m1,m2) in Cartesian coordinates or M=(alpha,beta,gamma) in barycentric coordinates in the
interior of the triangle tri; default is M="CC" i.e. circumcenter of tri.
Loops are allowed, so the diagonal entries are all equal to 1.
See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}).
}
\examples{
nx<-100
ny<-10

#set.seed(1)
Xp<-cbind(runif(nx),runif(nx))
#Xp<-cbind(runif(nx,.5,1.5),runif(nx))
#Xp<-cbind(runif(nx,1,2),runif(nx))
#Xp<-c(.5,.5)
Yp<-cbind(runif(ny),runif(ny))

plotDeltri(Xp,Yp,xlab="",ylab="")

M<-"CC"
#M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-c(1.3,1.3)

NumArcsASMT(Xp,Yp,M)

NumArcsASMT(c(.4,.2),Yp,M)

NumArcsASMT(Xp1,rbind(Yp1,Yp1),M)

dat.fr<-data.frame(a=Xp1)
NumArcsASMT(dat.fr,Yp,M)


}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IncMatASMT}}, \code{\link{IncMatPEtri}}, and \code{\link{IncMatCStri}}
}
