% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentSim1D.R
\name{IncMatCS1D}
\alias{IncMatCS1D}
\title{Incidence matrix for Central Similarity Proximity Catch Digraphs (CS-PCDs)
for 1D data - multiple interval case}
\usage{
IncMatCS1D(Xp, Yp, t, c = 0.5)
}
\arguments{
\item{Xp}{a set of 1D points which constitutes the vertices of the digraph.}

\item{Yp}{a set of 1D points which constitutes the end points of the intervals
that partition the real line.}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region.}

\item{c}{A positive real number in \eqn{(0,1)} parameterizing the center inside middle intervals
with the default \code{c=.5}.
For the interval, \eqn{(a,b)}, the parameterized center is \eqn{M_c=a+c(b-a)}.}
}
\value{
Incidence matrix for the CS-PCD with vertices being 1D data set, \code{Xp},
and \code{Yp} determines the end points of the intervals (the multi-interval case)
}
\description{
Returns the incidence matrix for the CS-PCD for a given 1D numerical data set, \code{Xp},
as the vertices of the digraph and \code{Yp} determines the end points of the intervals (in the multi-interval case).
Loops are allowed, so the diagonal entries are all equal to 1.

CS proximity region is constructed
with an expansion parameter \eqn{t>0} and a centrality parameter \eqn{c \in (0,1)}.

See also (\insertCite{ceyhan:revstat-2016;textual}{pcds}).
}
\examples{
\dontrun{
t<-2
c<-.4
a<-0; b<-10;
nx<-10; ny<-4

set.seed(1)
Xp<-runif(nx,a,b)
Yp<-runif(ny,a,b)

IM<-IncMatCS1D(Xp,Yp,t,c)
IM
dom.greedy(IM) #try also dom.exact(IM)  #might take a long time depending on nx
IndUBdom(IM,5)
}

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IncMatCS1D}}, \code{\link{IncMatPEtri}}, and \code{\link{IncMatPE}}
}
\author{
Elvan Ceyhan
}
