% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{IncMatCS1D}
\alias{IncMatCS1D}
\title{Incidence matrix for Central Similarity Proximity Catch Digraphs (CS-PCDs)
for 1D data - multiple interval case}
\usage{
IncMatCS1D(Xp, Yp, t, c)
}
\arguments{
\item{Xp}{a set of 1D points which constitutes the vertices of the digraph.}

\item{Yp}{a set of 1D points which constitutes the end points of the intervals
that partition the real line.}

\item{t}{A positive real number which serves as the expansion parameter in CS proximity region.}

\item{c}{A positive real number in \eqn{(0,1)} parameterizing the center inside \eqn{int=(a,b)}.
For the interval, \eqn{int=(a,b)}, the parameterized center is \eqn{M_c=a+c(b-a)}.}
}
\value{
Incidence matrix for the CS-PCD with vertices being 1D data set, \code{Xp},
and \code{Yp} determines the end points of the intervals (the multi-interval case)
}
\description{
Returns the incidence matrix for the CS-PCD for a given 1D numerical data set, \code{Xp},
as the vertices of the digraph and \code{Yp} determines the end points of the intervals (in the multi-interval case).
Loops are allowed, so the diagonal entries are all equal to 1.

CS proximity region is constructed
with an expansion parameter \eqn{t>0} and a centrality parameter \code{c} in \eqn{(0,1)}.

See also (\insertCite{ceyhan:revstat-2016;textual}{pcds}).
}
\examples{
t<-2
c<-.4
a<-0; b<-10;
nx<-10; ny<-4

set.seed(1)
Xp<-runif(nx,a,b)
Yp<-runif(ny,a,b)

IM<-IncMatCS1D(Xp,Yp,t,c)
IM
dom.greedy(IM)
\donttest{
dom.exact(IM)  #might take a long time depending on nx
}
IndUBdom(IM,5)

Arcs<-ArcsCSMI(Xp,Yp,t,c)
Arcs
summary(Arcs)
plot(Arcs)

IncMatCS1D(Xp,Yp+10,t,c)

t<-2
c<-.4
a<-0; b<-10;
#nx is number of X points (target) and ny is number of Y points (nontarget)
nx<-20; ny<-4;  #try also nx<-40; ny<-10 or nx<-1000; ny<-10;

Xp<-runif(nx,a,b)
Yp<-runif(ny,a,b)

IncMatCS1D(Xp,Yp,t,c)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IncMatCS1D}}, \code{\link{IncMatPEtri}}, and \code{\link{IncMatPEMT}}
}
