% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{int.circ.line}
\alias{int.circ.line}
\title{The points of intersection of a line and a circle}
\usage{
int.circ.line(pt1, pt2, cent, rad)
}
\arguments{
\item{pt1, pt2}{2D points that determine the straight line (i.e., through which the straight line passes)}

\item{cent}{A 2D point representing the center of the circle}

\item{rad}{A positive real number representing the radius of the circle}
}
\value{
point(s) of intersection between the circle and the line (if they do not intersect, the function
yields \code{NULL} as the output)
}
\description{
Returns the intersection point(s) of a line and a circle. The line is determined by the two points
\code{pt1} and \code{pt2} and the circle is centered at point \code{cent} and has radius \code{rad}.
If the circle does not intersect the line, the function yields \code{NULL};
if the circle intersects at only one point, it yields only that point;
otherwise it yields both intersection points as output.
When there are two intersection points, they are listed in the order of the x-coordinates of \code{pt1} and \code{pt2};
and if the x-coordinates of \code{pt1} and \code{pt2} are equal, intersection points are listed in the order of
y-coordinates of \code{pt1} and \code{pt2}.
}
\examples{
P1<-c(.3,.2)*100
P2<-c(.6,.3)*100
cent<-c(1.1,1.1)*100
rad<-2*100

int.circ.line(P1,P2,cent,rad)
int.circ.line(P2,P1,cent,rad)
int.circ.line(P1,P1+c(0,1),cent,rad)
int.circ.line(P1+c(0,1),P1,cent,rad)

dp2l(cent,P1,P2)
rad2<-dp2l(cent,P1,P2)$d
int.circ.line(P1,P2,cent,rad2)
int.circ.line(P1,P2,cent,rad=.8)
int.circ.line(P1,P2,cent,rad=.78)

#plot of the line and the circle
A<-c(.3,.2); B<-c(.6,.3); cent<-c(1,1); rad<-2 #check dp2l(cent,A,B)$dis, .3

IPs<-int.circ.line(A,B,cent,rad)

xr<-range(A[1],B[1],cent[1])
xf<-(xr[2]-xr[1])*.1 #how far to go at the lower and upper ends in the x-coordinate
x<-seq(xr[1]-rad-xf,xr[2]+rad+xf,l=20)  #try also l=100
lnAB<-Line(A,B,x)
y<-lnAB$y

Xlim<-range(x,cent[1])
Ylim<-range(y,A[2],B[2],cent[2]-rad,cent[2]+rad)
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plot(rbind(A,B,cent),pch=1,asp=1,xlab="x",ylab="y",
xlim=Xlim+xd*c(-.05,.05),ylim=Ylim+yd*c(-.05,.05))
lines(x,y,lty=1)
interp::circles(cent[1],cent[2],rad)
IP.txt<-c()
if (!is.null(IPs))
{
  for (i in 1:(length(IPs)/2))
    IP.txt<-c(IP.txt,paste("I",i, sep = ""))
}
txt<-rbind(A,B,cent,IPs)
text(txt+cbind(rep(xd*.03,nrow(txt)),rep(-yd*.03,nrow(txt))),c("A","B","M",IP.txt))

}
\seealso{
\code{\link{int.2lines}}
}
