% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropEdge2D.R
\name{NumArcsPEtri}
\alias{NumArcsPEtri}
\title{Number of arcs of Proportional Edge Proximity Catch Digraphs (PE-PCDs) - one triangle case}
\usage{
NumArcsPEtri(Xp, tri, r, M = c(1, 1, 1))
}
\arguments{
\item{Xp}{A set of 2D points which constitute the vertices of PE-PCD.}

\item{tri}{A \eqn{3 \times 2} matrix with each row representing a vertex of the triangle.}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region;
must be \eqn{\ge 1}.}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the triangle \code{tri} or the circumcenter of \code{tri}
which may be entered as "CC" as well;
default is \eqn{M=(1,1,1)} i.e., the center of mass of \code{tri}.}
}
\value{
A \code{list} with the elements
\item{num.arcs}{Number of arcs of the PE-PCD}
\item{num.in.tri}{Number of \code{Xp} points in the triangle, \code{tri}}
\item{ind.in.tri}{The vector of indices of the \code{Xp} points that reside in the triangle}
}
\description{
Returns the number of arcs of Proportional Edge Proximity Catch Digraphs (PE-PCDs) whose vertices are the
given 2D numerical data set, \code{Xp}. It also provides number of vertices (i.e., number of data points inside the triangle)
and indices of the data points that reside in the triangle.

PE proximity region \eqn{N_{PE}(x,r)} is defined with respect to the triangle, \code{tri}
for this function. PE proximity region is constructed with expansion parameter \eqn{r \ge 1} and vertex regions are
based on the center \eqn{M=(m_1,m_2)} in Cartesian coordinates or \eqn{M=(\alpha,\beta,\gamma)} in barycentric coordinates
in the interior of the triangle \code{tri} or based on circumcenter of \code{tri}; default is \eqn{M=(1,1,1)} i.e., the center
of mass of \code{tri}. For the number of arcs, loops are not
allowed so arcs are only possible for points inside the triangle \code{tri} for this function.

See also (\insertCite{ceyhan:Phd-thesis,ceyhan:arc-density-PE;textual}{pcds}).
}
\examples{
\dontrun{
A<-c(1,1); B<-c(2,0); C<-c(1.5,2);
Tr<-rbind(A,B,C);

n<-10  #try also n<-20
set.seed(1)
Xp<-runif.tri(n,Tr)$g

M<-as.numeric(runif.tri(1,Tr)$g)  #try also M<-c(1.6,1.0)

NumArcsPEtri(Xp,Tr,r=1.25,M)
NumArcsPEtri(Xp,Tr,r=1.5,M)
}

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{NumArcsPETe}}, \code{\link{NumArcsPE}}, \code{\link{NumArcsCStri}},
and \code{\link{NumArcsAStri}}
}
\author{
Elvan Ceyhan
}
