% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDFunctions.r
\name{plotArcsASMT}
\alias{plotArcsASMT}
\title{The plot of the arcs of Arc Slice Proximity Catch Digraph (AS-PCD) for a 2D data set -
multiple triangle case}
\usage{
plotArcsASMT(
  Xp,
  Yp,
  M = "CC",
  asp = NA,
  main = "",
  xlab = "",
  ylab = "",
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{Xp}{A set of 2D points which constitute the vertices of the AS-PCD}

\item{Yp}{A set of 2D points which constitute the vertices of the Delaunay triangulation. The Delaunay
triangles partition the convex hull of \code{Yp} points.}

\item{M}{"CC" which stands for circumcenter of each Delaunay triangle \code{tri} or 3D point in barycentric
coordinates which serves as a center in the interior of each Delaunay triangle;
default is \code{M}="CC" i.e. the circumcenter of each triangle}

\item{asp}{Numeric, giving the aspect ratio for y axis to x-axis y/x (default is NA),
see the official help for asp by typing "? asp"}

\item{main}{An overall title for the plot (default="")}

\item{xlab, ylab}{Titles for the x and y axes, respectively (default="" for both)}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x- and y-coordinate ranges (default=NULL for both)}

\item{...}{Additional plot parameters}
}
\value{
A plot of the arcs of the AS-PCD for a 2D data set \code{Xp} where AS proximity regions
are defined with respect to the Delaunay triangles based on \code{Yp} points; also plots the Delaunay triangles
based on \code{Yp} points
}
\description{
Plots the arcs of AS-PCD whose vertices are the data points in \code{Xp} and Delaunay triangles based on \code{Yp} points.

AS proximity regions are constructed with respect to the Delaunay triangles based on \code{Yp} points, i.e.,
AS proximity regions are defined only for \code{Xp} points inside the convex hull of \code{Yp} points.
That is, arcs may exist for \code{Xp} points only inside the convex hull of \code{Yp} points.

Vertex regions are based on the center \code{M}="CC"
for circumcenter of each Delaunay triangle or \eqn{M=(\alpha,\beta,\gamma)} in barycentric coordinates in the
interior of each Delaunay triangle; default is \code{M}="CC" i.e. circumcenter of each triangle.
See (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:mcap2012;textual}{pcds}) for more on AS-PCDs.
Also see (\insertCite{okabe:2000,ceyhan:comp-geo-2010,renka:1996;textual}{pcds}) for more on Delaunay triangulation and the TRIPACK algorithm.
}
\examples{
nx<-100
ny<-10

#set.seed(1)
Xp<-cbind(runif(nx,0,1),runif(nx,0,1))
#Xp<-cbind(runif(nx,0.5,1.5),runif(nx,0,1))
#Xp<-cbind(runif(nx,1,2),runif(nx,0,1))
Yp<-cbind(runif(ny,0,1),runif(ny,0,1))

M<-c(1,1,1) #M<-c(1,2,3) #M<-c(-1,2,3) #barycentric coordinates
#M<-"CC"
#M<-c(1.3,1.3)

plotArcsASMT(Xp,Yp,M,xlab="",ylab="")
plotArcsASMT(Xp,Yp,M,asp=1,xlab="",ylab="")

Xlim<-range(Xp[,1],Yp[,1])
Ylim<-range(Xp[,2],Yp[,2])
xd<-Xlim[2]-Xlim[1]
yd<-Ylim[2]-Ylim[1]

plotArcsASMT(Xp,Yp,M,xlab="",ylab="",xlim=Xlim+xd*c(-.01,.01),ylim=Ylim+yd*c(-.01,.01))
plotArcsASMT(Xp,Yp,M,asp=1,xlab="",ylab="",xlim=Xlim+xd*c(-.01,.01),ylim=Ylim+yd*c(-.01,.01))

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plotArcsAStri}}, \code{\link{plotArcsPEtri}}, \code{\link{plotArcsPEMT}},
\code{\link{plotArcsCStri}}, and \code{\link{plotArcsCSMT}}
}
