% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxRFuncs4PCDClasses.r
\name{print.Extrema}
\alias{print.Extrema}
\title{Print a \code{Extrema} object}
\usage{
\method{print}{Extrema}(x, ...)
}
\arguments{
\item{x}{A \code{Extrema} object.}

\item{\dots}{Additional arguments for the S3 method '\code{print}'.}
}
\value{
The \code{call} of the object of class '\code{Extrema}'
and also the \code{type} (i.e. a brief description) of the extrema).
}
\description{
Prints the \code{call} of the object of class '\code{Extrema}'
and also the \code{type} (i.e. a brief description) of the extrema).
}
\examples{
n<-20
dat<-runifTe(n)$gen.points
Ext<-cl2edgesTe(dat)
Ext
print(Ext)

}
\seealso{
\code{\link{summary.Extrema}}, \code{\link{print.summary.Extrema}}, and \code{\link{plot.Extrema}}
}
