% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFunctions.R
\name{print.Patterns}
\alias{print.Patterns}
\title{Print a \code{Patterns} \code{object}}
\usage{
\method{print}{Patterns}(x, ...)
}
\arguments{
\item{x}{A \code{Patterns} \code{object}.}

\item{\dots}{Additional arguments for the S3 method \code{'print'}.}
}
\value{
The \code{call} of the \code{object} of class \code{"Patterns"}
and also the \code{type} (or description) of the pattern).
}
\description{
Prints the \code{call} of the \code{object} of class \code{"Patterns"}
and also the \code{type} (or description) of the pattern).
}
\examples{
\dontrun{
nx<-10; #try also 20, 100, and 1000
ny<-5; #try also 1
e<-.15;
Y<-cbind(runif(ny),runif(ny)) #with default bounding box (i.e., unit square)

Xdt<-rseg.circ(nx,Y,e)
Xdt
print(Xdt)
}

}
\seealso{
\code{\link{summary.Patterns}}, \code{\link{print.summary.Patterns}}, and \code{\link{plot.Patterns}}
}
