% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFunctions.R
\name{print.summary.Extrema}
\alias{print.summary.Extrema}
\title{Print a summary of a \code{Extrema} \code{object}}
\usage{
\method{print}{summary.Extrema}(x, ...)
}
\arguments{
\item{x}{An \code{object} of class \code{"summary.Extrema"}, generated by \code{summary.Extrema}.}

\item{\dots}{Additional parameters for \code{print}.}
}
\value{
None
}
\description{
Prints some information about the \code{object}.
}
\seealso{
\code{\link{print.Extrema}}, \code{\link{summary.Extrema}}, and \code{\link{plot.Extrema}}
}
