% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFunctions.R
\name{print.summary.Lines3D}
\alias{print.summary.Lines3D}
\title{Print a summary of a \code{Lines3D} \code{object}}
\usage{
\method{print}{summary.Lines3D}(x, ...)
}
\arguments{
\item{x}{An \code{object} of class \code{"summary.Lines3D"}, generated by \code{summary.Lines3D}.}

\item{\dots}{Additional parameters for \code{print}.}
}
\value{
None
}
\description{
Prints some information about the \code{object}.
}
\seealso{
\code{\link{print.Lines3D}}, \code{\link{summary.Lines3D}}, and \code{\link{plot.Lines3D}}
}
