% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxGeometry.R
\name{slope}
\alias{slope}
\title{The slope of a line}
\usage{
slope(a, b)
}
\arguments{
\item{a, b}{2D points that determine the straight line (i.e., through which the straight line passes).}
}
\value{
Slope of the line joining 2D points \code{a} and \code{b}
}
\description{
Returns the slope of the line joining two distinct 2D points \code{a} and \code{b}.
}
\examples{
A<-c(-1.22,-2.33); B<-c(2.55,3.75)
slope(A,B)

slope(c(1,2),c(2,3))

}
\seealso{
\code{\link{Line}}, \code{\link{paraline}}, and \code{\link{perpline}}
}
\author{
Elvan Ceyhan
}
